### 2019-12-03 #1933 @kthyng
Hi @DavidLitwin! Can you please look through the list of volunteer reviewers and suggest about 5 who you think would be good fits? You can also recommend others that you know of even if they are not on the list.

https://bit.ly/joss-reviewers
