### 2019-12-04 #1933 @kthyng
Thanks @dvalters!

@nicgaspar (or others who have been mentioned): are you interested in reviewing this submission for JOSS? We ask for reviews within about 3 weeks if possible and you can learn more about review guidelines here: https://joss.readthedocs.io/en/latest/review_criteria.html
