### 2019-12-04 #1935 @whedon
[REVIEW]: GroundwaterDupuitPercolator: A Landlab component for groundwater flow
**Submitting author:** @davidlitwin (<a href="http://orcid.org/0000-0002-8097-4029">David Litwin</a>)
**Repository:** <a href="https://github.com/landlab/landlab" target ="_blank">https://github.com/landlab/landlab</a>
**Version:** v2.0.0b5
