### 2019-12-04 #1933 @rreinecke
@kthyng Is that a new review process? Just woke up to 15 emails without then even the chance to respond. I would have been interested to review as well - but if this is the new way Joss is assigning reviewers ... Could you explain this new process? Till a week ago the handling editor used to choose the reviewers - and only when they declined somebody else was asked (again by the editor not the person who's work is reviewed).
