### 2019-12-04 #1933 @kthyng
@rreinecke No, not new. The submitting author listed possible reviewers, as is always requested in the first comment in a pre-review issue and I as the handling editor additionally requested this list. The submitter author mentioned them specifically through github so people were aware of the suggestion before I subsequently wrote to them, and one volunteered from that list of mentions. I then specifically asked one person from the list of suggestions, and I typically start by asking possible reviewers who have not reviewed for us yet, or who have done so less than other possibilities, so as to avoid reviewer burnout. I prefer to specifically ask a smaller number of people rather than a bunch at once since then in that case it can fall in the cracks for awhile. We do not give a particular time frame for people to accept or decline the review.
