### 2019-12-05 #1933 @rreinecke
> @rreinecke No, not new. The submitting author listed possible reviewers, as is always requested in the first comment in a pre-review issue and I as the handling editor additionally requested this list. The submitter author mentioned them specifically through github so people were aware of the suggestion before I subsequently wrote to them, and one volunteered from that list of mentions. I then specifically asked one person from the list of suggestions, and I typically start by asking possible reviewers who have not reviewed for us yet, or who have done so less than other possibilities, so as to avoid reviewer burnout. I prefer to specifically ask a smaller number of people rather than a bunch at once since then in that case it can fall in the cracks for awhile. We do not give a particular time frame for people to accept or decline the review.

@kthyng Thank you for your thorough explanation. However, you have to be aware that this way I'm getting unnecessary spam for a review I'm not even involved in. If I'm not going to review, or to be more precise given the chance to accept or decline that review, I don't want to be mentioned at all. Or am I missing something here?
