### 2019-12-05 #1933 @kthyng
@rreinecke We don't have control over whether people are mentioned by submitters in reviews the way we are doing things, but we are always open to feedback on our process. Do you have any specific suggestions for improvements?
