### 2019-12-05 #1933 @rreinecke
> @rreinecke We don't have control over whether people are mentioned by submitters in reviews the way we are doing things, but we are always open to feedback on our process. Do you have any specific suggestions for improvements?

@kthyng of course you do. At least a little bit: 1) Only ask for 2 suggestions for possible reviewers and only ask for more if they decline or do not answer in a given time frame. Or 2) as many of the other editors are doing it just recommend two reviewers yourself. Yes everything will be slowed down a little bit but you at least reduce the possibility to get mail on submissions one is not even concerned with. Also with the approach you have chosen the reviewer assignment doesn't look like the process you described but more like a first come first serve approach - that can't be what you want to get high quality reviews. You have to understand my frustration to be notified on something that does not even give me the possibility to react but just fills up my inbox @danielskatz @kthyng .
