### 2019-12-05 #1933 @danielskatz
@rreinecke - one thing we should make clear in our documentation and instructions from editor is for the author to suggest reviewers without the @ so that there are no notifications that start at this point.  I'll open an issue and also work on a PR to try to address this.
