### 2019-12-05 #1933 @dvalters
@rreinecke - you are welcome to review in my place if you are interested in the submission. I've done a fair few reviews this year for JOSS and would be happy to pass the opportunity on to someone else.
