### 2019-12-06 7725913
plotting updates
With Allison's guidance, plot network links is in a nice function and plot network parcels is also in a nice function and nearly has all the desired functionality. One thing I can't figure out is how to get a contour legend for the network parcels to appear. Also, I have written a small bit of code to go from parcel attributes to link attributes and I know those bits of code exist in NST to go from link attributes to parcel attributes, but I have not put these into nice genearlizable functions yet. This would greatly improve functionality. I can work on this after AGU. This fu
[…]
