+++ landlab/plot/network_sediment_transporter/__init__.py
@@ -5,6 +5,13 @@ Created on Tue Jun 18 14:22:27 2019
 @author: pfeif
 """
 from .plot_network_links import plot_network_links
-from .other_example_plot import other_example_plot
+from .locate_parcel_xy import locate_parcel_xy
+from .plot_network_parcels import plot_network_parcels
+#from .other_example_plot import other_example_plot
 
-__all__ = ["plot_network_links","other_example_plot"]
+#__all__ = ["plot_network_links","other_example_plot"]
+__all__ = [
+        "plot_network_links",
+        "locate_parcel_xy",
+        "plot_network_parcels",
+           ]

[…]
