### 2019-12-08 ec9cb6c
improved zone boundary conditions description
