+++ landlab/components/species_evolution/zone_controller.py
@@ -32,11 +32,11 @@ class ZoneController(object):
     values         mask returned
     evaluated      by zone function
     0 0 0 0 0 0    · · · · · ·
-    0 0 4 5 4 0    · · * * * ·
-    0 0 6 3 0 0    · · * * · ·
+    0 0 0 5 4 2    · · · * * *
+    0 6 0 4 6 0    · * · * * ·
+    0 2 0 0 0 0    · * · · · ·
+    0 0 4 0 4 0    · · * · * ·
     0 0 0 0 0 0    · · · · · ·
-    0 6 4 0 4 0    · * * · * ·
-    2 0 0 0 0 0    * · · · · ·
 
     The above example is continued in the following four diagrams that
     demonstrate how individual zones are identified. Each zone is marked with a
@@ -48,11 +48,35 @@ class ZoneController(object):
     D8             D4             D8             D4
     min area = 0   min area = 0   min area = 2   min area = 2
     · · · · · ·    · · · · · ·    · · · · · ·    · · · · · ·
-    · · + + + ·    · · + + + ·    · · + + + ·    · · + + + ·
-    · · + + · ·    · · + + · ·    · · + + · ·    · · + + · ·
+    · · · + + +    · · · + + +    · · · + + +    · · · + + +
+    · x · + + ·    · x · + + ·    · x · + + ·    · x · + + ·
+    · x · · · ·    · x · · · ·    · x · · · ·    · x · · · ·
+    · · x · o ·    · · @ · o ·    · · x · · ·    · · · · · ·
     · · · · · ·    · · · · · ·    · · · · · ·    · · · · · ·
-    · x x · o ·    · @ @ · o ·    · x x · · ·    · x x · · ·
-    x · · · · ·    x · · · · ·    x · · · · ·    · · · · · ·
+
+    The grid boundary affects zone creation. Zones can include nodes along the
+    boundary, although because boundary nodes are not associated with cells
+    that have area, boundary nodes do not contribute to the area summation of
+    clusters. The area summation only takes into account the cells associated
+    with core nodes.
+    
+    Creation of zones along boundaries is illustrated below. A zone extent mask
+    different from the above example was produced by the hypothetical zone
+    function here. Again ``*`` indicates where a zone can exist. Distinct zones
+    that are created are in the other four grid schematics, where the symbols,
+    ``$`` and ``#`` indicate zones in addition to the symbols defined above.
+    Individual zone masks and the count of zones are affected by the use of
+    ``D8`` or ``D4`` along with the minimum area parameter, especially when
+    zone clusters are along grid boundaries.
+    
+    zone function  D8             D4             D8             D4
+    returned mask  min area = 0   min area = 0   min_area = 2   min_area = 2
+    * · · * * ·    + · · x x ·    + · · x x ·    + · · · · ·    · · · · · ·
+    · * · · * ·    · + · · x ·    · # · · x ·    · + · · · ·    · # · · · ·
+    · * · · · ·    · + · · · ·    · # · · · ·    · + · · · ·    · # · · · ·
+    * · · · * *    + · · · o o    $ · · · o o    + · · · o o    · · · · o o
+    · · · · * ·    · · · · o ·    · · · · o ·    · · · · o ·    · · · · o ·
+    · * * · · ·    · @ @ · · ·    · @ @ · · ·    · · · · · ·    · · · · · ·
 
     By default, ``ZoneSpecies`` are used with this controller, and the
     following paragraphs make that assumption. Speciation occurs when the
@@ -82,24 +106,26 @@ class ZoneController(object):
 
     T1a            T1b            T1c            T1d
     · · · · · ·    · · · · · ·    · · · · · ·    · · · · · ·
-    · · · · · ·    · · · · · ·    · · · * · *    · · · · · ·
[…]
