### 2019-12-08 40f780e
fixed lint issues
