### 2019-12-14 9adb21e
remove tutorials stage from Travis CI
