+++ .travis.yml
@@ -12,23 +12,6 @@ env:
 sudo: false
 jobs:
   include:
-  - stage: tutorials
-    if: branch = release
-    os: linux
-    install:
-    - conda env create --file environment-dev.yml
-    - conda activate landlab_dev
-    - pip install -e .
-    script:
-    - pushd ../..
-    - git clone https://github.com/landlab/tutorials.git landlab/tutorials
-    - pushd landlab/tutorials
-    - git fetch origin release
-    - git checkout -qf release
-    - conda install --file=requirements.txt
-    - conda install nbformat
-    - travis_wait 50 pytest -vvv
-
   - stage: lint
     os: linux
     script:
