### 2019-12-15 517f534
install nbformat for testing notebooks
