### 2019-12-15 c12731c
install jupyter for testing notebooks
