+++ .travis.yml
@@ -86,6 +86,6 @@ install:
 - pip install numpy
 - pip install -e .
 script:
-- pip install pytest pytest-cov pytest-datadir coveralls nbformat
+- pip install pytest pytest-cov pytest-datadir coveralls jupyter nbformat
 - pytest --cov=landlab --cov-report=xml:$(pwd)/coverage.xml --run-notebook -vvv
 after_success: coveralls
