### 2019-12-15 31f9c2f
install dask for notebooks
