### 2019-12-15 e63af8d
install holoviews for notebook tests
