### 2019-12-18 f740c82
first steps to merging
