+++ landlab/components/groundwater/dupuit_percolator.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """GroundwaterDupuitPercolator Component.
 
-@author: G Tucker, D Litwin
+@author: G Tucker, D Litwin, K Barnhart
 """
 
 import numpy as np
@@ -48,12 +48,36 @@ def get_link_hydraulic_conductivity(grid, K):
 
 
 class GroundwaterDupuitPercolator(Component):
-    """Simulate groundwater flow in a shallow unconfined aquifer.
+    r"""
+    Simulate groundwater flow in a shallow unconfined aquifer.
 
     The GroundwaterDupuitPercolator solves the Boussinesq equation for
-    flow in an unconfined aquifer over an impermeable aquifer base that may
-    have uneven slope. This method uses the Dupuit approximation that the
-    hydraulic gradient is equal to the slope of the water table.
+    flow in an unconfined aquifer over an impermeable aquifer base and
+    calculates groundwater return flow to the surface. This method uses the
+    Dupuit approximation that the hydraulic gradient is zero in the direction
+    normal to the aquifer base. For this reason, aquifer thickness
+    and regolith thickness are calculated in the direction normal to the
+    aquifer base. The hydraulic gradient is also calculated relative
+    to the aquifer base. For more details, see component documentation
+    :ref:`here <dupuit_theory>`.
+
+    Parameters
+    ----------
+    grid: ModelGrid
+            Landlab ModelGrid object
+    hydraulic_conductivity: float, field name, or array of float
+            saturated hydraulic conductivity, m/s.
+            Default = 0.001 m/s
+    porosity: float, field name or array of float
+            the porosity of the aquifer [-].
+            Default = 0.2
+    recharge_rate: float, field name, or array of float
+            Rate of recharge, m/s.
+            Default = 1.0e-8 m/s
+    regularization_f: float
+            factor controlling the smoothness of the transition between
+            surface and subsurface flow.
+            Default = 1e-2
 
     Examples
     --------
@@ -76,9 +100,9 @@ class GroundwaterDupuitPercolator(Component):
     >>> for i in range(100):
     ...     gdp.run_one_step(dt)
 
-    In an example that generates surface water leakage, the surface water flux
+    When the model generates groundwater return flow, the surface water flux
     out of the domain can be calculated only after a FlowAccumulator is run.
-    Here is a more advanced example with a sloping aquifer that returns surface water flow.
+    Below is a more complex model that demonstrates this case.
 
     >>> from landlab.components import FlowAccumulator
 
@@ -115,22 +139,34 @@ class GroundwaterDupuitPercolator(Component):
 
     Notes
     -----
-    Groundwater discharge per unit length, q, is calculated as:
+    Below is a summary of the theory and numerical implementation of
+    the ``GroundwaterDupuitPercolator``. A complete description can be found
+    :ref:`here <dupuit_theory>`.
 
-        q = - K H ( dH/dx cos(alpha) + sin(alpha) ),
+    Groundwater discharge per unit length, :math:`q`, is calculated as:
 
-    where K is hydraulic conductivity, H is aquifer thickness, alpha is
-    the slope angle of the aquifer base, and x is horizontal distance.
+    .. math::
+        q = -K_{sat} h \big( \nabla z \big) \cos^2 (\alpha)
 
-    Surface water discharge per unit area, qs, is calculated as:
+    where :math:`K_{sat}` is the saturated hydraulic conductivity, :math:`h` is
+    the aquifer thickness, and :math:`\alpha` is the slope angle of the aquifer base.
 
-        qs = G( H/(Z-Zb) ) * R( f - dq/dx)
+    Surface water discharge per unit area, :math:`q_s`, is calculated as:
 
[…]
