+++ docs/source/reference/components/groundwater.rst
@@ -1,7 +1,7 @@
-GroundwaterDupuitPercolator: model flow in a shallow unconfined aquifer using the Dupuit-Forscheimer approximation
-------------------------------------------------------------------------------------------------------------------
+GroundwaterDupuitPercolator: model flow in a shallow unconfined aquifer using the Dupuit-Forcheimer approximation
+----------------------------------------------------------------------------------------------------
 
-.. automodule:: landlab.components.groundwater
+.. automodule:: landlab.components.groundwater.dupuit_percolator
     :members:
     :undoc-members:
     :show-inheritance:
