### 2019-12-18 79acb4f
working towards functional tests and docs
