+++ docs/source/reference/components/groundwater.rst
@@ -1,5 +1,5 @@
 GroundwaterDupuitPercolator: model flow in a shallow unconfined aquifer using the Dupuit-Forcheimer approximation
-----------------------------------------------------------------------------------------------------
+-----------------------------------------------------------------------------------------------------------------
 
 .. automodule:: landlab.components.groundwater.dupuit_percolator
     :members:
