### 2019-12-18 c52f239
Fix velocity error from merge
