### 2019-12-18 5961b8e
fix redirects
