### 2019-12-18 ea2cf1c
Update GroundwaterDupuitPercolator (equivalent of 1047, but for v2_dev, #1076)
* first steps to merging

* working towards functional tests and docs

* refer to self._dhdt

* add linkcheck to toplevel makefile

* doc update

* split test into two

* Fix velocity error from merge

* fix redirects

* one more redirect

* workaround with linkcheck_ignore
