### 2019-12-18 782ac4c
add teaching notebooks
