+++ notebooks/conftest.py
@@ -1,9 +1,10 @@
 import os
+import pathlib
 
 import pytest
 
 
-_TEST_DIR = os.path.abspath(os.path.dirname(__file__))
+_TEST_DIR = pathlib.Path(__file__).absolute().parent
 _EXCLUDE = [
     "animate-landlab-output.ipynb",
     "cellular_automaton_vegetation_flat_domain.ipynb",
@@ -11,20 +12,32 @@ _EXCLUDE = [
 ]
 
 
-def _all_notebooks(path="."):
-    notebooks = []
-    for root, dirs, files in os.walk(path):
-        if ".ipynb_checkpoints" in root:
-            continue
-        for file in files:
-            if file.endswith(".ipynb") and (file not in _EXCLUDE):
-                notebooks.append(os.path.join(root, file))
-    return notebooks
+def collect_notebooks(src):
+    p = pathlib.Path(src)
+    if p.is_dir():
+        return set([_p.absolute() for _p in iter_notebooks_in_dir(p, src)])
+    else:
+        raise ValueError("{0}: not a directory".format(src))
+
+
+def iter_notebooks_in_dir(path, root):
+    for s in path.iterdir():
+        p = pathlib.Path(s)
+        normalized_path = "/" + p.resolve().relative_to(root).as_posix()
+
+        if p.is_dir():
+            normalized_path += "/"
+        if p.is_dir() and p.name not in (".git", ".ipynb_checkpoints"):
+            yield from iter_notebooks_in_dir(p, root)
+        elif p.is_file() and p.suffix == ".ipynb" and p.name not in _EXCLUDE:
+            yield p
 
 
 def pytest_generate_tests(metafunc):
     if "notebook" in metafunc.fixturenames:
-        metafunc.parametrize("notebook", _all_notebooks(_TEST_DIR))
+        metafunc.parametrize(
+            "notebook", sorted([str(name) for name in collect_notebooks(_TEST_DIR)])
+        )
 
 
 def pytest_configure(config):
