### 2019-12-19 add36a9
mark notebooks to exclude with pytest.skip
