### 2019-12-19 4aad077
add requirements file for landlab that setup.py reads
