### 2019-12-19 7abefab
add requirements for running the notebooks
