### 2019-12-19 f595a24
add requirements for CI
