+++ appveyor.yml
@@ -34,7 +34,7 @@ install:
   - cmd: conda update --yes --quiet conda
   - cmd: set PYTHONUNBUFFERED=1
   - cmd: conda config --set always_yes yes
-  - cmd: pip install twine pytest pytest-cov pytest-datadir jupyter nbformat dask holoviews
+  - cmd: pip install twine -r requirements-testing.txt -r notebooks/requirements.txt
   - cmd: conda info
   - cmd: conda list
 
