### 2019-12-19 4f4b9b8
require bmipy less than 1.0
