### 2019-12-19 8e47a2e
bit of update to the notebooks that help people move around in the repo
