### 2019-12-19 03aa30f
add exclude file for notebooks to skip when testing
