+++ notebooks/exclude.yml
@@ -1,0 +1,8 @@
+- file: animate-landlab-output.ipynb
+  reason: "requires ffmpeg, imagemagick"
+- file: cellular_automaton_vegetation_flat_domain.ipynb
+  reason: "not v2 compatible"
+- file: cellular_automaton_vegetation_DEM.ipynb
+  reason: "not v2 compatible"
+- file: stream_power_channels_class_notebook.ipynb
+  reason: "not v2 compatible"
