### 2019-12-19 0e70676
move environment.yml for binder to top-level folder
