+++ appveyor.yml
@@ -34,7 +34,7 @@ install:
   - cmd: conda update --yes --quiet conda
   - cmd: set PYTHONUNBUFFERED=1
   - cmd: conda config --set always_yes yes
-  - cmd: pip install twine -r requirements-testing.txt -r notebooks/requirements.txt
+  - cmd: pip install twine -r requirements-testing.txt -r requirements-notebooks.txt
   - cmd: conda info
   - cmd: conda list
 
