### 2019-12-19 8ccf790
update environment specifications
