### 2019-12-20 d99c05b
docs: add tutorial notebooks to repository (#1071)
* add landlab tutorial notebooks to repository

* remove tutorials stage from Travis CI

* test notebooks on Travis CI

* use --run-notebook option

* add notebook to markers

* add top-level conftest.py

* install nbformat for testing notebooks

* install jupyter for testing notebooks

* install dask for notebooks

* install holoviews for notebook tests

* move tutorial notebooks to tutorials folder

* add teaching notebooks

* use pathlib

* update teaching notebooks to v2; blacken

* mark notebooks to exclude with pytest.skip

* add groundw
[…]
