+++ setup.cfg
@@ -8,7 +8,7 @@ parentdir_prefix = landlab-
 
 [tool:pytest]
 minversion = 3.0
-testpaths = landlab tests
+testpaths = notebooks landlab tests
 norecursedirs = .* *.egg* build dist examples
 # usefixtures = suppress_resource_warning
 addopts =
@@ -25,6 +25,7 @@ doctest_optionflags =
     ALLOW_UNICODE
 markers =
     slow: marks tests as slow (deselect with '-m "not slow"')
+    notebook: marks tests as notebook (deselect with '-m "not notebook"')
 
 [isort]
 multi_line_output=3
