### 2019-12-20 c203595
landlab v2 (#1077)
* BREAKING CHANGE: Remove py27 support (#946)

* remove 27 from travis and appveyor

* remove 27 from setup.py

* remove __future__.print_function

* remove six.moves range

* remove six from env dev

* update contributing

* remove six.print

* removing six import

* six.stringIO

* string_types to str

* remove six from setup.py

* iteritems

* remove six from dist file

* misc six removes

* fix: add Seed Parameter to CA and update CI

* fix: add seed parameter to CA submodule (#920)

* wip

* CTS subclasses take seed, pass it to base class

* add test

* clean up in cfun
[…]
