### 2019-12-20 e99fcc4
fix 6 files that did not update in v2 transition
four files should have been deleted and two should have been updated.
