+++ docs/dupuit_theory.rst
@@ -1,162 +1,0 @@
-.. _dupuit_theory:
-
-========================================================
-Theory and Implementation of GroundwaterDupuitPercolator
-========================================================
-
-Governing Equations
-===================
-
-Variably-saturated groundwater flow is often assumed to be governed by
-the Richards equation, which describes how water content and/or total
-energy potential evolve in an idealized porous medium due to fluxes of
-water driven by gradients in total potential, :math:`h = z + p/ \gamma`,
-where :math:`z` is the elevation head, :math:`p` is the gage pressure,
-and :math:`\gamma` is the specific weight of water.
-
-.. math:: \frac{\partial \theta}{\partial t} = \frac{\partial}{\partial x} \bigg( k \frac{\partial h}{\partial x} \bigg) + \frac{\partial}{\partial y} \bigg( k \frac{\partial h}{\partial y} \bigg) + \frac{\partial}{\partial z} \bigg( k \frac{\partial h}{\partial z} \bigg)
-
-Here :math:`\theta` is the volumetric water content of the aquifer and
-:math:`k` is the hydraulic conductivity, which may be a function of
-:math:`\theta`. Here we use the widely applied Dupuit-Forcheimer
-approximation, which is valid when the aquifer is laterally extensive in
-comparison to its thickness, and the capillary fringe above the water
-table is relatively thin. If this is the case, the component of the
-hydraulic gradient normal to the aquifer base can be neglected, and the
-water table can be treated as a free surface. Consequently, the total
-head is equal to the water table elevation, :math:`h=z`. With these
-assumptions, an adjusted governing equation can be written for the time
-evolution of the water table elevation:
-
-.. math:: n \frac{\partial h}{\partial t} = \frac{\partial}{\partial x} \bigg( k_{sat} h \frac{\partial h}{\partial x} \bigg) + \frac{\partial}{\partial y} \bigg( k_{sat} h \frac{\partial h}{\partial y} \bigg)
-
-where :math:`n` is the drainable porosity, and :math:`k_{sat}` is the
-saturated hydraulic conductivity.
-
-.. figure:: /images/water_table_schematic.png
-   :alt: Aquifer schematic.
-   :align: center
-   :scale: 50%
-
-   Aquifer schematic showing vertical aquifer thickness :math:`h`, bed-normal
-   aquifer thickness :math:`\eta`, and water table elevation :math:`z`.
-
-
-When the aquifer base is sloping, the governing equations must be adjusted.
-Childs (1971) provides the governing equation for the groundwater specific discharge as:
-
-.. math:: q_{x'} = - k_{sat} \eta \frac{\partial z}{\partial x'}
-
-where :math:`x'` is the coordinate parallel to the impermeable base, and :math:`\eta`
-is the aquifer thickness perpendicular to the impermeable base ([2]_).
-The `GroundwaterDupuitPercolator` treats two additional fluxes that affect
-aquifer storage: groundwater return flow to the surface :math:`q_s`, and
-recharge from precipitation :math:`f`. Implementations of the Dupuit-Forcheimer
-model often encounter numerical instabilities as the water table intersects the
-land surface. To alleviate this problem, we use the regularization approach
-introduced by Marcais et al. (2017), which smooths the transition between
-surface and subsurface flow ([1]_). The complete governing equations in the
-base-parallel reference frame :math:`(x',y')` are:
-
[…]
