### 2019-12-20 447471e
Tidying
Minor tidying
