+++ docs/source/conf.py
@@ -124,7 +124,7 @@ html_theme = 'sphinxdoc'
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
-#html_theme_options = {}
+html_theme_options = {'github_banner': True}
 
 # Add any paths that contain custom themes here, relative to this directory.
 #html_theme_path = []
