+++ docs/source/getting_started/teach_yourself_landlab.rst
@@ -4,5 +4,5 @@ Self-study course for learning to use Landlab in your research
 ==============================================================
 
 This course is entirely hosted in our Jupyter notebook tutorials. Here is a
-link to the `syllabus <https://mybinder.org/v2/gh/landlab/tutorials/v2_dev?filepath=syllabus.ipynb>`_
+link to the `syllabus <https://mybinder.org/v2/gh/landlab/landlab/release?filepath=notebooks/tutorials/syllabus.ipynb>`_
 on Binder. 
