+++ docs/source/user_guide/tutorials.rst
@@ -3,102 +3,18 @@
 Tutorials
 =========
 
-The `Landlab tutorials <https://github.com/landlab/tutorials>`_
-repository contains `IPython
-notebooks <https://ipython.org/notebook.html>`_ that are both
-unexpanded (so you can run them in an IPython notebook viewer) and
-expanded (so you can read them like a regular text tutorial) along with
-code examples. Landlab's tutorials repo can also be accessed from
-`Binder <https://mybinder.org/v2/gh/landlab/tutorials/v2_dev>`_.
+The Landlab tutorials are Jupyter notebooks that give examples of Landlab core
+concepts, component introductions, and teaching exercises.
 
-IPython notebook tutorials
-~~~~~~~~~~~~~~~~~~~~~~~~~~
+Landlab's tutorials are located within the main Landlab repository in the
+`notebooks <https://github.com/landlab/landlab/tree/master/notebooks>`_ folder.
+They can also be accessed using
+`Binder <https://mybinder.org/v2/gh/landlab/landlab/release?filepath=notebooks/welcome.ipynb>`_.
 
-Instructions on how to run an IPython notebook can be found
-`here <https://github.com/landlab/tutorials/blob/release/README.md>`__.
-
-A short IPython notebook tutorial along with a screencast can be found
-`here <http://www.randalolson.com/2012/05/12/a-short-demo-on-how-to-use-ipython-notebook-as-a-research-notebook/>`__
-(the tutorial uses an example with statistics, but you can substitute
-Landlab!)
-
-`Click here to download all the
-tutorials <https://codeload.github.com/landlab/tutorials/zip/release>`_
-
-A suggested introduction to Landlab follows roughly this order:
-
-- `Introduction to Python and
-  NumPy <https://mybinder.org/v2/gh/landlab/tutorials/v2_dev?filepath=python_intro/Python_intro.ipynb>`_.
-  *Learn about:* The very basics of Python.
-- `Introduction to Landlab: example model of fault-scarp
-  degradation <https://mybinder.org/v2/gh/landlab/tutorials/v2_dev?filepath=fault_scarp/landlab-fault-scarp.ipynb>`_.
-  A short overview of some of the things Landlab can do.
-- `Introduction to the model grid
-  object <https://mybinder.org/v2/gh/landlab/tutorials/v2_dev?filepath=grid_object_demo/grid_object_demo.ipynb>`_.
-  Grid topology; how landlab represents data; connectivity of grid
-  elements.
-- `Introduction to Landlab data
-  fields <https://mybinder.org/v2/gh/landlab/tutorials/v2_dev?filepath=fields/working_with_fields.ipynb>`_.
-  How Landlab stores spatial data on the grid; a little on naming
-  conventions.
-- `Introduction to plotting output with
-  Landlab <https://mybinder.org/v2/gh/landlab/tutorials/v2_dev?filepath=plotting/landlab-plotting.ipynb>`_.
-  The basics of plotting with Landlab; combining matplotlib and out
-  plots; the all-powerful ``imshow_grid()`` function.
-- `Introduction to using the Landlab component
-  library <https://mybinder.org/v2/gh/landlab/tutorials/v2_dev?filepath=component_tutorial/component_tutorial.ipynb>`_.
-  The basics of working with and coupling components, using
-  *diffusion*, *stream power*, and a *storm generator* as examples.
-- `Using the gradient and flux-divergence
-  functions <https://mybinder.org/v2/gh/landlab/tutorials/v2_dev?filepath=gradient_and_divergence/gradient_and_divergence.ipynb>`_.
[…]
