### 2019-12-20 ec4ed37
update description of teaching notebooks.
