### 2019-12-20 e0e9f9e
bit more doc updating
