+++ docs/source/index.rst
@@ -21,12 +21,15 @@ across scales.
 The most current source code is always available from our `git
 repository <https://github.com/landlab/landlab>`_.
 
-We compile ongoing development :ref:`here <ongoing_development>` and good first
-issues :ref:`here <desired_contributions>`.
+If you are interested in the state of current development, we compile
+:ref:`ongoing development <ongoing_development>`. If you are interested in
+contributing but don't know how to get started (THANK YOU!),
+we compile :ref:`desired contributions <desired_contributions>` and have a
+:ref:`page all about development <development>`. 
 
 Many Jupyter notebooks exist describing use of Landlab. Find an overview
 :ref:`here <tutorials>`. A subset of these notebooks are designed specifically
-for the classroom. Information about them and how to set them up for classroom 
+for the classroom. Information about them and how to set them up for classroom
 use is described :ref:`on this page <teaching_tutorials>`.
 
 
