### 2019-12-21 d7e1bcd
Post 2.0 documentation update (#1081)
* updating links.

* remove unsupported option.

* add two new pages

* update description of teaching notebooks.

* add .rst extension

* bit more doc updating

* update link to binder

* update teaching tutorial welcome ipynb
