+++ landlab/components/groundwater/readme.md
@@ -2,6 +2,6 @@ Welcome to the README for the GroundwaterDupuitPercolator component submodule.
 
 The documentation is hosted on our [ReadTheDocs](https://landlab.readthedocs.io/en/master/landlab.components.groundwater_dupuit_percolator.html) page.
 
-There is a [jupyter notebook in the Landlab Tutorials repository](https://mybinder.org/v2/gh/landlab/tutorials/release?filepath=groundwater/groundwater_flow.ipynb) that describes the use of this submodule. The link takes you to a binder instance of this notebook.
+There is a [Jupyter notebook in the Landlab Tutorials repository](https://mybinder.org/v2/gh/landlab/landlab/release?filepath=notebooks/tutorials/groundwater/groundwater_flow.ipynb) that describes the use of this submodule. The link takes you to a binder instance of this notebook.
 
 If you have any questions, comments, issues, or bugs related to this submodule, please [open an Issue](https://github.com/landlab/landlab/issues/new) so we can respond.
