+++ landlab/components/species_evolution/base_species.py
@@ -1,10 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """Base Species of SpeciesEvolver."""
-from abc import ABCMeta, abstractmethod
-
-# Get ABC for compatibility with Python 2 and 3.
-ABC = ABCMeta('ABC', (object,), {'__slots__': ()})
+from abc import ABC, abstractmethod
 
 
 class Species(ABC):
