+++ landlab/components/species_evolution/species_evolution.py
@@ -193,20 +193,15 @@ class SpeciesEvolver(Component):
     """
     _name = 'SpeciesEvolver'
 
-    _output_var_names = (
-        'species__richness'
-    )
-
-    _var_units = {
-        'species__richness': 'None'
-    }
-
-    _var_mapping = {
-        'species__richness': 'node',
-    }
-
-    _var_doc = {
-        'species__richness': 'The number of species at each node.'
+    _info = {
+        "species__richness": {
+            "dtype": int,
+            "intent": "out",
+            "optional": False,
+            "units": "-",
+            "mapping": "node",
+            "doc": "The number of species at each node",
+        },
     }
 
     def __init__(self, grid, initial_time=0):
@@ -244,7 +239,9 @@ class SpeciesEvolver(Component):
 
         # Create a species richness field.
 
-        _ = grid.add_zeros('species__richness', at='node', noclobber=False)
+        _ = grid.add_zeros(
+            'species__richness', at='node', dtype=int, clobber=True
+        )
 
     @property
     def record_data_frame(self):
@@ -649,8 +646,8 @@ class SpeciesEvolver(Component):
 
         if species:
             masks = np.stack([s.range_mask for s in species])
-            richness_mask = sum(masks)
+            richness_mask = sum(masks).astype(int)
         else:
-            richness_mask = np.zeros(self._grid.number_of_nodes)
+            richness_mask = np.zeros(self._grid.number_of_nodes, dtype=int)
 
         return richness_mask
