### 2019-12-21 482516d
paper updates before migrate to landlab 2.0
I still have some more edits to make, I am part way through, but want to commit this before the update.
