+++ landlab/components/network_sediment_transporter/joss/paper.md
@@ -12,7 +12,7 @@ authors:
     orcid: 0000-0001-5682-455X
     affiliation: 2, 3
 
-  - name: Jon Czuba
+  - name: Jonathan A. Czuba
     orcid: 0000-0002-9485-2604
     affiliation: 4
 
@@ -37,25 +37,29 @@ bibliography: papers.bib
 
 # Summary
 
-Coarse sediment moves downstream through river networks. The transport rate of any particular sediment grain on the river bed surface is a function of both the hydraulics of that reach of river and the size distribution of the other grains in the reach. As sediment moves through a river system, grains may be deposited or eroded, burying and exposing other grains, and in the process changing the elevation and slope of each segment of river. This process of river channel evolution through the process of sediment transport is referred to as morphodynamics [@cite]. Computational morphodynamic models allow for the prediction of sediment pulse transport, such as that which occurs after dam removal [@citeCui] or landsliding events[@An2017], as well as the prediction of changes in river channel bed surface grain size.
+Coarse sediment moves downstream through river networks. The transport rate of any particular sediment grain on the river bed surface is a function of both the hydraulics of that reach of river and the size distribution of the other grains in the reach. As sediment moves through a river system, grains may be deposited or eroded, burying and exposing other grains, and in the process changing the elevation and slope of each segment of river. This process of river channel evolution through the process of sediment transport is referred to as morphodynamics [@cite]. Computational morphodynamic models allow for the prediction of sediment pulse transport, such as that which occurs after dam removal [@citeCui] or landsliding events[@An2017; @Benda&Dunne1997], as well as the prediction of changes in river channel bed surface grain size[@cite].
 
 Over the past two decades, computational morphodynamic models have ...
 [**JON, could you write two sentences here about the history of morphodynamic models?** ]
 Cui et al. [2006] ; Ferguson et al. [2015]
 Enter network models:CASCADE...
-Czuba (2018) introduced a network-based, Lagrangian bed material morphodynamic model that tracks the motion of individual units (referred to as “parcels”) of sediment through the river network. This approach improves on the existing morphodynamic models by: (1) accounting for the full river network, rather than a single longitudinal profile, (2) allowing the user to ‘tag’ particular sediment inputs and track their fate through time. These existing network sediment transport models, however, have two notable drawbacks: 1) they are written in a proprietary scripting language, and 2) they are not explicitly designed to be interoperable with other earth surface models, such as streamflow or landslide models.
+Czuba (2018) introduced a network-based, Lagrangian bed material morphodynamic model that tracks the motion of individual units (referred to as “parcels”) of sediment through a river network. This approach improves on the existing morphodynamic models by: (1) accounting for the full river network, rather than a single longitudinal profile, (2) allowing the user to ‘tag’ particular sediment inputs and track their fate through time. These existing network sediment transport models, however, have two notable drawbacks: 1) they are written in a proprietary scripting language, and 2) they are not explicitly designed to be interoperable with other Earth-surface models, such as streamflow or landslide models.
 
-Here, we present software that overcomes these two drawbacks, translating network sediment transport model of @Czuba2018 into Landlab, a modular, Python-based package for the modeling of Earth surface dynamics. Landlab is an Open Source Python package for modeling earth surface processes [@Hobley2017Creative]. It was designed as a modular framework, hosting a variety of process components such as flow routing, hillslope diffusion, and stream power erosion that function on a common set of landscape model grids. The original implementation of this model was in MatLab. ``NetworkSedimentTransporter`` is the newest of these components. We first describe computational infrastructure created in order to create the ``NetworkSedimentTransporter`` and then describe the new componet itself.
[…]
