### 2019-12-21 17073a3
2.0 documentation updates (#1083)
* Merge release into master (to try and sync sha keys) (#1079)

* Add tutorials stage to travis build (#1022)

* add tutorials stage to travis build

* fix doctest that assumes Dataset.attrs is OrderedDict

* get into the right directory.

* install additional dependencies

* update release for 1.11 (#1067)

* feat: Addition of LateralEroder to the Landlab Component library

* FirstPass

* changes to 14:30

At this point, chagnes to lateral erosion component (and dependencies) are such that I can import it from a driver!

* changes to 10:13am

* endofdayapril5
[…]
