### 2019-12-25 da7bcbf
resolved conflicts for PET component
- resolved conflicts for PET component
