+++ landlab/components/pet/potential_evapotranspiration_field.py
@@ -1,22 +1,15 @@
 import numpy as np
-
 from landlab import Component
-##<<<<<<< HEAD
 from ...utils.decorators import use_file_name_or_kwds
-import numpy as np
 import math
-##=======
-##>>>>>>> origin/v1
 
 from ...utils.decorators import use_file_name_or_kwds
 
-##<<<<<<< HEAD
-_VALID_METHODS = set(['Constant', 'PriestleyTaylor', 'MeasuredRadiationPT',
-                      'Cosine', 'PenmanMonteith'])
-##=======
-_VALID_METHODS = set(["Constant", "PriestleyTaylor", "MeasuredRadiationPT", "Cosine"])
-##>>>>>>> origin/v1
-
+_VALID_METHODS = set(['Constant',
+                      'Cosine',
+                      'MeasuredRadiationPT',
+                      'PenmanMonteith',
+                      'PriestleyTaylor'])
 
 def _assert_method_is_valid(method):
     if method not in _VALID_METHODS:
@@ -24,7 +17,6 @@ def _assert_method_is_valid(method):
 
 
 class PotentialEvapotranspiration(Component):
-
     """
     Potential Evapotranspiration Component calculates spatially distributed
     potential evapotranspiration based on input radiation factor (spatial
@@ -51,59 +43,6 @@ class PotentialEvapotranspiration(Component):
 
     .. codeauthor:: Sai Nudurupati and Erkan Istanbulluoglu
 
-##<<<<<<< HEAD
-    Construction::
-
-        PotentialEvapotranspiration(grid, method='Cosine',
-            Priestley_taylor_const=1.26, albedo=0.6,
-            latent_heat_of_vaporization=28.34, psychometric_const=0.066,
-            stefan_boltzmann_const=0.0000000567, solar_const=1366.67,
-            latitude=34., elevation_of_measurement=300, adjustment_coeff=0.18,
-            lt=0., nd=365., MeanTmaxF=12., delta_d=5., **kwds)
-
-    Parameters
-    ----------
-    grid: RasterModelGrid
-        A grid.
-    method: {'Constant', 'PriestleyTaylor', 'MeasuredRadiationPT',
-             'Cosine'}, optional
-        Priestley Taylor method will spit out radiation outputs too.
-    priestley_taylor_constant: float, optional
-        Alpha used in Priestley Taylor method.
-    albedo: float, optional
-        Albedo.
-    latent_heat_of_vaporization: float, optional
-        Latent heat of vaporization for water Pwhv (Wd/(m*mm^2)).
-    psychometric_const: float, optional
-        Psychometric constant (kPa (deg C)^-1).
-    stefan_boltzmann_const: float, optional
-        Stefan Boltzmann's constant (W/(m^2K^-4)).
-    solar_const: float, optional
-        Solar constant (W/m^2).
-    latitude: float, optional
-        Latitude (radians).
-    elevation_of_measurement: float, optional
-        Elevation at which measurement was taken (m).
-    adjustment_coeff: float, optional
-        adjustment coeff to predict Rs from air temperature (deg C)^-0.5.
-    lt: float, optional
-        lag between peak TmaxF and solar forcing (days).
-    nd: float, optional
-        Number of days in year (days).
-    MeanTmaxF: float, optional
-        Mean annual rate of TmaxF (mm/d).
-    delta_d: float, optional
-        Calibrated difference between max & min daily TmaxF (mm/d).
-    zm: float, required only for method(s): PenmanMonteith
-        Wind speed anemometer height (m).
-    zh: float, required only for method(s): PenmanMonteith (with
-    correction for Relative Humidity measurement elevation)
-        Relative Humidity measurement height (m).
-    rl: float, required only for method(s): PenmanMonteith
[…]
