### 2019-12-25 afd9fd9
resolved conflicts for SM component
resolved conflicts for SoilMoisture component
