+++ landlab/components/soil_moisture/soil_moisture_dynamics.py
@@ -1,16 +1,10 @@
-##<<<<<<< HEAD
-##=======
-import numpy as np
-##>>>>>>> origin/v1
 
+import numpy as np
+import six
 from landlab import Component
 
 from ...utils.decorators import use_file_name_or_kwds
-##<<<<<<< HEAD
-import numpy as np
-import six
-##=======
-##>>>>>>> origin/v1
+
 
 _VALID_METHODS = set(["Grid", "Multi"])
 
@@ -32,73 +26,12 @@ class SoilMoisture(Component):
     component requires fields of initial soil moisture, rainfall input (if
     any), time to the next storm and potential transpiration.
 
-##<<<<<<< HEAD
-    Construction::
-
-        SoilMoisture(grid, runon=0., f_bare=0.7, soil_ew=0.1,
-           intercept_cap_grass= 1., zr_grass=0.3, I_B_grass=20.,
-           I_V_grass=24., K_s_grass=42., pc_grass=0.43, fc_grass=0.56,
-           sc_grass=0.33, wp_grass=0.13, hgw_grass=0.1, beta_grass=13.8,
-           LAI_max_grass=2., LAIR_max_grass=2.88,
-           intercept_cap_shrub=1.5, zr_shrub=0.5, I_B_shrub=20.,
-           I_V_shrub=40., K_s_shrub=42., pc_shrub=0.43, fc_shrub=0.56,
-           sc_shrub=0.24, wp_shrub=0.13, hgw_shrub=0.1, beta_shrub=13.8,
-           LAI_max_shrub=2., LAIR_max_shrub=2.,
-           intercept_cap_tree=2., zr_tree=1.3, I_B_tree=20.,
-           I_V_tree=40., K_s_tree=42., pc_tree=0.43, fc_tree=0.56,
-           sc_tree=0.22, wp_tree=0.15, hgw_tree=0.1, beta_tree=13.8,
-           LAI_max_tree=4., LAIR_max_tree=4.,
-           intercept_cap_bare=1., zr_bare=0.15, I_B_bare=20.,
-           I_V_bare=20., K_s_bare=42., pc_bare=0.43, fc_bare=0.56, sc_bare=0.33,
-           wp_bare=0.13, hgw_bare=0.1, beta_bare=13.8,
-           LAI_max_bare=0.01, LAIR_max_bare=0.01)
-
-    Parameters
-    ----------
-    grid: RasterModelGrid
-        A grid.
-    runon: float, optional
-        Runon from higher elevation (mm).
-    f_bare: float, optional
-        Fraction to partition PET for bare soil (None).
-    soil_ew: float, optional
-        Residual Evaporation after wilting (mm/day).
-    intercept_cap: float, optional
-        Plant Functional Type (PFT) specific full canopy interception
-        capacity.
-    zr: float, optional
-        Root depth (m).
-    I_B: float, optional
-        Infiltration capacity of bare soil (mm/h).
-    I_V: float, optional
-        Infiltration capacity of vegetated soil (mm/h).
-    K_s: float, optional
-        Hydraulic conductivity of soil (mm/h).
-    pc: float, optional
-        Soil porosity (None).
-    fc: float, optional
-        Soil saturation degree at field capacity (None).
-    sc: float, optional
-        Soil saturation degree at stomatal closure (None).
-    wp: float, optional
-        Soil saturation degree at wilting point (None).
-    hgw: float, optional
-        Soil saturation degree at hygroscopic point (None).
-    beta: float, optional
-        Deep percolation constant = 2*b+3 where b is
-        water retention (None).
-    LAI_max: float, optional
-        Maximum leaf area index (m^2/m^2).
-    LAIR_max: float, optional
[…]
