### 2019-12-25 9c5bcab
resolved conflicts for Vegetation component
resolved conflicts for Vegetation component
