### 2019-12-26 de4dd26
resolved conflicts for plant_competition component
- resolved conflicts
- minor style edits
