+++ landlab/components/plant_competition_ca/plant_competition_ca.py
@@ -1,7 +1,6 @@
+
 import numpy as np
-
 from landlab import Component
-
 from ...utils.decorators import use_file_name_or_kwds
 
 _VALID_METHODS = set(["Grid"])
@@ -187,7 +186,6 @@ class VegCA(Component):
         assert_method_is_valid(self._method)
 
         super(VegCA, self).__init__(grid)
-##<<<<<<< HEAD
         """
         Note: If the field 'vegetation__plant_functional_type' is not
         input, a random field will be generated.
@@ -222,13 +220,6 @@ class VegCA(Component):
             VegType[locs_shrubs[tp[
                 locs_shrubs] < self._tpmax_sh_s]] = SHRUBSEEDLING
             grid.at_cell['plant__age'] = tp
-##=======
-
-        if "vegetation__plant_functional_type" not in self.grid.at_cell:
-            grid["cell"]["vegetation__plant_functional_type"] = np.random.randint(
-                0, 6, grid.number_of_cells
-            )
-##>>>>>>> origin/v1
 
         for name in self._input_var_names:
             if name not in self.grid.at_cell:
@@ -240,25 +231,7 @@ class VegCA(Component):
 
         self._cell_values = self.grid["cell"]
 
-##<<<<<<< HEAD
     def update(self, time_elapsed=1, edit_vegcov=False):
-##=======
-        VegType = grid["cell"]["vegetation__plant_functional_type"]
-        tp = np.zeros(grid.number_of_cells, dtype=int)
-        tp[VegType == TREE] = np.random.randint(
-            0, self._tpmax_tr, np.where(VegType == TREE)[0].shape
-        )
-        tp[VegType == SHRUB] = np.random.randint(
-            0, self._tpmax_sh, np.where(VegType == SHRUB)[0].shape
-        )
-        locs_trees = np.where(VegType == TREE)[0]
-        locs_shrubs = np.where(VegType == SHRUB)[0]
-        VegType[locs_trees[tp[locs_trees] < self._tpmax_tr_s]] = TREESEEDLING
-        VegType[locs_shrubs[tp[locs_shrubs] < self._tpmax_sh_s]] = SHRUBSEEDLING
-        grid["cell"]["plant__age"] = tp
-
-    def update(self, time_elapsed=1, Edit_VegCov=True):
-##>>>>>>> origin/v1
         """
         Update fields with current loading conditions.
 
@@ -271,24 +244,16 @@ class VegCA(Component):
             will be output, (i.e.) if a cell is vegetated the corresponding
             cell of the field will be 1, otherwise it will be 0.
         """
-##<<<<<<< HEAD
         self._VegType = self._cell_values['vegetation__plant_functional_type']
         self._CumWS = self._cell_values[
             'vegetation__cumulative_water_stress']
         self._live_index = self._cell_values['plant__live_index']
         self._tp = self._cell_values['plant__age'] + time_elapsed
 #        self._tp_check_1 = np.copy(self._tp)
-##=======
-        self._VegType = self._cell_values["vegetation__plant_functional_type"]
-        self._CumWS = self._cell_values["vegetation__cumulative_water_stress"]
-        self._live_index = self._cell_values["plant__live_index"]
-        self._tp = self._cell_values["plant__age"] + time_elapsed
-##>>>>>>> origin/v1
 
         # Check if shrub and tree seedlings have matured
         shrub_seedlings = np.where(self._VegType == SHRUBSEEDLING)[0]
         tree_seedlings = np.where(self._VegType == TREESEEDLING)[0]
-##<<<<<<< HEAD
[…]
