### 2019-12-27 6b43905
doc-tests fixed for radiation comonent
- fixed a couple of things in /components
- removed space in the Radiation component
