### 2019-12-27 f8cb09b
v2 update
now the tests folder and the main component file pass tests (other files don't necessarily pass)
