+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -1,18 +1,18 @@
 #!/usr/env/python
 
-"""    
+"""
 Landlab component that simulates the transport of bed material
-sediment through a 1-D river network, while tracking the resulting changes 
-in bed material grain size and river bed elevation. Model framework 
-described in Czuba (2018). Additions include: particle abrasion, variable 
-active layer thickness (Wong et al., 2007). 
+sediment through a 1-D river network, while tracking the resulting changes
+in bed material grain size and river bed elevation. Model framework
+described in Czuba (2018). Additions include: particle abrasion, variable
+active layer thickness (Wong et al., 2007).
 
 Fixes that need to happen:
 
     -- Check abrasion exponent units (per km or per m?)
 
     -- Channel width-- why is this anything other than a link attribute??
-    
+
     -- Flow depth -- double check to make sure we're going through flow depth array correctly
 
     -- JC: I found two items that I think should be changed in the _calc_transport_wilcock_crowe and I made these changes
@@ -35,28 +35,28 @@ Last edit ---
 import numpy as np
 
 # %% Import Libraries
-from landlab import BAD_INDEX_VALUE, Component
+from landlab import Component
+from landlab.components import FlowDirectorSteepest
 from landlab.data_record import DataRecord
 from landlab.grid.network import NetworkModelGrid
-from landlab.components import FlowDirectorSteepest
-from landlab.utils.decorators import use_file_name_or_kwds
 
 _SUPPORTED_TRANSPORT_METHODS = ["WilcockCrowe"]
 
-_OUT_OF_NETWORK = BAD_INDEX_VALUE - 1
+_OUT_OF_NETWORK = NetworkModelGrid.BAD_INDEX - 1
 
 
 _ACTIVE = 1
 _INACTIVE = 0
 
+
 class NetworkSedimentTransporter(Component):
     """
     Landlab component that simulates the transport of bed material
-    sediment through a 1-D river network, while tracking the resulting changes 
-    in bed material grain size and river bed elevation. Model framework 
-    described in Czuba (2018). Additions include: particle abrasion, variable 
-    active layer thickness (Wong et al., 2007). 
-    
+    sediment through a 1-D river network, while tracking the resulting changes
+    in bed material grain size and river bed elevation. Model framework
+    described in Czuba (2018). Additions include: particle abrasion, variable
+    active layer thickness (Wong et al., 2007).
+
 
     **Usage:**
     Option 1 - Basic::
@@ -68,145 +68,136 @@ class NetworkSedimentTransporter(Component):
                              )
 
     Examples
-    ----------  
+    ----------
     >>> import numpy as np
     >>> from landlab.components import FlowDirectorSteepest, NetworkSedimentTransporter
-    >>> from landlab.grid.network import NetworkModelGrid
-    >>> from landlab import BAD_INDEX_VALUE
+    >>> from landlab import NetworkModelGrid
     >>> from landlab.data_record import DataRecord
-    >>> _OUT_OF_NETWORK = BAD_INDEX_VALUE - 1
-    
-    The NetworkSedimentTransporter moves "parcels" of sediment down a network 
+    >>> _OUT_OF_NETWORK = NetworkModelGrid.BAD_INDEX - 1
+
+    The NetworkSedimentTransporter moves "parcels" of sediment down a network
     based on a given flow and a given sediment transport formulation. The river
     network is represented by a landlab NetworkModelGrid. Flow direction in the
-    network is determined using a landlab flow director. Sediment parcels are 
-    represented as items within a landlab ``DataRecord``. 
-    
-    Create a ``NetworkModelGrid`` to represent the river channel network. In this 
-    case, the grid is a single line of 4 nodes connected by 3 links. Each link represents a reach of river. 
[…]

[…]
