### 2019-12-31 07141f7
tests pass for PET component
- fixed styling for PET component
- added documentation
- tests pass
