+++ landlab/components/pet/potential_evapotranspiration_field.py
@@ -20,17 +20,83 @@ class PotentialEvapotranspiration(Component):
     """
     Potential Evapotranspiration Component calculates spatially distributed
     potential evapotranspiration based on input radiation factor (spatial
-    distribution of incoming radiation) using chosen method such as constant
-    or Priestley Taylor. Ref: Xiaochi et. al. 2013 for 'Cosine' method and
-    ASCE-EWRI Task Committee Report Jan 2005 for 'PriestleyTaylor' method.
+    distribution of incoming radiation) using one of the five ways
+    to calculate potential evapotranspiration. The methods
+    are 'Constant','Cosine', 'MeasuredRadiationPT',
+    'PenmanMonteith', and 'PriestleyTaylor'.
 
-    Note: Calling 'PriestleyTaylor' method would generate/overwrite shortwave &
-    longwave radiation fields.
+    Note: Calling 'PriestleyTaylor' or 'PenmanMonteith' methods
+    will overwrite the following fields:
+        "radiation__incoming_shortwave_flux",
+        "radiation__net_shortwave_flux",
+        "radiation__net_longwave_flux", and
+        "radiation__net_flux".
+    
+    As mentioned above, this component can be configured to use
+    one of the five ways to calculate potential evapotranspiration.
+    The methods are 'Constant','Cosine', 'MeasuredRadiationPT',
+    'PenmanMonteith', and 'PriestleyTaylor'.
+
+    'Constant' method: Receives the user-input const_potential_evapotranspiration
+    during a call to the 'update' method. This value is
+    then distributed spatially using the 'radiation__ratio_to_flat_surface'
+    field.
+
+    'Cosine' method: This method calculates the flat surface potential
+    evapotranspiration value as a sinusoidal function of the day of the year.
+    This method uses the inputs of MeanTmaxF, delta_d, lt, and nd
+    obtained during instantiation of the component or during a call
+    to the 'initialize' method, and current_time during a call to the
+    'update' method. This value is then distributed spatially using the
+    'radiation__ratio_to_flat_surface' field. [Equation 28 - Zhou et al. 2013]
+
+    'MeasuredRadiationPET' method: This method calculates the flat surface
+    potential evapotranspiration value using the inputs of
+    obs_radiation, albedo, and Tavg obtained during the call to
+    'update' method. This value is then distributed spatially
+    using the 'radiation__ratio_to_flat_surface' field.
+    [ASCE-EWRI Task Committee Report Jan 2005]
+
+    'PenmanMonteith' method: This method calculates the flat surface
+    potential evapotranspiration values using Penman equation.
+    This method uses the inputs of zm, zh, zveg, rl,
+    air_density, latent_heat_of_vaporization,
+    stefan_boltzmann_const, and albedo obtained during
+    instantiation of the component or during a call
+    to the 'initialize' method, and Tavg, radiation_sw,
+    wind_speed, relative_humidity, and co2_concentration
+    during a call to the 'update' method. This value is then
+    distributed spatially using the
+    'radiation__ratio_to_flat_surface' field.[Penman 1948]
+
+    'PriestleyTaylor' method: This method calculates the flat surface
+    potential evapotranspiration values using Priestley-Taylor
+    equation. This method uses the inputs of
+    stefan_boltzmann_const, latitude, elevation_of_measurement,
+    adjustment_coeff, priestley_taylor_constant, albedo,
+    psychometric_const, and latent_heat_of_vaporization obtained during
+    instantiation of the component or during a call
+    to the 'initialize' method, and current_time, Tmax,
[…]
