### 2019-12-31 b61ce40
tests pass for Soil Moisture component
- fixed doc tests and unit tests for Soil Moisture component
