+++ landlab/components/soil_moisture/soil_moisture_dynamics.py
@@ -45,6 +45,7 @@ class SoilMoisture(Component):
      'soil_moisture__saturation_fraction',
      'surface__evapotranspiration',
      'surface__runoff',
+     'surface__runon',
      'vegetation__water_stress']
     >>> sorted(SoilMoisture.units) # doctest: +NORMALIZE_WHITESPACE
     [('rainfall__daily_depth', 'mm'),
@@ -53,12 +54,14 @@ class SoilMoisture(Component):
      ('soil_moisture__saturation_fraction', 'None'),
      ('surface__evapotranspiration', 'mm'),
      ('surface__potential_evapotranspiration_rate', 'mm'),
+     ('surface__potential_evapotranspiration_rate__grass', 'mm'),
      ('surface__runoff', 'mm'),
+     ('surface__runon', 'mm'),
      ('vegetation__cover_fraction', 'None'),
      ('vegetation__live_leaf_area_index', 'None'),
      ('vegetation__plant_functional_type', 'None'),
      ('vegetation__water_stress', 'None')]
-    >>> grid['cell']['vegetation__plant_functional_type']= (
+    >>> grid.at_cell["vegetation__plant_functional_type"] = (
     ...            np.zeros(grid.number_of_cells, dtype=int))
     >>> SM = SoilMoisture(grid)
     >>> SM.grid.number_of_cell_rows
@@ -68,110 +71,110 @@ class SoilMoisture(Component):
     >>> SM.grid is grid
     True
     >>> import numpy as np
-    >>> np.allclose(grid.at_cell['soil_moisture__saturation_fraction'], 0.)
+    >>> np.allclose(grid.at_cell["soil_moisture__saturation_fraction"], 0.)
     True
-    >>> grid['cell']['surface__potential_evapotranspiration_rate']= np.array([
+    >>> grid.at_cell["surface__potential_evapotranspiration_rate"]= np.array([
     ...            0.2554777, 0.2554777 , 0.22110221, 0.22110221,
     ...            0.24813062, 0.24813062])
-    >>> grid['cell']['soil_moisture__initial_saturation_fraction']= (
+    >>> grid.at_cell["soil_moisture__initial_saturation_fraction"]= (
     ...        0.75 * np.ones(grid.number_of_cells))
-    >>> grid['cell']['vegetation__live_leaf_area_index']= (
+    >>> grid.at_cell["vegetation__live_leaf_area_index"]= (
     ...        2. * np.ones(grid.number_of_cells))
-    >>> grid['cell']['vegetation__cover_fraction']= (
+    >>> grid.at_cell["vegetation__cover_fraction"]= (
     ...        np.ones(grid.number_of_cells))
     >>> current_time = 0.5
-    >>> grid['cell']['rainfall__daily_depth'] = (
+    >>> grid.at_cell["rainfall__daily_depth"] = (
     ...        25. * np.ones(grid.number_of_cells))
     >>> current_time = SM.update(current_time)
-    >>> np.allclose(grid.at_cell['soil_moisture__saturation_fraction'], 0.)
+    >>> np.allclose(grid.at_cell["soil_moisture__saturation_fraction"], 0.)
     False
     """
 
     _name = "Soil Moisture"
 
     _input_var_names = (
-        'vegetation__cover_fraction',
-        'vegetation__live_leaf_area_index',
-        'surface__potential_evapotranspiration_rate',
-        'surface__potential_evapotranspiration_rate__grass',
-        'soil_moisture__initial_saturation_fraction',
-        'vegetation__plant_functional_type',
-        'rainfall__daily_depth',
+        "vegetation__cover_fraction",
+        "vegetation__live_leaf_area_index",
+        "surface__potential_evapotranspiration_rate",
[…]
