### 2020-01-01 82ae968
two more components are tested
- tests for vegetation_dynamics and plant_competition are passing
- added unit tests for resource_redistribution
- need to add documentation and doctests for resource_redistribution
