+++ landlab/components/resource_redistribution/resource_redistribution.py
@@ -34,9 +34,6 @@ class ResourceRedistribution(Component):
     What does this component output? Are there multiple
     processes? What are the key methods?
 
-    Construction::
-        ResourceRedistribution()
-
     Parameters:
     ----------
     grid: RasterModelGrid
@@ -75,35 +72,33 @@ class ResourceRedistribution(Component):
     Deal with them!
     """
 
-    _name = 'SpatialDisturbance'
-
-    __version__ = 'landlab version in which this component is created in'
+    _name = "Resource Redistribution"
 
     _input_var_names = (
-            'vegetation__plant_functional_type',
-            'soil__resources',
+            "vegetation__plant_functional_type",
+            "soil__resources",
             )
 
     _output_var_names = (
-            'vegetation__plant_functional_type',
-            'soil__resources',
+            "vegetation__plant_functional_type",
+            "soil__resources",
             )
 
     _var_units = {
-            'vegetation__plant_functional_type': None,
-            'soil__resources': None,
+            "vegetation__plant_functional_type": "None",
+            "soil__resources": "None",
             }
 
     _var_mapping = {
-            'vegetation__plant_functional_type': 'cell',
-            'soil__resources': 'cell',
+            "vegetation__plant_functional_type": "cell",
+            "soil__resources": "cell",
             }
 
     _var_doc = {
-            'vegetation__plant_functional_type':
-            'classification of plant type - bare=0, grass=1,' +
-            'shrub=2, burntgrass=3, burntshrub=4',
-            'soil__resources': 'level of soil resources',
+            "vegetation__plant_functional_type":
+                "classification of plant type - bare=0, grass=1,"
+                + "shrub=2, burntgrass=3, burntshrub=4",
+            "soil__resources": "level of soil resources",
             }
 
     @use_file_name_or_kwds
@@ -119,18 +114,18 @@ class ResourceRedistribution(Component):
                  **kwds):
         super(ResourceRedistribution, self).__init__(grid, **kwds)
 
-        name = 'vegetation__plant_functional_type'
+        name = "vegetation__plant_functional_type"
         if name not in self.grid.at_cell:
-            print('Since a cellular field of PFTs ' +
-                  'is not provided, the field ' +
-                  name + ' is initialized to 0 - all bare cells!')
-            self.grid.add_zeros('cell', name, units=self._var_units[name])
-        name = 'soil__resources'
+            print("Since a cellular field of PFTs " +
+                  "is not provided, the field " +
+                  name + " is initialized to 0 - all bare cells!")
+            self.grid.add_zeros("cell", name, units=self._var_units[name])
+        name = "soil__resources"
         if name not in self.grid.at_cell:
-            print('Since a cellular field for Resources ' +
-                  'is not provided, the field ' +
-                  name + ' is initialized to 1!')
-            self.grid.add_ones('cell', name, units=self._var_units[name])
+            print("Since a cellular field for Resources " +
+                  "is not provided, the field " +
+                  name + " is initialized to 1!")
+            self.grid.add_ones("cell", name, units=self._var_units[name])
 
         self.initialize(e=e, R_low_threshold=R_low_threshold,
                         R_threshold=R_threshold,
@@ -182,8 +177,8 @@ class ResourceRedistribution(Component):
         self._gr_mor_ws_thresh = gr_mor_ws_thresh
 
[…]
