### 2020-01-01 0f01d57
Added unit tests for spatial disturbance
- added unit tests for spatial_disturbance
- need to add documentation and doctests for spatial_disturbance
