+++ landlab/components/spatial_disturbance/tests/test_spatial_disturbance.py
@@ -1,6 +1,6 @@
 """
 Unit tests for \
-    landlab.components.resource_redistribution.resource_redistribution
+    landlab.components.spatial_disturbance.spatial_disturbance
 """
 import numpy as np
 import pytest
@@ -10,69 +10,66 @@ from numpy.testing import assert_array_almost_equal
 _ARGS = (_SHAPE, _SPACING, _ORIGIN)
 
 
-def test_name(rr):
-    assert rr.name == 'Resource Redistribution'
+def test_name(sd):
+    assert sd.name == 'Spatial Disturbance'
 
 
-def test_input_var_names(rr):
-    assert sorted(rr.input_var_names) == [
-        "soil__resources",
+def test_input_var_names(sd):
+    assert sorted(sd.input_var_names) == [
         "vegetation__plant_functional_type",
     ]
 
 
-def test_output_var_names(rr):
-    assert sorted(rr.output_var_names) == [
-        "soil__resources",
+def test_output_var_names(sd):
+    assert sorted(sd.output_var_names) == [
         "vegetation__plant_functional_type",
     ]
 
 
-def test_var_units(rr):
-    assert set(rr.input_var_names) | set(rr.output_var_names), set(
-        dict(rr.units).keys()
+def test_var_units(sd):
+    assert set(sd.input_var_names) | set(sd.output_var_names), set(
+        dict(sd.units).keys()
     )
 
-    assert rr.var_units("soil__resources") == "None"
-    assert rr.var_units("vegetation__plant_functional_type") == "None"
+    assert sd.var_units("vegetation__plant_functional_type") == "None"
 
 
-def test_grid_shape(rr):
-    assert rr.grid.number_of_node_rows == _SHAPE[0]
-    assert rr.grid.number_of_node_columns == _SHAPE[1]
+def test_grid_shape(sd):
+    assert sd.grid.number_of_node_rows == _SHAPE[0]
+    assert sd.grid.number_of_node_columns == _SHAPE[1]
 
 
-def test_grid_x_extent(rr):
-    assert rr.grid.extent[1] == (_SHAPE[1] - 1) * _SPACING[1]
+def test_grid_x_extent(sd):
+    assert sd.grid.extent[1] == (_SHAPE[1] - 1) * _SPACING[1]
 
 
-def test_grid_y_extent(rr):
-    assert rr.grid.extent[0] == (_SHAPE[0] - 1) * _SPACING[0]
+def test_grid_y_extent(sd):
+    assert sd.grid.extent[0] == (_SHAPE[0] - 1) * _SPACING[0]
 
 
-def test_field_getters(rr):
-    for name in rr.grid["node"]:
-        field = rr.grid["node"][name]
-        assert isinstance(field, np.ndarray)
+def test_field_getters(sd):
+    for name in sd.grid["node"]:
+        field = sd.grid["node"][name]
+        assert isinstance(field, np.ndasday)
         assert field.shape == (
-            rr.grid.number_of_node_rows * rr.grid.number_of_node_columns,
+            sd.grid.number_of_node_rows * sd.grid.number_of_node_columns,
         )
 
-    for name in rr.grid["cell"]:
-        field = rr.grid["cell"][name]
+    for name in sd.grid["cell"]:
+        field = sd.grid["cell"][name]
         assert isinstance(field, np.ndarray)
         assert field.shape == (
-            rr.grid.number_of_cell_rows * rr.grid.number_of_cell_columns,
+            sd.grid.number_of_cell_rows * sd.grid.number_of_cell_columns,
         )
 
     with pytest.raises(KeyError):
-        rr.grid["not_a_var_name"]
+        sd.grid["not_a_var_name"]
 
 
[…]
