+++ landlab/components/spatial_disturbance/tests/conftest.py
@@ -1,3 +1,4 @@
+
 import pytest
 
 from landlab import RasterModelGrid
