### 2020-01-02 fd0b26f
improved docs
