+++ landlab/components/species_evolution/zone.py
@@ -2,19 +2,29 @@
 # -*- coding: utf-8 -*-
 """Zone functions and class of SpeciesEvolver."""
 from collections import OrderedDict
+from enum import IntEnum, unique
 
 import numpy as np
 
 
-# Define connection types.
+@unique
+class Connection(IntEnum):
+    """Zone connection type.
 
-_NONE_TO_NONE = 'none_to_none'
-_NONE_TO_ONE = 'none_to_one'
-_ONE_TO_NONE = 'one-to-none'
-_ONE_TO_ONE = 'one-to-one'
-_ONE_TO_MANY = 'one-to-many'
-_MANY_TO_ONE = 'many-to-one'
-_MANY_TO_MANY = 'many-to-many'
+    Connection type represents the connectivity of zones in two time steps. It
+    is described with the pattern, x-to-y where x and y are the descriptive
+    counts (none, one, or many) of zone(s) at the earlier and later time step,
+    respectively. For example, a connection type of one-to-many is a zone in
+    the earlier time step that spatially overlaps multiple zones in the later
+    time step.
+    """
+    NONE_TO_NONE = 0
+    NONE_TO_ONE = 1
+    ONE_TO_NONE = 2
+    ONE_TO_ONE = 3
+    ONE_TO_MANY = 4
+    MANY_TO_ONE = 5
+    MANY_TO_MANY = 6
 
 
 def _update_zones(grid, prior_zones, new_zones, record):
@@ -27,12 +37,7 @@ def _update_zones(grid, prior_zones, new_zones, record):
     step zones referred to as the predecessor zones.
 
     The type of connection between predecessor and successor zones is described
-    by the number of predecessors and successors that overlap each other. The
-    connection type is represented by a string with the pattern, x-to-y where x
-    and y are the descriptive counts (none, one, or many) of zone(s) at the
-    prior and current time step, respectively. For example, a connection type
-    of one-to-many is a zone in the prior time step that spatially overlaps
-    multiple zones in the current time step.
+    ``Connection``. The `_conn_type` property of zones are set by this method.
 
     In the `many` connections, a rule determines which of the prior zones
     persist as the zone in the current time step. The zone with the greatest
@@ -62,17 +67,10 @@ def _update_zones(grid, prior_zones, new_zones, record):
     capture_ct = 0
     area_captured = [0]
 
-    # Get `successors`, the zones of `time`.
+    # Get `successors`.
 
     if len(prior_zones) == 0 and len(new_zones) == 0:
         successors = []
-
-    elif len(prior_zones) > 0 and len(new_zones) == 0:
-        successors = []
-        conn_type = _determine_connection_type(1, 0)
-        for p in prior_zones:
-            p._conn_type = conn_type
-            p._successors = []
 
     elif len(prior_zones) == 0 and len(new_zones) > 0:
         successors = new_zones
@@ -80,6 +78,13 @@ def _update_zones(grid, prior_zones, new_zones, record):
         for n in new_zones:
             n._conn_type = conn_type
             n._successors = [n]
+
+    elif len(prior_zones) > 0 and len(new_zones) == 0:
+        successors = []
+        conn_type = _determine_connection_type(1, 0)
+        for p in prior_zones:
+            p._conn_type = conn_type
+            p._successors = []
 
     elif len(prior_zones) > 0 and len(new_zones) > 0:
         successors = []
@@ -106,6 +111,10 @@ def _update_zones(grid, prior_zones, new_zones, record):
             ps_i_ns = _intersecting_zones(ns_mask, ps_index_map, prior_zones)
             ps_i_ns_ct = len(ps_i_ns)
 
+            if ps_i_ns_ct == 0:
+                ps_i_ns_ct = 1
+                ps_i_ns = p
+
[…]
