+++ landlab/components/species_evolution/__init__.py
@@ -1,12 +1,12 @@
 from .species_evolution import SpeciesEvolver
 from .zone import Zone
-from .zone_species import ZoneSpecies
+from .zone_taxon import ZoneTaxon
 from .zone_controller import ZoneController
 
 
 __all__ = [
     'SpeciesEvolver',
     'Zone',
-    'ZoneSpecies',
+    'ZoneTaxon',
     'ZoneController'
 ]
