+++ docs/source/reference/components/species_evolution.rst
@@ -1,0 +1,7 @@
+SpeciesEvolver: Evolve life in a landscape
+------------------------------------------
+
+.. automodule:: landlab.components.species_evolution
+    :members:
+    :undoc-members:
+    :show-inheritance:

[…]
