### 2020-01-02 c543d02
quick test to make sure I have updated everything
I pulled the new changes to landlab. It looks like everything is in the same location except for the paper which I have found. I added a comment to the paper and the the NST_setup to make sure everything looks good from my end.
