### 2020-01-02 4944f8a
intermediate updates
- this is work in progress
- added documentation to ResourceRedistribution
TODO
- add more documentation
- add doctests
