+++ landlab/components/resource_redistribution/resource_redistribution.py
@@ -9,11 +9,9 @@ Sujith Ravi's 2009 Landscape Ecology paper.
 0: Bare; 1: Grass; 2: Shrub; 3: Burnt Grass; 4: Burnt Shrub;
 """
 
-#%% Import Packages
 import numpy as np
 from landlab import Component
 from ...utils.decorators import use_file_name_or_kwds
-#from __future__ import print_function
 
 
 #%% Declare Global Variables (If any)
@@ -24,7 +22,7 @@ SHRUB = 2
 BURNTGRASS = 3
 BURNTSHRUB = 4
 
-#%%
+
 class ResourceRedistribution(Component):
     """
     What is this component (brief description)?
@@ -33,31 +31,6 @@ class ResourceRedistribution(Component):
 
     What does this component output? Are there multiple
     processes? What are the key methods?
-
-    Parameters:
-    ----------
-    grid: RasterModelGrid
-        grid, Landlab's RasterModelGrid object
-    V: array_like
-        Vegetation Plant Functional Type; shape = [grid.number_of_cells]
-        BARE = 0; GRASS = 1; SHRUB = 2; BURNTGRASS = 3; BURNTSHRUB = 4;
-        TREE = 5; BURNTTREE = 6; SHRUBSEED = 7; TREESEED = 8
-    n_fires: int, optional
-        Number of fires to be created
-    fire_area_mean: float, optional
-        mean area of uniform distribution to sample fire size
-    fire_area_dev: float, optional
-        standard deviation of uniform distribution to sample fire size
-    sh_susc: float, optional
-        susceptibility of shrubs to burn
-    tr_susc: float, optional
-        susceptibility of trees to burn
-    gr_susc: float, optional
-        susceptibility of grass to burn
-    sh_seed_susc: float, optional
-        susceptibility of shrub seedlings to burn
-    tr_seed_susc: float, optional
-        susceptibility of tree seedlings to burn
 
     Examples
     --------
@@ -102,16 +75,86 @@ class ResourceRedistribution(Component):
             }
 
     @use_file_name_or_kwds
-    def __init__(self, grid, e=0.1, R_low_threshold=-2.,
-                 R_threshold=2., R_dep_threshold=1.,
-                 Rth_gr=0.4, Rth_sh=0.8, P_gr_regrwth=0.25,
-                 P_sh_regrwth=0.25, Pen=0.05, Pgrz=0.01,
-                 P_gr=0.5, sh_max_age=600, sh_seedling_max_age=18,
-                 sh_seedling_mor_dis=0., sh_mor_dis_low_thresh_age=300,
-                 sh_mor_dis_low_slp=0.01, sh_mor_dis_high_slp=0.99,
-                 P_sh_fire_mor=0.75, P_gr_fire_mor=1.,
-                 sh_mor_ws_thresh=0.01, gr_mor_ws_thresh=0.08,
-                 **kwds):
+    def __init__(
+         self,
+         grid,
+         e=0.1,
+         R_low_threshold=-2.,
+         R_threshold=2.,
+         R_dep_threshold=1.,
+         Rth_gr=0.4,
+         Rth_sh=0.8,
+         P_gr_regrwth=0.25,
+         P_sh_regrwth=0.25,
+         Pen=0.05,
+         Pgrz=0.01,
+         P_gr=0.5,
+         sh_max_age=600,
+         sh_seedling_max_age=18,
+         sh_seedling_mor_dis=0.,
+         sh_mor_dis_low_thresh_age=300,
+         sh_mor_dis_low_slp=0.01,
+         sh_mor_dis_high_slp=0.99,
+         P_sh_fire_mor=0.75,
+         P_gr_fire_mor=1.,
+         sh_mor_ws_thresh=0.01,
[…]
