+++ docs/landlab.components.species_evolution.rst
@@ -1,7 +1,0 @@
-SpeciesEvolver: Evolve species in a landscape
----------------------------------------------
-
-.. automodule:: landlab.components.species_evolution
-    :members:
-    :undoc-members:
-    :show-inheritance:
