+++ landlab/components/species_evolution/__init__.py
@@ -1,4 +1,10 @@
-from .species_evolution import SpeciesEvolver
+#!/usr/bin/env python
+"""
+.. codeauthor:: Nathan Lyons <nlyons@tulane.edu>
+
+.. sectionauthor:: Nathan Lyons <nlyons@tulane.edu>
+"""
+from .species_evolver import SpeciesEvolver
 from .zone import Zone
 from .zone_taxon import ZoneTaxon
 from .zone_controller import ZoneController
