+++ docs/source/reference/components/species_evolution.rst
@@ -1,7 +1,23 @@
 SpeciesEvolver: Evolve life in a landscape
 ------------------------------------------
 
-.. automodule:: landlab.components.species_evolution
+.. automodule:: landlab.components.species_evolution.species_evolver
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+ZoneController: Controls zones and populates them with taxa
+-----------------------------------------------------------
+
+.. automodule:: landlab.components.species_evolution.zone_controller
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+ZoneTaxon: A zone-based taxon
+-----------------------------
+
+.. automodule:: landlab.components.species_evolution.zone_taxon
     :members:
     :undoc-members:
     :show-inheritance:
