### 2020-01-03 0452c41
JOSS updates
I have finished revising the JOSS paper for now.
