+++ joss/in_preparation/network_sediment_transporter/paper.md
@@ -37,33 +37,32 @@ bibliography: papers.bib
 
 # Summary
 
-# Jan. 2 quick change to test Czuba updates
-Coarse sediment moves downstream through river networks. The transport rate of any particular sediment grain on the river bed surface is a function of both the hydraulics of that reach of river and the size distribution of the other grains in the reach. As sediment moves through a river system, grains may be deposited or eroded, burying and exposing other grains, and in the process changing the elevation and slope of each segment of river. This process of river channel evolution through the process of sediment transport is referred to as morphodynamics [@cite]. Computational morphodynamic models allow for the prediction of sediment pulse transport, such as that which occurs after dam removal [@citeCui] or landsliding events[@An2017; @Benda&Dunne1997], as well as the prediction of changes in river channel bed surface grain size[@cite].
+Coarse sediment (sand, gravel, and cobbles) moves downstream through river networks. The transport rate of any particular sediment grain on the river bed surface is a function of both the hydraulics of that reach of river and the size distribution of the other grains in the reach. As sediment moves through a river system, grains may be deposited or eroded, burying and exposing other grains, and in the process changing the elevation and slope of each segment of river. This process of river channel evolution through the process of sediment transport is referred to as morphodynamics [@ParkerEbook,http://hydrolab.illinois.edu/people/parkerg/morphodynamics_e-book.htm; @Church&Ferguson2015]. Computational morphodynamic models allow for the prediction of sediment pulse transport, such as that which occurs after dam removal [@Cuietal2006a; @Cuietal2006b; @Cui2007a] or landsliding events [@An2017; @Benda&Dunne1997], as well as the prediction of changes in river channel bed surface grain size [@Ferguson et al. 2015].
 
-Over the past two decades, computational morphodynamic models have ...
-[**JON, could you write two sentences here about the history of morphodynamic models?** ]
-Cui et al. [2006] ; Ferguson et al. [2015]
-Enter network models:CASCADE...
-Czuba (2018) introduced a network-based, Lagrangian bed material morphodynamic model that tracks the motion of individual units (referred to as “parcels”) of sediment through a river network. This approach improves on the existing morphodynamic models by: (1) accounting for the full river network, rather than a single longitudinal profile, (2) allowing the user to ‘tag’ particular sediment inputs and track their fate through time. These existing network sediment transport models, however, have two notable drawbacks: 1) they are written in a proprietary scripting language, and 2) they are not explicitly designed to be interoperable with other Earth-surface models, such as streamflow or landslide models.
+Most computational morphodynamic models take a Eulerian approach, which tracks changes in bed elevation through time as a function of the spatial gradient in sediment flux [e.g., @ParkerEbook,http://hydrolab.illinois.edu/people/parkerg/morphodynamics_e-book.htm]. These models directly compute bed elevation change and sediment flux throughout the domain. One of the major drawbacks with Eulerian morphodynamic models is the difficulty in being able to 'tag' individual sediment particles to answer questions about how an individual sediment particle/input may move, when it might arrive, and what affect it will have on river morphology when it arrives downstream. To overcome this drawback and to more easily extend morphodynamic models to entire river networks, recent work has focused on developing river-network based Lagrangian sediment transport models, which track the locations of individual sediment units on a river network.  
 
[…]
