### 2020-01-05 a4d6f42
improved docstrings
