### 2020-01-06 8344cfd
minor doc revisions
