+++ docs/source/index.rst
@@ -108,7 +108,7 @@ table.
 +----------------------------------+-----------------------------------------------------------------------------------------------------+
 | Component                        | Citation                                                                                            |
 +==================================+=====================================================================================================+
-| Cellular Automaton                | `Tucker et al. (2015) <https://www.geosci-model-dev.net/9/823/2016/>`_                              |
+| Cellular Automaton               | `Tucker et al. (2015) <https://www.geosci-model-dev.net/9/823/2016/>`_                              |
 +----------------------------------+-----------------------------------------------------------------------------------------------------+
 | Flexure                          | `Hutton and Syvitski (2008) <https://www.sciencedirect.com/science/article/pii/S0098300408000587>`_ |
 +----------------------------------+-----------------------------------------------------------------------------------------------------+
