### 2020-01-09 #1935 @nicgaspar
Hi @kthyng! I'm behind as always. I just realized that I have a COI. I am still willing to go ahead with the review if you think it is OK. But I actively collaborate with Katy Barnhart and Greg Tucker, both of whom are co-authors. As in we are working on papers together now, and have published together in the last year.

I feel I can be unbiased, but I should let you know. Sorry this took so long.
