### 2020-01-10 3cb24f0
Small edit to handling of negative slopes
