### 2020-01-11 ee28f24
handle vert+rect hex graph nodes_at_right_edge
