### 2020-01-12 8c988f0
lint hex.py
