### 2020-01-12 6e4e04f
tutorial output cleared
