### 2020-01-12 6aaba9b
component to model species evolution (#1065)
* First commit

* species tracked by identifiers

* Added plot_trees function for phylogenetics

* renamed seamless folder and file, incorporated species class

* componentized seamless

* tree troubleshotting print statements added

* species range is now tracked in species class

* species handles updating range

* added plot_species_range

* get_species_by_identifier fixed to set input as uuid

* update imported modules

* added stream_species as a subclass of species

* modified area captured calculation

* implemented habitat patches

* begun p
[…]
