+++ landlab/components/__init__.py
@@ -41,6 +41,7 @@ from .sink_fill import SinkFiller, SinkFillerBarnes
 from .soil_moisture import SoilInfiltrationGreenAmpt, SoilMoisture
 from .space import Space
 from .spatial_precip import SpatialPrecipitationDistribution
+from .species_evolution import SpeciesEvolver
 from .steepness_index import SteepnessFinder
 from .stream_power import (
     FastscapeEroder,
@@ -104,6 +105,7 @@ COMPONENTS = [
     SoilInfiltrationGreenAmpt,
     Space,
     SpatialPrecipitationDistribution,
+    SpeciesEvolver,
     SteepnessFinder,
     StreamPowerEroder,
     StreamPowerSmoothThresholdEroder,
