### 2020-01-13 #1935 @kthyng
Hi @rreinecke! You were interested in this JOSS submission before — any chance you still are and are available to review it? It turns out we had a conflict of interest with another reviewer.
