### 2020-01-13 dbc7da0
text edits
