+++ docs/source/development/index.rst
@@ -22,4 +22,3 @@ The most important things to remember are to:
    install/index
    contribution/index
    practices/index
-   joss_workflow
