+++ docs/source/development/contribution/joss_workflow.rst
@@ -8,7 +8,7 @@ The `Journal of Open Source Software (JOSS) <https://joss.theoj.org>`_ is a
 venue for publications on open source software. If you are considering
 preparing a contribution to Landlab for a JOSS publication, please make sure to
 review the JOSS `submission requirements
-<https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements`_
+<https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements>`_
 
 
 Parts of Landlab have already been published by JOSS, and multiple parts of
@@ -21,7 +21,7 @@ journal.
 
 
 - Create a pull request to merge the contribution into Landlab. See
-  :ref:`Develop your own component or utility <developer_a_component>`.
+  :ref:`Develop your own component or utility <landlab_component_dev_page>`.
 - Following the above merge, create a new branch for the JOSS review process.
   Name the paper ``paper.md``, which follows JOSS requirements, and place this
   file in a new folder, ``/landlab/joss/<name_of_contribution>/`` with your other
@@ -35,17 +35,17 @@ journal.
       $ @whedon prepare pdf from branch my-custom-branch-name
 
 - Once the JOSS review is complete and if the paper is accepted:
-  - Create a pull request to merge the JOSS review branch into the ``master``
-    branch.
-  - Merge ``master`` into ``release``, and update version number. See
-    :ref:`Create a Landlab release <dev_guide_releases>`.
-  - Ensure that the release is correctly created and distributed.
-  - Update the JOSS review pull request with the correct archive DOI and
-    version number.
-  - JOSS publishes the paper.
-  - Once publication is complete, make another pull request to move and rename
-    the ``paper.md`` to ``landlab/joss/published/<name_of_module_or_component>/<new_paper_name>.md``.
-    The paper can no longer be named ``paper.md`` because JOSS expects paper in
-    review to have this name, and to avoid conflicts with other Landlab paper
-    contributions in review. Note that the paper folder should be moved to
-    within the ``published`` directory.
+- Create a pull request to merge the JOSS review branch into the ``master``
+  branch.
+- Merge ``master`` into ``release``, and update version number. See
+  :ref:`Create a Landlab release <dev_guide_releases>`.
+- Ensure that the release is correctly created and distributed.
+- Update the JOSS review pull request with the correct archive DOI and
+  version number.
+- JOSS publishes the paper.
+- Once publication is complete, make another pull request to move and rename
+  the ``paper.md`` to ``landlab/joss/published/<name_of_module_or_component>/<new_paper_name>.md``.
+  The paper can no longer be named ``paper.md`` because JOSS expects paper in
+  review to have this name, and to avoid conflicts with other Landlab paper
+  contributions in review. Note that the paper folder should be moved to
+  within the ``published`` directory.
