+++ docs/source/development/install/install.rst
@@ -69,11 +69,22 @@ imports Landlab from the working copy you just cloned. This ensures that
 any changes you make to your copy of the code is seen by Python the
 *next* time you import Landlab.
 
+The conda environment described by ``environment-dev.yml`` contains the minimal
+set of dependencies necessary to run the Landlab tests and notebooks, and keep
+the codebase clean and tidy. It may not include some of your favorite
+development tools (e.g., spyder). See below for how to install additional
+packages into the conda environment.
+
+In addition, this environment does not have everything needed to build the
+documentation. These requirements are specified in the file
+``landlab/docs/environment.yml``.
+
 Conda Environment Tips
 ----------------------
 
 *   In order to use this environment, you will need to activate it every time
-    you open a new terminal instance.
+    you open a new terminal instance using the command
+    ``conda activate landlab_dev``.
 *   If you use python tools for your work that are not in the conda environment
     but you have previously installed them on your computer (e.g., spyder),
     you will need to add them to the environment. Use the standard terminal
