+++ landlab/components/species_evolution/zone_controller.py
@@ -17,8 +17,8 @@ class ZoneController(object):
     spatially continuous grid nodes.
 
     This controller creates zones using the initialization parameter,
-    zone_function``. This function identifies all of the grid nodes where zones
-    are to be created. A zone is created for each cluster of spatially
+    ``zone_function``. This function identifies all of the grid nodes where
+    zones are to be created. A zone is created for each cluster of spatially
     continuous nodes. Zones are updated also using this function when the
     ``run_one_step`` method of this controller is called.
 
@@ -26,8 +26,9 @@ class ZoneController(object):
     how zones are created. The grid contains six columns and six rows. In this
     example, the function returns True where node values are greater than 0.
     Nodes marked with an ``*`` are nodes that will belong to a zone because the
-    mask is True at these nodes. All other nodes are marked with a ``·``. A zone
-    is created for each cluster of continuous nodes where the mask is True.
+    mask is True at these nodes. All other nodes are marked with a ``·``. A
+    zone is created for each cluster of continuous nodes where the mask is True.
+    ::
 
     values         mask returned
     evaluated      by zone function
@@ -44,6 +45,7 @@ class ZoneController(object):
     where diagonal neighbors are included or ``D4`` where diagonal neighbors
     are excluded. A minimum zone area can be enforced with the ``minimum_area``
     initialization parameter.
+    ::
 
     D8             D4             D8             D4
     min area = 0   min area = 0   min area = 2   min area = 2
@@ -67,6 +69,7 @@ class ZoneController(object):
     symbols defined above. Individual zone masks and the count of zones are
     affected by the use of ``D8`` or ``D4`` along with the minimum area
     parameter, especially when zone clusters are along the grid parameter.
+    ::
 
     zone function  D8             D4             D8             D4
     returned mask  min area = 0   min area = 0   min_area = 2   min_area = 2
@@ -88,6 +91,7 @@ class ZoneController(object):
     zones. The grid represents the time, ``T0`` with the nodes of a zone
     marked with ``x``. The following examples will use D8 neighborhoods and a
     minimum zone area of 0.
+    ::
 
     T0
     · · · · · ·
@@ -103,6 +107,7 @@ class ZoneController(object):
     the zone when at least one zone node overlaps between the two time steps.
     However, in ``T1d``, no nodes overlaps, therefore taxa do not disperse from
     the zone in T0 to the zone in T1d.
+    ::
 
     T1a            T1b            T1c            T1d
     · · · · · ·    · · · · · ·    · · · · · ·    · · · · · ·
@@ -118,6 +123,7 @@ class ZoneController(object):
     zone fragmentations can be viewed in the ``record_data_frame`` attribute.
     In the T1e example, the fragmentation count for time 1 would be 2 because
     2 zones that fragmented from a prior zone were recognized at this time.
+    ::
 
     T1e
     · · · · · ·
@@ -141,6 +147,7 @@ class ZoneController(object):
     assumed to be the prior zone and the others are considered captured zones.
     The number of zone captures can be viewed in the ``record_data_frame``
     attribute.
+    ::
 
     T2
     · · · · · ·
