+++ landlab/components/species_evolution/README.md
@@ -1,0 +1,20 @@
+# SpeciesEvolver: evolve life in a simulated landscape
+
+## Documentation
+
+Landlab documentation is hosted on this [ReadTheDocs page](https://landlab.readthedocs.io/en/master), including Landlab installation instructions. SpeciesEvolver is included with Landlab.
+
+SpeciesEvolver documentation is located [here](https://landlab.readthedocs.io/en/master/reference/components/species_evolution.html).
+
+## SpeciesEvolver tutorial
+
+A tutorial of this component exists in the form of a Jupyter Notebook. It is
+located in ``/landlab/notebooks/tutorials`` titled, ``Introduction_to_SpeciesEvolver.ipynb``.
+
+A static version of the same tutorial can be viewed [here](https://nbviewer.jupyter.org/github/landlab/landlab/blob/master/notebooks/tutorials/species_evolution/Introduction_to_SpeciesEvolver.ipynb).
+
+## Issues
+
+[Open an Issue here](https://github.com/landlab/landlab/issues) where we can
+respond to your questions, comments, issues, or any identified bugs related to
+Landlab including SpeciesEvolver.
