### 2020-01-14 #1935 @rreinecke
@DavidLitwin before continuing with my review could you please provide me detailed information on where to find the new groundwater module? It appears that you have submitted the landlab main repository to this JOSS review - Is there a reason it is not in a separate repository and used as submodule?
Also the contribution of submitting author @DavidLitwin is not clear. From the number of total commits in the landlab project I cannot derive a major contribution.
