### 2020-01-14 #1935 @DavidLitwin
Hi @rreinecke and @dvalters, you can find the groundwater module in the components folder, [here](https://github.com/landlab/landlab/tree/master/landlab/components/groundwater). I think this is how this has been handled for all landlab components - they become part of the main repository so that they are readily available for anyone who uses landlab. Details on that are probably more a question for @kbarnhart.

As for contributions. The groundwater model was started by Greg Tucker. I did most of the core work from there. I adapted it for sloping aquifer base, incorporated the functions necessary to allow for seepage to the surface, wrote the adaptive timestep solver, wrote setters and getters and methods to calculate fluxes, wrote all current documentation and the tutorial. @kbarnhart has since made several commits so that the component is compatible with landlab 2.0. Does that help?
