+++ landlab/components/groundwater/readme.md
@@ -1,6 +1,6 @@
 Welcome to the README for the GroundwaterDupuitPercolator component submodule.
 
-The documentation is hosted on our [ReadTheDocs](https://landlab.readthedocs.io/en/master/landlab.components.groundwater_dupuit_percolator.html) page.
+The documentation is hosted on our [ReadTheDocs](https://landlab.readthedocs.io/en/master/reference/components/groundwater.html) page.
 
 There is a [Jupyter notebook in the Landlab Tutorials repository](https://mybinder.org/v2/gh/landlab/landlab/release?filepath=notebooks/tutorials/groundwater/groundwater_flow.ipynb) that describes the use of this submodule. The link takes you to a binder instance of this notebook.
 
