+++ landlab/components/species_evolution/zone_controller.py
@@ -30,14 +30,14 @@ class ZoneController(object):
     zone is created for each cluster of continuous nodes where the mask is True.
     ::
 
-    values         mask returned
-    evaluated      by zone function
-    0 0 0 0 0 0    · · · · · ·
-    0 0 0 5 4 2    · · · * * *
-    0 6 0 4 6 0    · * · * * ·
-    0 2 0 0 0 0    · * · · · ·
-    0 0 4 0 4 0    · · * · * ·
-    0 0 0 0 0 0    · · · · · ·
+        values         mask returned
+        evaluated      by zone function
+        0 0 0 0 0 0    · · · · · ·
+        0 0 0 5 4 2    · · · * * *
+        0 6 0 4 6 0    · * · * * ·
+        0 2 0 0 0 0    · * · · · ·
+        0 0 4 0 4 0    · · * · * ·
+        0 0 0 0 0 0    · · · · · ·
 
     The above example is continued in the following four diagrams that
     demonstrate how individual zones are identified. Each zone is marked with a
@@ -47,14 +47,14 @@ class ZoneController(object):
     initialization parameter.
     ::
 
-    D8             D4             D8             D4
-    min area = 0   min area = 0   min area = 2   min area = 2
-    · · · · · ·    · · · · · ·    · · · · · ·    · · · · · ·
-    · · · + + +    · · · + + +    · · · + + +    · · · + + +
-    · x · + + ·    · x · + + ·    · x · + + ·    · x · + + ·
-    · x · · · ·    · x · · · ·    · x · · · ·    · x · · · ·
-    · · x · o ·    · · @ · o ·    · · x · · ·    · · · · · ·
-    · · · · · ·    · · · · · ·    · · · · · ·    · · · · · ·
+        D8             D4             D8             D4
+        min area = 0   min area = 0   min area = 2   min area = 2
+        · · · · · ·    · · · · · ·    · · · · · ·    · · · · · ·
+        · · · + + +    · · · + + +    · · · + + +    · · · + + +
+        · x · + + ·    · x · + + ·    · x · + + ·    · x · + + ·
+        · x · · · ·    · x · · · ·    · x · · · ·    · x · · · ·
+        · · x · o ·    · · @ · o ·    · · x · · ·    · · · · · ·
+        · · · · · ·    · · · · · ·    · · · · · ·    · · · · · ·
 
     The grid perimeter affects zone creation. Zones can include perimeter nodes
     (the nodes along grid edges), although because perimeter nodes are not
@@ -71,14 +71,14 @@ class ZoneController(object):
     parameter, especially when zone clusters are along the grid parameter.
     ::
 
-    zone function  D8             D4             D8             D4
-    returned mask  min area = 0   min area = 0   min_area = 2   min_area = 2
-    * · · * * ·    + · · x x ·    + · · x x ·    + · · · · ·    · · · · · ·
-    · * · · * ·    · + · · x ·    · # · · x ·    · + · · · ·    · # · · · ·
-    · * · · · ·    · + · · · ·    · # · · · ·    · + · · · ·    · # · · · ·
-    * · · · * *    + · · · o o    $ · · · o o    + · · · o o    · · · · o o
-    · · · · * ·    · · · · o ·    · · · · o ·    · · · · o ·    · · · · o ·
-    · * * · · ·    · @ @ · · ·    · @ @ · · ·    · · · · · ·    · · · · · ·
+        zone function  D8             D4             D8             D4
+        returned mask  min area = 0   min area = 0   min_area = 2   min_area = 2
+        * · · * * ·    + · · x x ·    + · · x x ·    + · · · · ·    · · · · · ·
[…]
