### 2020-01-14 #1935 @kbarnhart
@rreinecke I'll add a few other comments regarding this submission and contributions. Hopefully this clarifies things, but if not, please let me know and I'll try and address additional questions.

This contribution specifically focuses on the GroundwaterDupuitPercolator component within the Landlab package, rather than the Landlab package itself. Because JOSS is set up for the submission to provide a repository (e.g., https://github.com/landlab/landlab), this means that the repository of record is github/landlab/landlab, but the scope of the contribution is the the component (https://github.com/landlab/landlab/tree/master/landlab/components/groundwater), the associated documentation, and examples. At present there is no way in the JOSS workflow to formally indicate that this submission focuses on only the GroundwaterDupuitPercolator component.

The first time someone considered submitting a component from Landlab to JOSS (it was me, [link to issue discussing this](https://github.com/openjournals/joss/issues/443)) there was a discussion about whether it would make the most sense to put it in a separate repository and then have Landlab consider it as a submodule. We decided that this was not the best approach because it would make maintenance of the component and the package more difficult. Additionally, keeping the component within the Landlab package means that a separate submodule doesn't need to re-implement all of the documentation-building, code testing, and CI specification package documents, but can take advantage of the setup that the landlab package already has.

While @DavidLitwin has made relatively few commits to the Landlab package, he is almost entirely responsible for this component, its documentation, and examples. My description of coauthor contributions would be as follows:

- @gregtucker started the component's development
- @gregtucker and Ciaran provided most of the theoretical guidance to @DavidLitwin
- @gregtucker and I provided operational and numerical guidance to @DavidLitwin
- @DavidLitwin did almost all of the coding, and created the documentation and examples.
