### 2020-01-14 #1935 @DavidLitwin
Thank you for clarifying @kbarnhart. I agree with the coauthor contributions as you have described them.
