+++ docs/source/development/practices/dev_guide_releases.rst
@@ -20,7 +20,7 @@ The Release Checklist
    This can either be done locally or on the GitHub website. Here are
    instructions to do this locally:
 
-   In the `landlab/landlab` repository, checkout the release branch and tag it.
+   In the `landlab/landlab` repository, checkout the `release` branch and tag it.
 
    .. code-block:: bash
 
@@ -39,6 +39,13 @@ The Release Checklist
    * Assuming its tests pass, merge the PR autogenerated in the landlab-feedstock.
 
    * Verify the builds on PyPI and on conda-forge.
+
+5. Now tag the current version of `master` with a "-dev" and push the tag.
+
+   .. code-block:: bash
+
+       $ git tag vX.X.X-dev
+       $ git push --tags
 
 Notes
 -----
