### 2020-01-15 #1935 @rreinecke
Thank you @kbarnhart and @DavidLitwin for clarifying the authorship.

Concerning the submodule I strongly disagree but that is something you need to decide and will not be part of this review. Nevertheless, I already have a problem in verifying the functionality of the software as I cannot find according unit tests (and how to run them). If they are somewhere else in the project please add a detailed description to your repository e.g. a list of submodules etc.. Please also explain there where to find which software component. I'm aware that there is another official Landlab page but as my time, and the time of the other reviewer and of course potential future users, is limited please consider to extend your readme. Also install instructions are missing together with a list of dependencies. Furthermore, did you implement a simple example model which demonstrates the capabilities of the new package? If so where? And where do I find the according documentation?

Please revise your repository in a way that I'm able to answer all of these questions on my own and let me know when it is implemented. Please do not describe these things in this thread. They should be self explanatory for every new user.
