### 2020-01-15 landlab/landlab#1105 @rreinecke
JOSS Review groundwater component: JOSS paper
Please consider adding a short paragraph on:
State of the field: Do the authors describe how this software compares to other commonly-used packages?
