### 2020-01-15 #1935 @dvalters
> Furthermore, did you implement a simple example model which demonstrates the capabilities of the new package?
There is a binder demo of the new component here (linked to in the component readme) which I'm using to explore the model.

https://mybinder.org/v2/gh/landlab/landlab/release?filepath=notebooks/tutorials/groundwater/groundwater_flow.ipynb
