### 2020-01-15 landlab/landlab#1106 @rreinecke
JOSS review groundwater component: Tutorial
Can you please add a step that verifies whether the mass balance is conserved also with varying step sizes and recharge rates?

What about different boundary conditions? Could you at least summarize what else is possible?
