### 2020-01-15 landlab/landlab#1107 @rreinecke
JOSS review groundwater component: Equations and documentation
You say that you are using the Dupuit–Forchheimer assumption to calculate groundwater flow right? That requires a lot of assumptions and is often used wrongly. Could you please show me where in the documentation this issue is discussed? If not please make it visible to potential users what the limitations of your approach are and what assumption need to hold! I think this is pivotal.
