### 2020-01-15 #1935 @rreinecke
@dvalters I've added a question concerning the used equations here: landlab/landlab#1107

Just to have it on record and be transparent (this is not a review request only a general suggestion and note, I don't think it is a conflict of interest): @dvalters I have developed a Modflow clone that allows to efficiently calculate 3d Darcy flow. Maybe that is of interest as an alternative to your use of Depuit. Just for consideration after the review for future work.

Nonetheless I think you should openly discuss the limitations of your approach so that users are applying it correctly.
