### 2020-01-15 #1935 @kbarnhart
@rreinecke I will open an issue in the landlab repository associated with your [prior comment](https://github.com/openjournals/joss-reviews/issues/1935#issuecomment-574524938) and address it with a PR. Once that PR is live I will ask you to comment on whether I have addressed the issues you raised.
