### 2020-01-15 7fde5fb
added readme to species evolver (#1103)
* added readme to SpeciesEvolver
