### 2020-01-15 33dcdbf
work on improving the readme
