+++ landlab/components/groundwater/readme.md
@@ -1,7 +1,39 @@
-Welcome to the README for the GroundwaterDupuitPercolator component submodule.
+## Welcome to the README for the GroundwaterDupuitPercolator component submodule.
 
-The documentation is hosted on our [ReadTheDocs](https://landlab.readthedocs.io/en/master/reference/components/groundwater.html) page.
+### Installation
+To use this component, you will need to install Landlab. Two options for
+installation are available:
+[a pre-packaged binary](https://landlab.readthedocs.io/en/master/install/index.html)
+distributed through PyPI or conda-forge and a
+[source code installation](https://landlab.readthedocs.io/en/master/development/install/index.html#developer-install).
 
-There is a [Jupyter notebook in the Landlab Tutorials repository](https://mybinder.org/v2/gh/landlab/landlab/release?filepath=notebooks/tutorials/groundwater/groundwater_flow.ipynb) that describes the use of this submodule. The link takes you to a binder instance of this notebook.
+The dependencies of the Landlab package are described [here]().
 
-If you have any questions, comments, issues, or bugs related to this submodule, please [open an Issue](https://github.com/landlab/landlab/issues/new) so we can respond.
+### Documentation
+The documentation specific to this component is housed within the Landlab
+documentation. There are two pages in the documentation that are most relevant
+to this component:
+- [The component API](https://landlab.readthedocs.io/en/master/reference/components/groundwater.html).
+- [A page](https://landlab.readthedocs.io/en/master/reference/components/dupuit_theory.html#dupuit-theory)
+describing the theory and numerical implementation of this component.
+
+If you are new to Landlab and components, we recommend that you also look at the
+[User Guide](https://landlab.readthedocs.io/en/master/user_guide/index.html),
+in particular, the page on the [model grid](https://landlab.readthedocs.io/en/master/user_guide/grid.html), and [components](https://landlab.readthedocs.io/en/master/user_guide/components.html).
+
+### Tutorials
+There is a [Jupyter notebook in the Landlab Tutorials repository](https://mybinder.org/v2/gh/landlab/landlab/release?filepath=notebooks/tutorials/groundwater/groundwater_flow.ipynb)
+that describes the use of the `GroundwaterDupuitPercolator`.
+The link takes you to a binder instance of this notebook. Its filepath within
+the repository is `notebooks/tutorials/groundwater/groundwater_flow.ipynb`
+
+### Tests of this Component
+Along with the rest of the Landlab package, this component uses
+[`pytest`](https://docs.pytest.org/en/latest/) 
+to  discover and run its tests. General information about running the Landlab
+tests can be found [here](https://landlab.readthedocs.io/en/master/development/install/test.html#testing).
+
+### Getting Help
+If you have any questions, comments, issues, or bugs related to this submodule,
+please [open an Issue](https://github.com/landlab/landlab/issues/new) so we can
+respond.
