+++ joss/species_evolver/paper.md
@@ -1,0 +1,52 @@
+---
+title: 'SpeciesEvolver: A Landlab component to evolve life in simulated landscapes'
+tags:
+  - Python
+  - Jupyter Notebook
+  - Landlab
+authors:
+  - name: Nathan J. Lyons
+    orcid: 0000-0001-6965-3374
+    affiliation: 1
+  - name: James S. Albert
+    orcid: 0000-0001-5477-1749
+    affiliation: 2
+  - name: Nicole M. Gasparini
+    orcid: 0000-0002-0803-3697
+    affiliation: 1
+affiliations:
+ - name: Tulane University, Department of Earth and Environmental Sciences
+   index: 1
+ - name: University of Louisiana Lafayette, Department of Biology
+   index: 2
+date: 15 January 2020
+bibliography: paper.bib
+---
+
+# Summary
+
+The surface of the Earth and its biota evolve together. Climate and tectonics ultimately drive the physical and chemical surface processes that evolve landscape structure, including the connectivity of landscape portions that facilitate or impede movement of organismal populations [@Stanley:1979; Antonelli:2018]. Impeded organismal movement reduces gene flow among populations and genetic diversity within populations, increasing the probability of species extinction [@Bohonak:1999]. Long-term geographic separation of populations (i.e. allopatry) is a mechanism of speciation as populations genetically diverge due to reproductive isolation [@Coyne:1992]. Speciation within the same area (i.e. sympatry) can emerge as subpopulations specialize in different resources, and along environmental gradients (e.g. in surface air temperature) where a continuum of reproductive isolation develops, among other proposed mechanisms [@Dieckmann:1999; @Doebeli:2003]. These macroevolutionary processes—dispersal, extinction, and speciation—determine regional biodiversity [@Taylor:1993].
+
+Biodiversity is influenced by changes in landscape connectivity, nevertheless linking the evolution of a landscape with its biota is complicated by limited landform and genetic preservation, disparate timescales of surface and macroevolutionary processes, and landscape heterogeneity (e.g., species-dense assemblages, lithological variability), among numerous other factors. Discovery of surface dynamics has been aided by numerical landscape evolution models [@Tucker:2010]. These models act as digital laboratories where researchers can apply current theory to simulate complex surface dynamics while quantitatively exploring new ideas. Yet to be developed and widely shared is a modeling tool that integrates macroevolutionary processes with numerical representations of surface processes. Such a tool can help research communities overcome the complications of linking the evolution of life and landscapes.
+
+We built ``SpeciesEvolver`` to simulate taxa that evolve at geologic and macroevolutionary timescales in two-dimensional landscapes. This software is adapted from SEAMLESS (Spatially Explicit Area Model of Landscape Evolution by SimulationS) that models organismal diversification in one-dimensional space [@Albert:2016]. ``SpeciesEvolver`` operates in two dimensions, built for extension, and is a component of the Landlab modeling toolkit.
+
+Landlab is an open source Python package that provides tools to build numerical models of surface dynamics [@Hobley:2017]. A landscape is represented by a model grid with configurable spatial dimensions. The surface processes components of Landlab drive change in landscape attributes (e.g. topographic elevation). The use of multiple components in a model effectively couples the processes because the processes work with the same grid and its landscape attributes. Building ``SpeciesEvolver`` into Landlab allows its users to examine taxa evolution alongside surface processes-driven landscape changes.
+
[…]
