### 2020-01-15 d6c71fc
various doc improvements
