+++ docs/source/development/install/install.rst
@@ -54,13 +54,17 @@ will contain the file ``setup.py``) and run the commands below.
 
 Landlab has a number of dependencies to run, test, and develop with. These are
 described in more detail :ref:`here <dependencies>`. We have created a conda
-environment file which contains everything you will need for development. The
-next step is it create this environment:
+environment file which contains everything you will need for development. Note
+that while this environment has everything you'll need to run Landlab and its
+notebooks, it may not have all of your favorite development tools (e.g.,
+spyder). See below for instructions on adding additional things to the
+environment.
 
-Install all of the dependencies required by Landlab into a new environment
-called *landlab_dev*. Read more about
+Read more about
 `conda environments <https://docs.conda.io/projects/conda/en/latest/user-guide/getting-started.html#managing-environments>`_.
 on the conda webpage.
+
+The next step is it create this environment, called *landlab_dev*.
 
 .. code-block:: bash
 
