### 2020-01-15 14aba0b
add Python 3.8 builds to AppVeyor and Travis
