### 2020-01-15 0c36cf2
additional readme updates
