### 2020-01-15 c229b28
deploy for Python 3.8 on Mac
