+++ .travis.yml
@@ -32,11 +32,14 @@ jobs:
     stage: deploy
     if: tag =~ v.*$
     os: osx
-    env: CONDA_ENV=3.7
+    env: CONDA_ENV=3.8
     script:
     - pip install twine wheel
     - python setup.py bdist_wheel
     - twine upload -u landlab -p$PYPI_PASS dist/*landlab*
+  - <<: *deploy
+    os: osx
+    env: CONDA_ENV=3.7
   - <<: *deploy
     os: osx
     env: CONDA_ENV=3.6
