### 2020-01-15 landlab/landlab#1108 @kbarnhart
I've created two PRs to address this issue.

First, #1110 updates core landlab documentation regarding installation, tests, and dependencies.

Second, #1112 updates the submodule Readme.

Once these two PRs are merged, I'll ping @rreinecke to asses whether we have addressed his concerns, or if there are still outstanding issues that need to be addressed on this point.
