### 2020-01-15 ebdf638
removed figure section
