+++ joss/species_evolver/paper.md
@@ -1,49 +1,51 @@
 ---
-title: 'SpeciesEvolver: A Landlab component to evolve life in simulated landscapes'
+title: 'GroundwaterDupuitPercolator: A Landlab component for groundwater flow'
 tags:
   - Python
-  - Jupyter Notebook
   - Landlab
 authors:
-  - name: Nathan J. Lyons
-    orcid: 0000-0001-6965-3374
+  - name: David G. Litwin
+    orcid: 0000-0002-8097-4029
     affiliation: 1
-  - name: James S. Albert
-    orcid: 0000-0001-5477-1749
-    affiliation: 2
-  - name: Nicole M. Gasparini
-    orcid: 0000-0002-0803-3697
-    affiliation: 1
+  - name: Gregory E. Tucker
+    orcid: 0000-0003-0364-5800
+    affiliation: 2, 3, 4
+  - name: Katherine R. Barnhart
+    orcid: 0000-0001-5682-455X
+    affiliation: 2, 3
+  - name: Ciaran J. Harman
+    orcid: 0000-0002-3185-002X
+    affiliation: 1, 5
 affiliations:
- - name: Tulane University, Department of Earth and Environmental Sciences
-   index: 1
- - name: University of Louisiana Lafayette, Department of Biology
-   index: 2
-date: 15 January 2020
-bibliography: paper.bib
+  - name: Johns Hopkins University, Department of Environmental Health and Engineering
+  - index: 1
+  - name: University of Colorado at Boulder, Department of Geological Sciences
+    index: 2
+  - name: University of Colorado at Boulder, Cooperative Institute for Research in Environmental Sciences
+    index: 3
+  - name: University of Colorado at Boulder, Community Surface Dynamics Modeling System Integration Facility
+    index: 4
+  - name: Johns Hopkins University, Department of Earth and Planetary Science
+  - index: 5  
+date: 18 November 2019
+bibliography: papers.bib
 ---
 
 # Summary
+A large portion of the water that enters a catchment as precipitation percolates through soil and rock before exiting to water bodies or returning to the atmosphere as evapotranspiration. In many places, the discharge of water stored in the subsurface is a primary source of streamflow, and thus controls the ways in which catchments respond to stochastic variations in precipitation and climate [@beck_global_2013; @jasechko_pronounced_2014]. Previous studies have shown the importance of groundwater for a diverse range of processes, from transpiration to solute export [@maxwell_connections_2016; @van_verseveld_role_2009], and across diverse timescales, from rainfall-runoff response to landscape evolution [@huang_modelling_2006; @sklash_role_1979]. Of particular relevance to landscape evolution, groundwater can be an important control on the occurrence of overland flow, as the interaction of the water table with the ground surface controls the spatial extent of saturation and groundwater return flow [@dunne_partial_1970].
+Variably-saturated groundwater flow is often assumed to be governed by the Richards equation, which describes how water content and/or total energy potential evolve in an idealized porous medium due to fluxes of water driven by gradients in total potential. Numerical solutions to the Richards equation are computationally expensive [e.g. @kirkland_algorithms_1992], often limiting their applications. For computational efficiency, we use the widely applied Dupuit-Forcheimer approximation, which simplifies the Richards equation when aquifers are laterally extensive in comparison to their thickness, and the capillary fringe above the water table is relatively thin [e.g. @childs_drainage_1971; @troch_hillslope-storage_2003]. In this case, the water table is modeled as a free surface with groundwater flow driven by gradients in water table elevation. When the model assumptions are valid, this greatly reduces the model complexity while still producing water table elevations and discharges comparable to Richards equation solutions [@hilberts_hillslope-storage_2004].
 
[…]
