+++ joss/species_evolver/paper.bib
@@ -1,0 +1,124 @@
+@article{Albert:2016,
+    author = {Albert, James S and Schoolmaster Jr, Donald R and Tagliacollo, Victor and Duke-Sylvester, Scott M},
+    doi = {10.1093/sysbio/syw080},
+    journal = {Systematic Biology},
+    number = {2},
+    pages = {167--182},
+    title = {Barrier Displacement on a Neutral Landscape: Toward a Theory of Continental Biogeography},
+    url = {https://dx.doi.org/10.1093/sysbio/syw080},
+    volume = {66},
+    year = {2016}
+}
+
+@article{Antonelli:2018,
+    author = {Antonelli, Alexandre and Kissling, W Daniel and Flantua, Suzette GA and Berm{\'u}dez, Mauricio A and Mulch, Andreas and Muellner-Riehl, Alexandra N and Kreft, Holger and Linder, H Peter and Badgley, Catherine and Fjeldsa, Jon and Fritz, Susanne A and Rahbek, Carsten and Herman, Fr{\'e}d{\'e}ric and Hooghiemstra, Henry and Hoorn, Carina},
+    doi = {10.1038/s41561-018-0236-z},
+    journal = {Nature Geoscience},
+    number = {10},
+    pages = {718--725},
+    title = {Geological and climatic influences on mountain biodiversity},
+    url = {https://doi.org/10.1038/s41561-018-0236-z},
+    volume = {11},
+    year = {2018}
+}
+
+@article{Bohonak:1999,
+    author = {Bohonak, Andrew J},
+    doi = {10.1086/392950},
+    journal = {The Quarterly Review of Biology},
+    number = {1},
+    pages = {21--45},
+    title = {Dispersal, gene flow, and population structure},
+    url = {https://doi.org/10.1086/392950},
+    volume = {74},
+    year = {1999}
+}
+
+@article{Coyne:1992,
+    author = {Coyne, Jerry A},
+    doi = {10.1038/355511a0},
+    journal = {Nature},
+    number = {6360},
+    pages = {511-515},
+    title = {Genetics and speciation},
+    url = {https://dx.doi.org/10.1038/355511a0},
+    volume = {355},
+    year = {1992}
+}
+
+@article{Dieckmann:1999,
+    author = {Dieckmann, Ulf and Doebeli, Michael},
+    doi = {10.1038/22521},
+    journal = {Nature},
+    number = {6742},
+    pages = {354--357},
+    title = {On the origin of species by sympatric speciation},
+    url = {https://dx.doi.org/10.1038/22521},
+    volume = {400},
+    year = {1999}
+}
+
+@article{Doebeli:2003,
+    author = {Doebeli, Michael and Dieckmann, Ulf},
+    doi = {10.1038/nature01274},
+    journal = {Nature},
+    number = {6920},
+    pages = {259--264},
+    title = {Speciation along environmental gradients},
+    url = {https://dx.doi.org/10.1038/nature01274},
+    volume = {421},
+    year = {2003}
+}
+
+@article{Hobley:2017,
+    author = {Hobley, Daniel EJ and Adams, Jordan M and Nudurupati, Sai Siddhartha and Hutton, Eric WH and Gasparini, Nicole M and Istanbulluoglu, Erkan and Tucker, Gregory E},
+    doi = {10.5194/esurf-5-21-2017},
+    journal = {Earth Surface Dynamics},
+    number = {1},
+    pages = {21--46},
+    title = {Creative computing with Landlab: an open-source toolkit for building, coupling, and exploring two-dimensional numerical models of Earth-surface dynamics},
+    url = {https://doi.org/10.5194/esurf-5-21-2017},
+    volume = {5},
+    year = {2017}
+}
+
[…]
