+++ docs/source/development/install/install.rst
@@ -54,11 +54,7 @@ will contain the file ``setup.py``) and run the commands below.
 
 Landlab has a number of dependencies to run, test, and develop with. These are
 described in more detail :ref:`here <dependencies>`. We have created a conda
-environment file which contains everything you will need for development. Note
-that while this environment has everything you'll need to run Landlab and its
-notebooks, it may not have all of your favorite development tools (e.g.,
-spyder). See below for instructions on adding additional things to the
-environment.
+environment file which contains everything you will need for development.
 
 Read more about
 `conda environments <https://docs.conda.io/projects/conda/en/latest/user-guide/getting-started.html#managing-environments>`_.
@@ -69,7 +65,6 @@ The next step is it create this environment, called *landlab_dev*.
 .. code-block:: bash
 
    $ conda env create --file=environment-dev.yml
-
 
 The conda environment described by ``environment-dev.yml`` contains the minimal
 set of dependencies necessary to run the Landlab tests and notebooks, and keep
