### 2020-01-15 58f2e00
remove figure
