### 2020-01-15 e0bbb1a
remove isbn
