+++ joss/species_evolver/paper.bib
@@ -1,3 +1,4 @@
+
 @article{Albert:2016,
     author = {Albert, James S and Schoolmaster Jr, Donald R and Tagliacollo, Victor and Duke-Sylvester, Scott M},
     doi = {10.1093/sysbio/syw080},
@@ -93,7 +94,6 @@
 
 @book{Stanley:1979,
     author = {Stanley, Steven M},
-    isbn={0716710927},
     publisher={WH Freeman}
     title = {Macroevolution: pattern and process},
     year = {1993}
