### 2020-01-15 c0196d4
reduced words
