### 2020-01-15 landlab/landlab#1106 @DavidLitwin
Yes, can do. I think that all of these are good ideas. It won't happen today but I will work on it when I'm able.
