### 2020-01-15 18e0cf7
changed figure format to jpg
