### 2020-01-15 0be6300
changed figure format to png
