### 2020-01-15 b14af45
set path to Python 3.8 installation on AppVeyor
