+++ appveyor.yml
@@ -11,7 +11,7 @@ environment:
 
     - TARGET_ARCH: x64
       CONDA_NPY: 111
-      CONDA_INSTALL_LOCN: C:\\Miniconda38-x64
+      CONDA_INSTALL_LOCN: C:\\Python38
       CONDA_PY: 3.8
 
     - TARGET_ARCH: x64
