### 2020-01-16 de767a1
Bibtex fix (#1114)
