+++ joss/species_evolver/paper.bib
@@ -94,7 +94,7 @@
 @book{Stanley:1979,
     author = {Stanley, Steven M},
     isbn={0716710927},
-    publisher={WH Freeman}
+    publisher={WH Freeman},
     title = {Macroevolution: pattern and process},
     year = {1993}
 }
