### 2020-01-16 cd10f11
changed figure to png
