### 2020-01-16 landlab/landlab#1108 @rreinecke
@kbarnhart Certainly. I'm referring to the main README in the landlab project. I know this review is for the groundwater component only, but as you choose to maintain it in this repository my comments will of course also address possible issues with the landlab project itself.

I'm missing a comprehensible README that:
* Tells me how to install the software
* What the dependencies are
* Were to find examples
* Maybe a short example
* How to run the tests
* Guides me through the package structure of the project
* Where to find the extended documentation
* And tells me who is responsible and how to contribute

I think everything is somewhere in your project and probably on your homepage but it is very cumbersome to find in the gitlab repository (which is under review not your homepage). Please revise the main README. They are many other github projects that do a great job in providing the information mentioned above. Maybe checkout other landing pages to improve yours.
