### 2020-01-16 6b20387
install with pip, not conda, on AppVeyor
