### 2020-01-16 7858cc4
use full path to python and pip
