### 2020-01-16 0492408
fix syntax error in appveyor config script
