### 2020-01-16 2c46b24
set PATH on AppVeyor
