### 2020-01-16 b5465ef
run commands with WITH_COMPILER
