+++ appveyor.yml
@@ -30,26 +30,26 @@ init:
 
 install:
   - "SET PATH=%PYTHON%;%PYTHON%\\Scripts;%PATH%"
-  - "pip install twine -r requirements-testing.txt -r requirements-notebooks.txt"
-  - "python -VV"
+  - "%WITH_COMPILER% pip install twine -r requirements-testing.txt -r requirements-notebooks.txt"
+  - "%WITH_COMPILER% python -VV"
 
 build: false
 
 test_script:
   - "%WITH_COMPILER% pip install numpy"
-  - "pip install netcdf4==1.5.2"
-  - "pip install -e ."
-  - "pytest -vvv"
+  - "%WITH_COMPILER% pip install netcdf4==1.5.2"
+  - "%WITH_COMPILER% pip install -e ."
+  - "%WITH_COMPILER% pytest -vvv"
 
 after_test:
-  - "python setup.py bdist_wheel"
+  - "%WITH_COMPILER% python setup.py bdist_wheel"
 
 artifacts:
   # Archive the generated conda package in the ci.appveyor.com build report.
   - path: 'dist\*'
 
 deploy_script:
-  - cmd: "python .ci/appveyor/pypi_upload.py"
+  - cmd: "%WITH_COMPILER% python .ci/appveyor/pypi_upload.py"
 
 notifications:
   - provider: Slack
