### 2020-01-16 a2471d3
remove os section for AppVeyor
