+++ appveyor.yml
@@ -21,7 +21,7 @@ environment:
 platform:
   - x64
 
-os: Previous Visual Studio 2015
+# os: Previous Visual Studio 2015
 # os: Visual Studio 2013
 
 init:
