### 2020-01-16 landlab/landlab#1108 @kbarnhart
@rreinecke thanks for clarifying. I'll re-ping you here when I've revised.
