### 2020-01-16 64e00c0
main readme and contributing updates.
