### 2020-01-16 53e8093
start on additional doc updates
