+++ docs/source/index.rst
@@ -6,17 +6,29 @@
 Landlab: A modular Earth Surface Dynamics modeling library
 ==========================================================
 
+Landlab is an open-source Python-language package for numerical modeling of
+Earth surface dynamics. it contains
 
-Landlab is a Python-based library that allows scientists and students to
-build numerical landscape models. Designed for disciplines that quantify
-earth surface dynamics such as geomorphology, hydrology, glaciology, and
-stratigraphy, it can also be used in related fields.
+* A gridding engine which represents the model domain. Regular and irregular
+  grids are supported.
+* A library of process components, each of which represents a physical process
+  (e.g., generation of rain, erosion by flowing water). These components have
+  a common interface and can be combined based on a users needs.
+* Utilities which support general numerical methods, file input/output, and
+  visualization.
 
-Landlab provides components to compute flows (such as water, sediment,
-glacial ice, volcanic material, or landslide debris) across a gridded
-terrain. With its robust, reusable components, Landlab allows scientists
-to quickly build landscape model experiments and compute mass balance
-across scales. Watch the webinar `Landlab Toolkit Overview <https://csdms.colorado.edu/wiki/Presenters-0407>`_ at CSDMS to learn more.
+In addition Landlab contains a set of Jupyter notebook tutorials providing
+introduction to core concepts and examples of use.
+
+Landlab was desiged for disciplines that quantify Earth surface dynamics such
+as geomorphology, hydrology, glaciology, and stratigraphy, it can also be used
+in related fields. Scientists who want to use a this type of model often build
+their own unique model from the ground up, re-coding the basic building blocks
+of their landscape model rather than taking advantage of codes that have
+already been written.
+
+Watch the webinar `Landlab Toolkit Overview <https://csdms.colorado.edu/wiki/Presenters-0407>`_
+ at CSDMS to learn more.
 
 The most current source code is always available from our `git
 repository <https://github.com/landlab/landlab>`_.
@@ -66,18 +78,19 @@ by the following grants:
       `1450412 <https://www.nsf.gov/awardsearch/showAward?AWD_ID=1450412&HistoricalAwards=false>`_).
     * A NSF EAR Postdoctoral Fellowship to Katy Barnhart
       (`1725774 <https://www.nsf.gov/awardsearch/showAward?AWD_ID=1725774&HistoricalAwards=false>`_).
-    * TODO: Add CSDMS grants, add any other grant.
 
 Citing Landlab
 ==============
 
 If you use any portion of Landlab, you must cite the following paper:
 
-`Hobley, D. E. J. <https://www.earth-surf-dynam.net/5/21/2017/>`_, Adams, J. M., Nudurupati, S. S., Hutton, E. W. H., Gasparini, N. M., Istanbulluoglu, E. and Tucker, G. E., 2017, Creative computing with Landlab: an open-source toolkit for building, coupling, and exploring two-dimensional numerical models of Earth-surface dynamics, Earth Surface Dynamics, 5, p 21-46, 10.5194/esurf-5-21-2017.
+`Hobley, D. E. J. <https://www.earth-surf-dynam.net/5/21/2017/>`__, Adams,
+J. M., Nudurupati, S. S., Hutton, E. W. H., Gasparini, N. M., Istanbulluoglu,
+E. and Tucker, G. E., 2017, Creative computing with Landlab: an open-source
+toolkit for building, coupling, and exploring two-dimensional numerical models
+of Earth-surface dynamics, Earth Surface Dynamics, 5, p 21-46,
+10.5194/esurf-5-21-2017.
 
-In addition, many components have an additional citation. The table below lists
-these. We also provide a handy way to get a bibtex file of the components you
-use. It is described below.
 
 BibTeX format:
 ::
@@ -100,49 +113,6 @@ BibTeX format:
[…]
