+++ docs/source/component_citation.rst
@@ -1,0 +1,38 @@
+.. _cite_components
+
+How to Cite Components
+======================
+
+
+In addition, many components have an additional citation. The table below lists
+these. We also provide a handy way to get a bibtex file of the components you
+use. It is described below.
+
+The following is a summary table of component and utility citations:
+
+TODO: we need to add to citations from Table 5 in Hobley et al. (2017) to this
+table.
+
++----------------------------------+-----------------------------------------------------------------------------------------------------+
+| Component                        | Citation                                                                                            |
++==================================+=====================================================================================================+
+| Cellular Automaton               | `Tucker et al. (2015) <https://www.geosci-model-dev.net/9/823/2016/>`_                              |
++----------------------------------+-----------------------------------------------------------------------------------------------------+
+| Flexure                          | `Hutton and Syvitski (2008) <https://www.sciencedirect.com/science/article/pii/S0098300408000587>`_ |
++----------------------------------+-----------------------------------------------------------------------------------------------------+
+| LakeMapperBarnes                 | `Barnes et al. (2014) <https://linkinghub.elsevier.com/retrieve/pii/S0098300413001337>`_            |
++----------------------------------+-----------------------------------------------------------------------------------------------------+
+| LandslideProbability             | `Strauch et al. (2017) <https://www.earth-surf-dynam.net/6/49/2018/esurf-6-49-2018.html>`_          |
++----------------------------------+-----------------------------------------------------------------------------------------------------+
+| LateralEroder                    | `Langston and Tucker (2018) <https://www.earth-surf-dynam.net/6/1/2018/>`_                          |
++----------------------------------+-----------------------------------------------------------------------------------------------------+
+| Lithology                        | `Barnhart et al. (2018) <https://joss.theoj.org/papers/10.21105/joss.00979>`_                       |
++----------------------------------+-----------------------------------------------------------------------------------------------------+
+| LithoLayers                      | `Barnhart et al. (2018) <https://joss.theoj.org/papers/10.21105/joss.00979>`_                       |
++----------------------------------+-----------------------------------------------------------------------------------------------------+
+| OverlandFlow                     | `Adams et al. (2017) <https://www.geosci-model-dev-discuss.net/gmd-2016-277/>`_                     |
++----------------------------------+-----------------------------------------------------------------------------------------------------+
+| SPACE                            | `Shobe et al. (2017) <https://www.geosci-model-dev.net/10/4577/2017/>`_                             |
++----------------------------------+-----------------------------------------------------------------------------------------------------+
+| SpatialPrecipitationDistribution | `Singer et al. (2018) <https://www.geosci-model-dev.net/11/3713/2018/gmd-11-3713-2018.html>`_       |
++----------------------------------+-----------------------------------------------------------------------------------------------------+
