+++ docs/source/development/package_organization.rst
@@ -1,0 +1,4 @@
+.. _organization
+
+Package Organization
+--------------------
