### 2020-01-16 9fb47a9
Address @mcflugen comments
