### 2020-01-17 4b8218f
improve cite as attribute
