### 2020-01-17 bb21e8d
improve citation documentation
