### 2020-01-17 c6d008b
misc docs
