+++ docs/source/development/index.rst
@@ -22,3 +22,4 @@ The most important things to remember are to:
    install/index
    contribution/index
    practices/index
+   package_organization
