+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -252,7 +252,7 @@ class NetworkSedimentTransporter(Component):
             )
             raise ValueError(msg)
 
-        self._num_parcels = self._parcels.dataset.element_id.size
+        self._num_parcels = self._parcels.number_of_items
 
         self.parcel_attributes = [
             "time_arrival_in_link",
@@ -326,13 +326,6 @@ class NetworkSedimentTransporter(Component):
             )
             raise ValueError(msg)
 
-        if "drainage_area" not in self._grid.at_link:
-            msg = (
-                "NetworkSedimentTransporter: drainage_area must be assigned"
-                "to the grid links"
-            )
-            raise ValueError(msg)
-
         # create field for channel slope if it doesnt exist yet.
         if "channel_slope" not in self._grid.at_link:
             self._channel_slope = self._grid.zeros(at="node")
@@ -343,13 +336,6 @@ class NetworkSedimentTransporter(Component):
         if "time_arrival_in_link" not in self._parcels.dataset:
             msg = (
                 "NetworkSedimentTransporter: time_arrival_in_link must be"
-                "assigned to the parcels"
-            )
-            raise ValueError(msg)
-
-        if "starting_link" not in self._parcels.dataset:
-            msg = (
-                "NetworkSedimentTransporter: starting_link must be"
                 "assigned to the parcels"
             )
             raise ValueError(msg)
@@ -425,6 +411,9 @@ class NetworkSedimentTransporter(Component):
             self._parcels.dataset.element_id[:, -1] == _OUT_OF_NETWORK
         )
         self._this_timesteps_parcels[self._parcels_off_grid, -1] = False
+        
+        self._num_parcels = self._parcels.number_of_items
+        # ^ needs to run just in case we've added more parcels
 
     def _update_channel_slopes(self):
         """Re-calculate channel slopes during each timestep."""
@@ -570,13 +559,17 @@ class NetworkSedimentTransporter(Component):
             self._parcels.dataset.active_layer == _ACTIVE
         ) * (self._this_timesteps_parcels)
 
-        # print("active_parcel_records",self._active_parcel_records)
-
-        vol_act = self._parcels.calc_aggregate_value(
-            np.sum, "volume", at="link", filter_array=self._active_parcel_records
-        )
-
-        vol_act[np.isnan(vol_act) == 1] = 0
+       #print("active_parcel_records",self._active_parcel_records)
+        
+        if np.any(self._active_parcel_records)== True:         
+            vol_act = self._parcels.calc_aggregate_value(
+                np.sum, "volume", at="link", 
+                filter_array=self._active_parcel_records
+            )
+            vol_act[np.isnan(vol_act) == 1] = 0
+            
+        else:
+            vol_act = np.zeros_like(vol_tot)
 
         self.vol_stor = (vol_tot - vol_act) / (1 - self.bed_porosity)
 
@@ -699,10 +692,15 @@ class NetworkSedimentTransporter(Component):
         )
         vol_tot[np.isnan(vol_tot) == 1] = 0
 
-        vol_act = self._parcels.calc_aggregate_value(
-            np.sum, "volume", at="link", filter_array=self._active_parcel_records
-        )
-        vol_act[np.isnan(vol_act) == 1] = 0
+        if np.any(self._active_parcel_records)== True:         
+            vol_act = self._parcels.calc_aggregate_value(
+                np.sum, "volume", at="link", 
+                filter_array=self._active_parcel_records
+            )
+            vol_act[np.isnan(vol_act) == 1] = 0
+            
+        else:
+            vol_act = np.zeros_like(vol_tot)
 
         # find active sand.
         findactivesand = (
[…]
