### 2020-01-17 71ff464
return fill_value when filter is all false
