### 2020-01-17 b112075
Documentation updates (#1110)
* various doc improvements

* a bit more about docs

* bug: missing `

* remove duplicate information.

* Address @mcflugen comments
