### 2020-01-17 3c1e027
add all component references
passes pretty and lint
