+++ landlab/components/erosion_deposition/erosion_deposition.py
@@ -54,9 +54,38 @@ class ErosionDeposition(_GeneralizedErosionDeposition):
     transport and erosion, and a hybrid response as described by Davy & Lague.
 
     Component written by C. Shobe, K. Barnhart, and G. Tucker.
+
+    References
+    ----------
+    **Required Software Citation(s) Specific to this Component**
+
+    Barnhart, K., Glade, R., Shobe, C., Tucker, G. (2019). Terrainbento 1.0: a
+    Python package for multi-model analysis in long-term drainage basin
+    evolution. Geoscientific Model Development  12(4), 1267--1297.
+    https://dx.doi.org/10.5194/gmd-12-1267-2019
+
+    **Additional References**
+
+    Davy, P., Lague, D. (2009). Fluvial erosion/transport equation of landscape
+    evolution models revisited Journal of Geophysical Research  114(F3),
+    F03007. https://dx.doi.org/10.1029/2008jf001146
+
     """
 
     _name = "ErosionDeposition"
+
+    _cite_as = """
+    @article{barnhart2019terrain,
+      author = {Barnhart, Katherine R and Glade, Rachel C and Shobe, Charles M and Tucker, Gregory E},
+      title = {{Terrainbento 1.0: a Python package for multi-model analysis in long-term drainage basin evolution}},
+      doi = {10.5194/gmd-12-1267-2019},
+      pages = {1267---1297},
+      number = {4},
+      volume = {12},
+      journal = {Geoscientific Model Development},
+      year = {2019},
+    }
+    """
 
     _info = {
         "flow__link_to_receiver_node": {
