+++ landlab/components/species_evolution/__init__.py
@@ -6,7 +6,13 @@
 """
 from .species_evolver import SpeciesEvolver
 from .zone import Zone
-from .zone_controller import ZoneController
 from .zone_taxon import ZoneTaxon
+from .zone_controller import ZoneController
 
-__all__ = ["SpeciesEvolver", "Zone", "ZoneTaxon", "ZoneController"]
+
+__all__ = [
+    'SpeciesEvolver',
+    'Zone',
+    'ZoneTaxon',
+    'ZoneController'
+]
