+++ landlab/components/species_evolution/species_evolver.py
@@ -10,7 +10,6 @@ import numpy as np
 from pandas import DataFrame, isnull
 
 from landlab import Component
-
 from .record import Record
 
 
@@ -197,20 +196,8 @@ class SpeciesEvolver(Component):
     ``ZoneTaxon`` as well as the parameters used in this example (default
     values were used because optional parameters were not set). Different
     behavior can be achieved by subclassing ``ZoneTaxon`` or ``Taxon``.
-
-    References
-    ----------
-    **Required Software Citation(s) Specific to this Component**
-
-    A publication has been submitted to the Journal of Open Source Software
-
-    **Additional References**
-
-    None Listed
-
     """
-
-    _name = "SpeciesEvolver"
+    _name = 'SpeciesEvolver'
 
     _info = {
         "taxa__richness": {
@@ -239,21 +226,21 @@ class SpeciesEvolver(Component):
         # Create data structures.
 
         self._record = Record(initial_time)
-        self._record.set_value("taxa", 0)
+        self._record.set_value('taxa', 0)
 
-        self._taxa = OrderedDict(
-            [
-                ("uid", []),
-                ("appeared", []),
-                ("latest_time", []),
-                ("extant", []),
-                ("object", []),
-            ]
+        self._taxa = OrderedDict([
+            ('uid', []),
+            ('appeared', []),
+            ('latest_time', []),
+            ('extant', []),
+            ('object', [])]
         )
 
         # Create a taxa richness field.
 
-        _ = grid.add_zeros("taxa__richness", at="node", dtype=int, clobber=True)
+        _ = grid.add_zeros(
+            'taxa__richness', at='node', dtype=int, clobber=True
+        )
 
     @property
     def record_data_frame(self):
@@ -288,10 +275,10 @@ class SpeciesEvolver(Component):
         The DataFrame is created from a data structure within the component.
         """
         cols = list(self._taxa.keys())
-        cols.remove("uid")
-        cols.remove("object")
-        df = DataFrame(self._taxa, columns=cols, index=self._taxa["uid"])
-        df.index.name = "uid"
+        cols.remove('uid')
+        cols.remove('object')
+        df = DataFrame(self._taxa, columns=cols, index=self._taxa['uid'])
+        df.index.name = 'uid'
 
         return df
 
@@ -312,7 +299,7 @@ class SpeciesEvolver(Component):
 
         # Create a dictionary of the taxa to update, `taxa_evolving`.
 
-        extant_taxa = np.array(self._taxa["object"])[self._taxa["extant"]]
+        extant_taxa = np.array(self._taxa['object'])[self._taxa['extant']]
         taxa_evolving = OrderedDict.fromkeys(extant_taxa, True)
 
         # Run stages of taxa evolution.
@@ -413,7 +400,7 @@ class SpeciesEvolver(Component):
         """
         time = self._record.latest_time
 
-        t_recorded = self._taxa["object"]
+        t_recorded = self._taxa['object']
 
         # Update previously introduced taxa.
 
@@ -422,9 +409,9 @@ class SpeciesEvolver(Component):
         for taxon in t_introduced:
             # Update taxon data.
 
-            idx = self._taxa["object"].index(taxon)
-            self._taxa["latest_time"][idx] = time
-            self._taxa["extant"][idx] = taxon.extant
+            idx = self._taxa['object'].index(taxon)
+            self._taxa['latest_time'][idx] = time
+            self._taxa['extant'][idx] = taxon.extant
 
         # Set the data of new taxa.
 
@@ -433,27 +420,28 @@ class SpeciesEvolver(Component):
         for taxon in t_new:
[…]
