+++ tests/components/species_evolution/test_base_taxon.py
@@ -22,10 +22,10 @@ class TestTaxon(Taxon):
 
     def _evolve(self, dt, stage, record):
         if stage == 0:
-            record.set_value("vara", 1)
+            record.set_value('vara', 1)
         elif stage == 1:
             TestTaxon(parent=self)
-            record.set_value("varb", 2)
+            record.set_value('varb', 2)
 
         return stage < 1
 
@@ -33,7 +33,7 @@ class TestTaxon(Taxon):
 def test_base_subclass():
     tt = TestTaxon()
     assert isinstance(tt, Taxon)
-    assert tt.__repr__() == "<TestTaxon, uid=None>"
+    assert tt.__repr__() == '<TestTaxon, uid=None>'
 
     np.testing.assert_equal(tt.range_mask, True)
 
@@ -43,7 +43,7 @@ def test_base_subclass():
     output = tt._evolve(10, 0, record)
     np.testing.assert_equal(output, True)
 
-    d = OrderedDict([("time", [0, 10]), ("vara", [np.nan, 1])])
+    d = OrderedDict([('time', [0, 10]), ('vara', [np.nan, 1])])
     np.testing.assert_equal(record._dict, d)
 
     output = tt._evolve(10, 1, record)
@@ -53,7 +53,9 @@ def test_base_subclass():
     np.testing.assert_equal(type(tt.children[0]), TestTaxon)
     np.testing.assert_equal(tt, tt.children[0].parent)
 
-    d = OrderedDict([("time", [0, 10]), ("vara", [np.nan, 1]), ("varb", [np.nan, 2])])
+    d = OrderedDict(
+        [('time', [0, 10]), ('vara', [np.nan, 1]), ('varb', [np.nan, 2])]
+    )
     np.testing.assert_equal(record._dict, d)
 
     tt.extant = False
