+++ tests/components/species_evolution/test_zone_objects.py
@@ -9,23 +9,25 @@ import pytest
 
 from landlab import RasterModelGrid
 from landlab.components import SpeciesEvolver
-from landlab.components.species_evolution import ZoneController, ZoneTaxon, zone as zn
+from landlab.components.species_evolution import zone as zn
+from landlab.components.species_evolution import ZoneController
+from landlab.components.species_evolution import ZoneTaxon
 
 
 @pytest.fixture()
 def zone_example_grid():
     mg = RasterModelGrid((5, 7), 2)
-    z = mg.add_zeros("node", "topographic__elevation")
+    z = mg.add_zeros('node', 'topographic__elevation')
     return mg, z
 
 
 def zone_func(grid):
-    z = grid.at_node["topographic__elevation"]
+    z = grid.at_node['topographic__elevation']
     return z == 1
 
 
 def zone_func_with_vars(grid, var1, var2):
-    z = grid.at_node["topographic__elevation"]
+    z = grid.at_node['topographic__elevation']
     return np.all([z == 1, grid.x_of_node > var1, grid.y_of_node > var2], 0)
 
 
@@ -50,17 +52,17 @@ def test_none_to_one(zone_example_grid):
 
     np.testing.assert_equal(len(sc.zones), 0)
 
-    expected_df = pd.DataFrame(
-        {
-            "time": [0],
-            "zones": [0],
-            "fragmentations": [np.nan],
-            "captures": [np.nan],
-            "area_captured_sum": [np.nan],
-            "area_captured_max": [np.nan],
-        }
+    expected_df = pd.DataFrame({
+        'time': [0],
+        'zones': [0],
+        'fragmentations': [np.nan],
+        'captures': [np.nan],
+        'area_captured_sum': [np.nan],
+        'area_captured_max': [np.nan]}
     )
-    pd.testing.assert_frame_equal(sc.record_data_frame, expected_df, check_like=True)
+    pd.testing.assert_frame_equal(
+        sc.record_data_frame, expected_df, check_like=True
+    )
 
     # Create a zone for time 1.
 
@@ -72,17 +74,17 @@ def test_none_to_one(zone_example_grid):
     np.testing.assert_equal(len(sc.zones), 1)
     np.testing.assert_equal(sc.zones[0]._conn_type, zn.Connection.NONE_TO_ONE)
 
-    expected_df = pd.DataFrame(
-        {
-            "time": [0, 1],
-            "zones": [0, 1],
-            "fragmentations": [np.nan, 0],
-            "captures": [np.nan, 0],
-            "area_captured_sum": [np.nan, 0],
-            "area_captured_max": [np.nan, 0],
-        }
+    expected_df = pd.DataFrame({
+        'time': [0, 1],
+        'zones': [0, 1],
+        'fragmentations': [np.nan, 0],
+        'captures': [np.nan, 0],
+        'area_captured_sum': [np.nan, 0],
+        'area_captured_max': [np.nan, 0]}
     )
-    pd.testing.assert_frame_equal(sc.record_data_frame, expected_df, check_like=True)
+    pd.testing.assert_frame_equal(
+        sc.record_data_frame, expected_df, check_like=True
+    )
 
     np.testing.assert_equal(se.record_data_frame.taxa.sum(), 0)
 
@@ -102,17 +104,17 @@ def test_one_to_none(zone_example_grid):
     np.testing.assert_equal(len(sc.zones), 1)
     zone = sc.zones[0]
 
-    expected_df = pd.DataFrame(
-        {
-            "time": [0],
-            "zones": [1],
-            "fragmentations": [np.nan],
-            "captures": [np.nan],
-            "area_captured_sum": [np.nan],
-            "area_captured_max": [np.nan],
-        }
[…]
