### 2020-01-21 099ca00
move scratch to an untested folder (to be removed)
