### 2020-01-21 85c2e35
bit of component cleanup
- remove not necessary field, 
- make attributes all private
- use reach_length instead of link length.
