+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -42,6 +42,15 @@ _SUPPORTED_TRANSPORT_METHODS = ["WilcockCrowe"]
 
 _OUT_OF_NETWORK = NetworkModelGrid.BAD_INDEX - 1
 
+_REQUIRED_PARCEL_ATTRIBUTES = [
+    "time_arrival_in_link",
+    "abrasion_rate",
+    "density",
+    "active_layer",
+    "location_in_link",
+    "D",
+    "volume",
+]
 
 _ACTIVE = 1
 _INACTIVE = 0
@@ -87,13 +96,19 @@ class NetworkSedimentTransporter(Component):
 
     >>> nmg = NetworkModelGrid((y_of_node, x_of_node), nodes_at_link)
 
-    Add channel and topographic variables to the NetworkModelGrid.
-    >>> nmg.at_node["topographic__elevation"] = [3., 2., 1., 0.] # m
-    >>> nmg.at_node["bedrock__elevation"] = [3., 2., 1., 0.] # m
-    >>> nmg.at_link["drainage_area"] = [10e6, 10e6, 10e6]  # m2
-    >>> nmg.at_link["channel_slope"] = [0.001, 0.001, 0.001]
-    >>> nmg.at_link["link_length"] = [100., 100., 100.]  # m
-    >>> nmg.at_link["channel_width"] = (15 * np.ones(np.size(nmg.at_link["drainage_area"])))
+    Add required channel and topographic variables to the NetworkModelGrid.
+
+    >>> _ = nmg.add_field("bedrock__elevation", [3., 2., 1., 0.], at="node") # m
+    >>> _ = nmg.add_field("reach_length", [100., 100., 100.], at="link")  # m
+    >>> _ = nmg.add_field("channel_width", (15 * np.ones(nmg.size("link"))), at="link")
+
+    Add ``topographic__elevation`` to the grid because the ``FlowDirectorSteepest``
+    will look to it to determine the direction of sediment transport through the
+    network. Each time we run the ``NetworkSedimentTransporter`` the
+    topography will be updated based on the bedrock elevation and the
+    distribution of alluvium.
+
+    >>> _ = nmg.add_field("topographic__elevation", np.copy(nmg.at_node["bedrock__elevation"]), at="node")
 
     Run ``FlowDirectorSteepest`` to determine the direction of sediment transport through the network.
     >>> flow_director = FlowDirectorSteepest(nmg)
@@ -136,15 +151,15 @@ class NetworkSedimentTransporter(Component):
     Instantiate the model run
 
     >>> nst = NetworkSedimentTransporter(
-    >>>         nmg,
-    >>>         one_parcel,
-    >>>         flow_director,
-    >>>         example_flow_depth,
-    >>>         bed_porosity=0.03,
-    >>>         g=9.81,
-    >>>         fluid_density=1000,
-    >>>         transport_method="WilcockCrowe",
-    >>>     )
+    ...         nmg,
+    ...         one_parcel,
+    ...         flow_director,
+    ...         example_flow_depth,
+    ...         bed_porosity=0.03,
+    ...         g=9.81,
+    ...         fluid_density=1000,
+    ...         transport_method="WilcockCrowe",
+    ...     )
 
     >>> dt = 60  # (seconds) 1 min timestep
 
@@ -163,9 +178,17 @@ class NetworkSedimentTransporter(Component):
     __version__ = "1.0"
 
     _info = {
-        "channel_slope": {
+        "bedrock__elevation": {
             "dtype": float,
             "intent": "in",
+            "optional": False,
+            "units": "m",
+            "mapping": "node",
+            "doc": "elevation of the bedrock surface",
+        },
+        "channel_slope": {
+            "dtype": float,
+            "intent": "out",
             "optional": False,
             "units": "m/m",
             "mapping": "link",
[…]
