### 2020-01-21 3bd1d08
split up tests into different files, also make them all run
