+++ tests/components/network_sediment_transporter/test_init.py
@@ -21,7 +21,6 @@ def test_basic_init(
         bed_porosity=0.3,
         g=9.81,
         fluid_density=1000,
-        channel_width="channel_width",
         transport_method="WilcockCrowe",
     )
 
@@ -37,7 +36,6 @@ def test_bad_flow_director(example_nmg, example_parcels, example_flow_depth):
             bed_porosity=0.03,
             g=9.81,
             fluid_density=1000,
-            channel_width="channel_width",
             transport_method="WilcockCrowe",
         )
 
@@ -55,7 +53,6 @@ def test_bad_network_model_grid(
             bed_porosity=0.03,
             g=9.81,
             fluid_density=1000,
-            channel_width="channel_width",
             transport_method="WilcockCrowe",
         )
 
@@ -71,7 +68,6 @@ def test_bad_parcels(example_nmg, example_flow_director, example_flow_depth):
             bed_porosity=0.03,
             g=9.81,
             fluid_density=1000,
-            channel_width="channel_width",
             transport_method="WilcockCrowe",
         )
 
@@ -89,7 +85,6 @@ def test_bad_porosity(
             bed_porosity=-0.03,
             g=9.81,
             fluid_density=1000,
-            channel_width="channel_width",
             transport_method="WilcockCrowe",
         )
 
@@ -107,365 +102,5 @@ def test_bad_transport_method(
             bed_porosity=0.03,
             g=9.81,
             fluid_density=1000,
-            channel_width="channel_width",
             transport_method="bad_transport_method",
         )
-
-
-def test_no_flow_no_transport(example_nmg, example_parcels, example_flow_director):
-
-    timesteps = 3
-    no_flow = np.zeros([(timesteps + 1), (example_nmg.number_of_links)])
-
-    nst = NetworkSedimentTransporter(
-        example_nmg,
-        example_parcels,
-        example_flow_director,
-        no_flow,
-        bed_porosity=0.03,
-        g=9.81,
-        fluid_density=1000,
-        channel_width="channel_width",
-        transport_method="WilcockCrowe",
-    )
-
-    dt = 60 * 60 * 24  # (seconds) daily timestep
-
-    for t in range(0, (timesteps * dt), dt):
-        nst.run_one_step(dt)
-
-        # Need to define original_node_elev after a timestep has passed.
-        if t / (60 * 60 * 24) == 1:
-            original_node_elev = example_nmg.at_node["topographic__elevation"]
-
-    # No flow? Parcels should stay in same locations in links
-    assert_array_almost_equal(
-        example_parcels.dataset.location_in_link[:, 0],
-        example_parcels.dataset.location_in_link[:, 1],
-    )
-
-    # No flow? Total parcel volume should stay the same
-    assert_array_equal(
-        np.sum(example_parcels.dataset["volume"].values, axis=0)[0],
-        np.sum(example_parcels.dataset["volume"].values, axis=0)[1],
-    )
-
-    # No flow? Elevations should stay the same
-    assert_array_equal(
-        original_node_elev, example_nmg.at_node["topographic__elevation"]
-    )
-
-
-def test_parcel_leaves(example_nmg, example_flow_depth, example_flow_director):
-
-    time = [0.0]  # probably not the sensible way to do this...
-
-    items = {"grid_element": "link", "element_id": np.array([[6]])}
-
-    initial_volume = np.array([[1]])
[…]
