+++ tests/components/network_sediment_transporter/test_transport.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pytest
 from numpy.testing import assert_array_almost_equal, assert_array_equal
 
 from landlab.components import FlowDirectorSteepest, NetworkSedimentTransporter
