+++ tests/components/network_sediment_transporter/test_filo.py
@@ -1,6 +1,5 @@
 import numpy as np
-import pytest
-from numpy.testing import assert_array_almost_equal, assert_array_equal
+from numpy.testing import assert_array_equal
 
 from landlab.components import FlowDirectorSteepest, NetworkSedimentTransporter
 from landlab.data_record import DataRecord
