+++ tests/components/network_sediment_transporter/test_init.py
@@ -1,9 +1,6 @@
-import numpy as np
 import pytest
-from numpy.testing import assert_array_almost_equal, assert_array_equal
 
-from landlab.components import FlowDirectorSteepest, NetworkSedimentTransporter
-from landlab.data_record import DataRecord
+from landlab.components import NetworkSedimentTransporter
 from landlab.grid.network import NetworkModelGrid
 
 _OUT_OF_NETWORK = NetworkModelGrid.BAD_INDEX - 1
@@ -13,7 +10,7 @@ def test_basic_init(
     example_nmg, example_parcels, example_flow_director, example_flow_depth
 ):
 
-    nst = NetworkSedimentTransporter(
+    _ = NetworkSedimentTransporter(
         example_nmg,
         example_parcels,
         example_flow_director,
