### 2020-01-21 b5fc357
exempt NST from this part of test components
