+++ tests/components/test_components.py
@@ -19,6 +19,7 @@ def test_component_metadata(Comp):
         "HackCalculator",
         "Lithology",
         "LithoLayers",
+        "NetworkSedimentTransporter",
         "Profiler",
         "SoilMoisture",
         "Vegetation",
