+++ docs/source/development/index.rst
@@ -15,6 +15,9 @@ The most important things to remember are to:
     `master` branch; and
   * Keep your fork's `master` and development branches up to date with changes
     in the main Landlab repository.
+  * Consider creating a small Pull Request to update the page on
+    :ref:`ongoing development <ongoing_development>` so that others know what
+    you are working on.
 
 .. toctree::
    :maxdepth: 2
@@ -23,3 +26,14 @@ The most important things to remember are to:
    contribution/index
    practices/index
    package_organization
+   
+There are a few pages that are particularly important for the maintenance of
+the package. These are:
+
+* :ref:`directory organization <organization>`
+* :ref:`testing protocol <testing>`
+* :ref:`dependency organization <dependencies>`, and
+* :ref:`release workflow <dev_releases>`
+
+If package maintainers change any of these, the prior pages likely need
+updating.
