### 2020-01-21 0ff9ccb
improve contributing docs
