+++ docs/source/development/contribution/index.rst
@@ -13,4 +13,24 @@ Please review the :ref:`development practices <development_practices>`.
    recommendations
    ongoing_development
    desired_contributions
+
+Contributions that change high level Landlab organization should update
+:ref:`this page <organization>` in the documentation.
+
+
+A number of researchers have used the Journal of Open Source Software (JOSS)
+and Geoscientific Model Development as a journal outlet to publish their
+contribution to Landlab. There are a few aspects of the submission workflow for
+JOSS that are non-standard and are described below.
+
+.. toctree::
+   :maxdepth: 2
+
    joss_workflow
+
+In addition, the JOSS editorial board and the Landlab core development team are
+presently (Jan 2019) working on defining some guidelines regarding what
+contributions to Landlab fit the scope of a JOSS submission. These will be
+summarized here when finalized. If you have any questions regarding whether
+your potential JOSS submission is appropriate, the best thing to do is to
+make a pre-submission inquiry with JOSS.
