### 2020-01-21 a30154f
address @nathonlyons comments (except missing links)
