### 2020-01-21 6dafdad
remove duplicate statement
