### 2020-01-21 c274ad7
revising contribution guidelines
including removing duplicated material.
