+++ CONTRIBUTING.md
@@ -1,6 +1,6 @@
 # Contributing to landlab
 
-Thank you for contributing to the landlab! We appreciate
+Thank you for contributing to Landlab! We appreciate
 your help as this is largely as volunteer effort! :heart: :heart: :heart:
 
 # How to contribute
@@ -8,8 +8,9 @@ your help as this is largely as volunteer effort! :heart: :heart: :heart:
 ## Reporting Bugs
 
 Before creating a bug report, please do at least a cursory check that the
-bug has not already been reported. If it has, add a comment to the existing
-issue instead of opening a new one.
+bug has not already been reported by searching the Issues portion of the
+GitHub repository. If it has, add a comment to the existing issue instead of
+opening a new one.
 
 ### Submitting a Bug Report
 
@@ -29,7 +30,9 @@ to track down the source of the problem.
    that demonstrates the steps** as, for example, a bash script
    along with input files. This example should reproduce your
    problem with as few lines of code as possible and easily
-   reproducible my another person.
+   reproducible my another person. Such an example almost certainly will not
+   include an input file or any dependencies beyond those required by the
+   `landlab_dev` conda environment.
 *  **Describe the behavior you are seeing after these steps**.
 *  **Describe the behavior you expect to see after these steps**.
 
@@ -37,7 +40,8 @@ Additionally, the answers to the following questions about your run
 environment will be helpful.
 
 *  **Which version of landlab are you using?** This could be a specific
-   git sha or a release number.
+   git sha or a release number. The best way to find this information is to
+   import landlab and evaluate `landlab.__version__`
 *  **What is he name and version of you OS?**
 *  **What compiler are you using?**
 *  **How did you build landlab (if using the development version)?**
@@ -54,14 +58,7 @@ Changes to landlab should be submitted as
 *  Create a topic branch that contains your changes.
 *  Open a new [GitHub pull request](https://github.com/landlab/landlab/pull/new/master).
 *  Ensure the pull request description clearly describes the problem
-   and solution.  Include the relevant issue number.
-
-## Styleguides
-
-Use the [PEP8 style guide](https://www.python.org/dev/peps/pep-0008/).
-You may want to use tools like
-[Prospector](http://prospector.landscape.io/en/master/) to help out
-with this.
+   and solution. Include the relevant issue number.
 
 ### Git Commit Messages
 
@@ -91,58 +88,15 @@ with this.
 
 ## Adding new components
 
-If you would like to create a new component, we have just a few
-conventions that we would like you to follow.
+If you would like to create a new component, we a few conventions that we would
+like you to follow.
 
-* Create a new folder under `landlab/components` that will hold your
-  new component.
+Please visit [this part](https://landlab.readthedocs.io/en/master/development/index.html)
+of the main Landlab documentation page to read about developer installation,
+guidelines to contributing code, and our software development practices.
 
-### Landlab 2 is Python >=3.6 only.
-
-* Landlab 2 is the standard starting in January 2020.
-* All tests will be run with Python 3.6, 3.7, 3.8.
-* If you need to introduce a new dependency, that dependency must
-  be compatible with Python  3.6+.
-
-### Documentation
-
-* All public functions, classes, methods, etc. must have a docstring
-  that follows the [numpydoc](https://github.com/numpydoc/numpydoc)
-  conventions.
-* Every `.py` file must contain a module-level docstring at the top
-  of the file that describes what the purpose of the file is.
-* Additionally, you will need to add a ReStructuredText (.rst) file
-  so that the documentation of your new part of Landlab can be
-  autogenerated. See [this section of the creating a component](https://landlab.readthedocs.io/en/latest/development/contribution/index.html) documentation
-  for an example of how to do this.
-
-### Testing
-
[…]
