### 2020-01-21 96bcf9e
incorporating material removed from contributing.md into the docs.
