### 2020-01-21 3be6726
address @gregtucker comments
