### 2020-01-21 331de35
make @gregtucker's revision in the readme.rst as well
