### 2020-01-21 98aae3b
first round of improvements to nst
passes tests
