+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -781,20 +781,26 @@ class NetworkSedimentTransporter(Component):
         """Method to update parcel location for each parcel in the active
         layer.
         """
-
-        # we need to make sure we are pointing to the array rather than making copies
+        # determine where parcels are starting
         current_link = self._parcels.dataset.element_id[
             :, self._time_idx
-        ]  # same as Linkarray, this will be updated below
+        ]
+
+        # determine location within link where parcels are starting.
         location_in_link = self._parcels.dataset.location_in_link[
             :, self._time_idx
-        ]  # updated below
+        ]
+
+        # determine how far each parcel needs to travel this timestep.
         distance_to_travel_this_timestep = (
             self._pvelocity * dt
         )  # total distance traveled in dt at parcel virtual velocity
         # ^ movement in current and any DS links at this dt is at the same velocity as in the current link
         # ... perhaps modify in the future(?) or ensure this type of travel is kept to a minimum
         # ... or display warnings or create a log file when the parcel jumps far in the next DS link
+
+        # Accumulate the total distance traveled by a parcel for abrasion rate
+        # calculations.
         if np.size(self._distance_traveled_cumulative) != np.size(
             distance_to_travel_this_timestep
         ):
@@ -805,18 +811,18 @@ class NetworkSedimentTransporter(Component):
             self._distance_traveled_cumulative += distance_to_travel_this_timestep
             # ^ accumulates total distanced traveled for testing abrasion
 
-        # print("Median travel distance of active layer parcels this timestep = ",
-        #       np.median(
-        #               distance_to_travel_this_timestep[
-        #                       self._parcels.dataset.active_layer[:,-1]==1
-        #                       ]
-        #               ))
+        # get the downstream link at link:
+        downstream_link_at_link = self._fd.link_to_flow_receiving_node[self._fd.downstream_node_at_link()]
 
-        #        if self._time_idx == 1:
-        #            print("t", self._time_idx)
+        # active parcels on the network:
+        in_network = self._parcels.dataset.element_id.values[:, self._time_idx] != _OUT_OF_NETWORK
+        active = distance_to_travel_this_timestep>0.0
+        active_parcel_ids = np.nonzero(in_network*active)[0] # this line broken.
 
-        for p in range(self._parcels.number_of_items):
+        # for each parcel.
+        for p in active_parcel_ids:
 
+            # Step 1: Move parcel far enough downstream.
             distance_to_exit_current_link = self._grid.at_link["reach_length"][
                 int(current_link[p])
             ] * (1 - location_in_link[p])
@@ -827,96 +833,93 @@ class NetworkSedimentTransporter(Component):
             ] * (location_in_link[p])
 
             running_travel_distance_in_dt = 0  # initialize to 0
-
             distance_left_to_travel = distance_to_travel_this_timestep[p]
-            # if parcel in network at end of last timestep
+            while (
+                running_travel_distance_in_dt + distance_to_exit_current_link
+            ) <= distance_to_travel_this_timestep[p]:
+                # distance_left_to_travel > 0:
+                # ^ loop through until you find the link the parcel will reside in after moving
+                # ... the total travel distance
 
-            if self._parcels.dataset.element_id[p, self._time_idx] != _OUT_OF_NETWORK:
-
-                while (
+                # update running travel distance now that you know the parcel will move through the
+                # ... current link
+                running_travel_distance_in_dt = (
                     running_travel_distance_in_dt + distance_to_exit_current_link
-                ) <= distance_to_travel_this_timestep[p]:
-                    # distance_left_to_travel > 0:
-                    # ^ loop through until you find the link the parcel will reside in after moving
[…]
