### 2020-01-21 650db18
this variable isn't used outside of this function. so I'm removing it from the component
