### 2020-01-21 f61491c
reduce some repetition.
