+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -124,6 +124,7 @@ class NetworkSedimentTransporter(Component):
     ...     np.tile(1, (timesteps + 1, 1))) # 2 meter flow depth
 
     Define the sediment characteristics that will be used to create the parcels ``DataRecord``
+
     >>> items = {"grid_element": "link",
     ...          "element_id": np.array([[0]])}
 
@@ -140,6 +141,7 @@ class NetworkSedimentTransporter(Component):
 
     Create the sediment parcel DataRecord. In this case, we are creating a single
     sediment parcel with all of the required attributes.
+
     >>> one_parcel = DataRecord(
     ...     nmg,
     ...     items=items,
@@ -318,7 +320,7 @@ class NetworkSedimentTransporter(Component):
         self._fluid_density = fluid_density
         self._time_idx = 0
         self._time = 0.0
-        self._distance_traveled_cumulative = np.zeros([self._num_parcels])
+        self._distance_traveled_cumulative = np.zeros(self._num_parcels)
 
         # check the transport method is valid.
         if transport_method in _SUPPORTED_TRANSPORT_METHODS:
@@ -440,10 +442,12 @@ class NetworkSedimentTransporter(Component):
                 d_i = Darray[Linkarray == i]
                 vol_i = Volarray[Linkarray == i]
                 rhos_i = Rhoarray[Linkarray == i]
+
+
                 vol_tot_i = np.sum(vol_i)
 
                 d_mean_active[i] = np.sum(d_i * vol_i) / (vol_tot_i)
-                self.d_mean_active = d_mean_active
+                self._d_mean_active = d_mean_active
 
                 rhos_mean_active[i] = np.sum(rhos_i * vol_i) / (vol_tot_i)
                 self._rhos_mean_active = rhos_mean_active
@@ -458,11 +462,11 @@ class NetworkSedimentTransporter(Component):
         taustar = tau / (
             (self._rhos_mean_active - self._fluid_density)
             * self._g
-            * self.d_mean_active
+            * self._d_mean_active
         )
 
         self._active_layer_thickness = (
-            0.515 * self.d_mean_active * (3.09 * (taustar - 0.0549) ** 0.56)
+            0.515 * self._d_mean_active * (3.09 * (taustar - 0.0549) ** 0.56)
         )  # in units of m
 
         self._active_layer_thickness[
@@ -546,7 +550,6 @@ class NetworkSedimentTransporter(Component):
 
         self._vol_stor = (vol_tot - vol_act) / (1 - self._bed_porosity)
 
-    # %%
     def _adjust_node_elevation(self):
         """Adjusts slope for each link based on parcel motions from last
         timestep and additions from this timestep.
@@ -689,8 +692,8 @@ class NetworkSedimentTransporter(Component):
             vol_act_sand = np.zeros(self._grid.number_of_links)
 
         frac_sand = np.zeros_like(vol_act)
-        frac_sand[vol_act != 0] = vol_act_sand[vol_act != 0] / vol_act[vol_act != 0]
-        frac_sand[np.isnan(frac_sand)] = 0
+        frac_sand[vol_act != 0.0] = vol_act_sand[vol_act != 0.0] / vol_act[vol_act != 0.0]
+        frac_sand[np.isnan(frac_sand)] = 0.0
 
         # Calc attributes for each link, map to parcel arrays
         for i in range(self._grid.number_of_links):
@@ -701,14 +704,14 @@ class NetworkSedimentTransporter(Component):
             rhos_act_i = Rhoarray[active_here]
             vol_act_tot_i = np.sum(vol_act_i)
[…]
