+++ tests/components/network_sediment_transporter/test_transport.py
@@ -122,7 +122,6 @@ def test_defined_parcel_transport():
     for t in range(0, (timesteps * dt), dt):
         nst.run_one_step(dt)
         distance_traveled[np.int(t / dt)] = nst._distance_traveled_cumulative
-        active_layer_thickness_array[np.int(t / dt)] = nst._active_layer_thickness_array
     # NEED TO CALCULATE THINGS HERE.
     # Transport distance should match?
     Distance_Traveled_Should_Be = [
