### 2020-01-21 c7b343b
some reorg in calc_wilcock crowe
