+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -635,15 +635,29 @@ class NetworkSedimentTransporter(Component):
 
         Note: could have options here (e.g. Wilcock and Crowe, FLVB, MPM, etc)
         """
+        # Initialize _pvelocity, the virtual velocity of each parcel (link length / link travel time)
+        self._pvelocity = np.zeros(self._num_parcels)
+
+        # Calculate total volume of sediment at each link.
+        vol_tot = self._parcels.calc_aggregate_value(
+            np.sum, "volume", at="link", filter_array=self._find_now, fill_value=0.0
+        )
+
+        # Calculate active volume of sediment at each link.
+        if np.any(self._active_parcel_records):
+            vol_act = self._parcels.calc_aggregate_value(
+                np.sum, "volume", at="link", filter_array=self._active_parcel_records, fill_value=0.0
+            )
+        else:
+            vol_act = np.zeros_like(vol_tot)
+
         # parcel attribute arrays from DataRecord
 
         Darray = self._parcels.dataset.D[:, self._time_idx]
         Activearray = self._parcels.dataset.active_layer[:, self._time_idx].values
         Rhoarray = self._parcels.dataset.density.values
         Volarray = self._parcels.dataset.volume[:, self._time_idx].values
-        Linkarray = self._parcels.dataset.element_id[
-            :, self._time_idx
-        ].values  # link that the parcel is currently in
+        Linkarray = self._parcels.dataset.element_id[:, self._time_idx].values  # link that the parcel is currently in
 
         R = (Rhoarray - self._fluid_density) / self._fluid_density
 
@@ -657,25 +671,6 @@ class NetworkSedimentTransporter(Component):
         active_layer_thickness_array = np.zeros(self._num_parcels) * np.nan
         #        rhos_mean_active = np.zeros(self._num_parcels)
         #        rhos_mean_active.fill(np.nan)
-        self._Ttimearray = np.zeros(self._num_parcels)
-        # ^ Ttimearray is the time to move through the entire length of a link
-        self._pvelocity = np.zeros(self._num_parcels)
-        # ^ pvelocity is the parcel virtual velocity = link length / link travel time
-
-        # Calculate bed statistics for all of the links
-        vol_tot = self._parcels.calc_aggregate_value(
-            np.sum, "volume", at="link", filter_array=self._find_now
-        )
-        vol_tot[np.isnan(vol_tot) == 1] = 0
-
-        if np.any(self._active_parcel_records):
-            vol_act = self._parcels.calc_aggregate_value(
-                np.sum, "volume", at="link", filter_array=self._active_parcel_records
-            )
-            vol_act[np.isnan(vol_act) == 1] = 0
-
-        else:
-            vol_act = np.zeros_like(vol_tot)
 
         # find active sand.
         findactivesand = (
@@ -700,7 +695,7 @@ class NetworkSedimentTransporter(Component):
         # Calc attributes for each link, map to parcel arrays
         for i in range(self._grid.number_of_links):
 
-            active_here = np.where(np.logical_and(Linkarray == i, Activearray == 1))[0]
+            active_here = np.nonzero(np.logical_and(Linkarray == i, Activearray == _ACTIVE))[0]
             d_act_i = Darray[active_here]
             vol_act_i = Volarray[active_here]
             rhos_act_i = Rhoarray[active_here]
