### 2020-01-21 fc8ab71
use xarray functions
