### 2020-01-22 cb956fc
Additional Doc Updates (#1117)
* start on additional doc updates

* typos

* add a tree view for package organization.

* improve cite as attribute

* add references

* create package organziation.

* improve citation documentation

* misc docs

* add all component references 

passes pretty and lint

* fix circular import

* improve formatting

* four GSA dois don't work (gsa pubs seem to have link check issues)

* undo all changes to compont/species_evolution submodule

* undo changes to tests/components/species_evolution

* fix lint/docs

* address @nathanlyons comments

* add another examp
[…]
