+++ landlab/components/drainage_density/drainage_density.py
@@ -8,25 +8,25 @@ from landlab import Component
 
 
 class DrainageDensity(Component):
-
-    r"""Calculate drainage density over a DEM.
+    r"""
+    Calculate drainage density over a DEM.
 
     Landlab component that implements the distance to channel algorithm of
     Tucker et al., 2001.
 
-    This component requires EITHER a channel__mask array with 1's
+    This component requires EITHER a ``channel__mask array`` with 1's
     where channels exist and 0's elsewhere, OR a set of coefficients
     and exponents for a slope-area relationship and a
     channelization threshold to compare against that relationship.
 
-    If an array is provided it MUST be of type np.uint8. See the example below
-    for how to make such an array.
+    If an array is provided it MUST be of type ``np.uint8``. See the example
+    below for how to make such an array.
 
-    The channel__mask array will be assigned to an at-node field with the name
-    `channel__mask`. If the channel__mask was originaly created from a passed
-    array, a user can update this array to change the mask.
+    The ``channel__mask`` array will be assigned to an at-node field with the
+    name ``channel__mask``. If the channel__mask was originaly created from a
+    passed array, a user can update this array to change the mask.
 
-    If the channel__mask is created using an area coefficent,
+    If the ``channel__mask`` is created using an area coefficent,
     slope coefficient, area exponent, slope exponent, and channelization
     threshold, the location of the mask will be re-update when
     calculate_drainage_density is called.
@@ -38,29 +38,106 @@ class DrainageDensity(Component):
 
     .. math::
 
-        C_A A^{m_r} C_s S^{n_r} > T_c
+       C_A A^{m_r} C_s S^{n_r} > T_c
 
     where :math:`A` is the drainage density and :math:`S` is the local slope,
     will be marked as channel nodes.
 
-   ``calculate_drainage_density`` function returns drainage density for the model
-    domain.
-
-    ``calculate_drainage_density`` calculates the distance from every node to the
+    The ``calculate_drainage_density`` function returns drainage density for the
+    model domain. This function calculates the distance from every node to the
     nearest channel node :math:`L` along the flow line of steepest descent
     (assuming D8 routing).
 
-
     This component stores this distance a field, called:
-    ``surface_to_channel__minimum_distance``.
-
-    The drainage density is then calculated (after Tucker et al., 2001):
+    ``surface_to_channel__minimum_distance``. The drainage density is then
+    calculated (after Tucker et al., 2001):
 
     .. math::
 
-        D_d = \frac{1}{2\overline{L}}
+       D_d = \frac{1}{2\overline{L}}
 
     where :math:`\overline{L}` is the mean L for the model domain.
+
+    Examples
+    --------
+    >>> import numpy as np
+    >>> from landlab import RasterModelGrid
+    >>> from landlab.components import FlowAccumulator, FastscapeEroder
+    >>> mg = RasterModelGrid((10, 10))
+    >>> _ = mg.add_zeros('node', 'topographic__elevation')
+    >>> np.random.seed(50)
+    >>> noise = np.random.rand(100)
+    >>> mg.at_node['topographic__elevation'] += noise
[…]
