+++ docs/source/citation_registry.rst
@@ -1,0 +1,11 @@
+.. _cite_as:
+
+:orphan:
+
+Landlab Citation Registry
+-------------------------
+
+.. automodule:: landlab._registry
+    :members:
+    :undoc-members:
+    :show-inheritance:
