### 2020-01-22 c58bfcb
Update Main Repository Readme.md and GroundwaterDupuitPercolator readme.md (#1112)
* work on improving the readme

* additional readme updates

* main readme and contributing updates.

* address @nathonlyons comments (except missing links)

* remove duplicate statement

* add links to documentation.

* revising contribution guidelines 

including removing duplicated material.

* make @gregtucker's revision in the readme.rst as well
