+++ README.rst
@@ -20,33 +20,248 @@
  :target: https://mybinder.org/v2/gh/landlab/landlab/release?filepath=notebooks/welcome.ipynb
 
 =======
-landlab
+Landlab
 =======
 
-The Landlab project creates an environment in which scientists can build a
-numerical landscape model without having to code all of the individual
-components. Landscape models compute flows of mass, such as water, sediment,
-glacial ice, volcanic material, or landslide debris, across a gridded terrain
-surface. Landscape models have a number of commonalities, such as operating on
-a grid of points and routing material across the grid. Scientists who want to
-use a landscape model often build their own unique model from the ground up,
-re-coding the basic building blocks of their landscape model rather than
-taking advantage of codes that have already been written.
+What does Landlab do?
+---------------------
 
-More information can be found at the website:
+Landlab is an open-source Python-language package for numerical modeling of
+Earth surface dynamics. It contains
 
-http://landlab.github.io
+* A gridding engine which represents the model domain. Regular and irregular
+  grids are supported.
+* A library of process components, each of which represents a physical process
+  (e.g., generation of rain, erosion by flowing water). These components have
+  a common interface and can be combined based on a user's needs.
+* Utilities that support general numerical methods, file input/output, and
+  visualization.
 
-This open-source manuscript is a gateway for entering the Landlab world:
+In addition Landlab contains a set of Jupyter notebook tutorials providing
+an introduction to core concepts and examples of use.
 
-https://www.earth-surf-dynam.net/5/21/2017/
+Landlab was desiged for disciplines that quantify Earth surface dynamics such
+as geomorphology, hydrology, glaciology, and stratigraphy. It can also be used
+in related fields. Scientists who use this type of model often build
+their own unique model from the ground up, re-coding the basic building blocks
+of their landscape model rather than taking advantage of codes that have
+already been written. Landlab saves practitioners from the need for this kind
+ of re-invention by providing standardized components that they can re-use.
 
-Two main installation options exist for Landlab. Most people will likely want to
-`install the conda package <https://landlab.readthedocs.io/en/latest/install/index.html>`_.
+Watch the webinar `Landlab Toolkit Overview <https://csdms.colorado.edu/wiki/Presenters-0407>`_
+ at CSDMS to learn more.
+
+How do I install Landlab?
+-------------------------
+
+First you'll need a Python distribution and either the conda or pip package
+manager. If you don't know what you want, we recommend the Anaconda Python
+distribution.
+
+Two main installation options exist for Landlab. Most people will likely want
+to
+`install a prepackaged binary <https://landlab.readthedocs.io/en/latest/install/index.html>`_.
+We distribute through both conda-forge and pip.
+
+Conda Environment with Binary
+`````````````````````````````
+
+To create a conda environment that installs a pre-packaged binary and all the
+dependencies necessary to run the notebooks, clone the repository, navigate to
+within the top level directory and use the following command:
+
+.. code-block:: bash
+
+    $ conda env create --file=environment.yml
+
+Then activate the environment and open the welcome notebook execute the
+following:
+
+  .. code-block:: bash
+
+      $ conda activate landlab_notebooks
+      $ jupyter notebook notebooks/welcome.ipynb
+
+Developer Installation
+``````````````````````
+
 Individuals interested in modifying the Landlab source code should follow the
[…]
