### 2020-01-22 landlab/landlab#1108 @kbarnhart
@DavidLitwin could you look through the changes that I've made to the readme and the docs and verify that I've addressed @rreinecke 's concerns?
