### 2020-01-22 2d05bac
Update soil_moisture_dynamics.py
minor correction
