### 2020-01-22 landlab/landlab#1108 @DavidLitwin
@kbarnhart I think it looks good in general. A few notes on the groundwater readme:
- the word "markdown" appears at the beginning of the document, is that supposed to be there?
- can the TODO for requirements be updated now?
- I'm still not sure we've satisfied the request to know where unit tests live. I will admit, with the new LL2.0 arrangement, it took me a second as well to find them in the testing folder. Do you think we should include more explicit directions on where to find them, or is it sufficient to give instructions on how to run them?
