### 2020-01-22 bb2db39
Update gdp docstring
Add notes about Dupuit-Forcheimer assumptions
