### 2020-01-22 21aee9b
add more info about where unit tests are located. few other changes.
