### 2020-01-22 landlab/landlab#1105 @DavidLitwin
I have addressed this [here](https://github.com/landlab/landlab/commit/bb2db392b868a9f96f6d2a8829911d3187856d7c) by updating the docstring for the component to caution users about model assumptions.

The supplemental theory page [here](https://landlab.readthedocs.io/en/master/reference/components/dupuit_theory.html#dupuit-theory) also describes the assumptions made when using the Dupuit-Forcheimer approximation, and how this affects the governing equations solved.
