### 2020-01-22 3654b37
vectorize moving downstream
