+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -408,10 +408,16 @@ class NetworkSedimentTransporter(Component):
         active or storage layer during this timestep, then updates node
         elevations.
         """
-
-        vol_tot = self._parcels.calc_aggregate_value(
-            np.sum, "volume", at="link", filter_array=self._this_timesteps_parcels, fill_value=0.
-        )
+        try:
+            vol_tot = self._parcels.calc_aggregate_value(
+                np.sum,
+                "volume",
+                at="link",
+                filter_array=self._this_timesteps_parcels,
+                fill_value=0.0,
+            )
+        except ValueError:
+            vol_tot = np.zeros(self._grid.size("link"))
         # Wong et al. (2007) approximation for active layer thickness.
         # NOTE: calculated using grain size and grain density calculated for
         # the active layer grains in each link at the **previous** timestep.
@@ -420,7 +426,9 @@ class NetworkSedimentTransporter(Component):
         # active layer.
 
         current_parcels = self._parcels.dataset.isel(time=self._time_idx)
-        if self._time_idx == 0: #KRB-Q: Why only the first timestep. Parcels will move around... so shouldn't this happen every timestep.
+        if (
+            self._time_idx == 0
+        ):  # KRB-Q: Why only the first timestep. Parcels will move around... so shouldn't this happen every timestep.
 
             # In the first full timestep, we need to calc grain size & rho_sed.
             # Assume all parcels are in the active layer for the purposes of
@@ -430,26 +438,33 @@ class NetworkSedimentTransporter(Component):
             # has already been calculated (e.g. during 'zeroing' runs)
 
             # Calculate mean values for density and grain size (weighted by volume).
-            sel_parcels = current_parcels.where(current_parcels.element_id != _OUT_OF_NETWORK)
+            sel_parcels = current_parcels.where(
+                current_parcels.element_id != _OUT_OF_NETWORK
+            )
 
-            d_weighted = sel_parcels.D*sel_parcels.volume
-            rho_weighted = sel_parcels.density*sel_parcels.volume
+            d_weighted = sel_parcels.D * sel_parcels.volume
+            rho_weighted = sel_parcels.density * sel_parcels.volume
             d_weighted.name = "d_weighted"
             rho_weighted.name = "rho_weighted"
 
-            grouped_by_element = xr.merge((sel_parcels.element_id,
-                             sel_parcels.volume,
-                             d_weighted,
-                             rho_weighted)).groupby("element_id")
+            grouped_by_element = xr.merge(
+                (sel_parcels.element_id, sel_parcels.volume, d_weighted, rho_weighted)
+            ).groupby("element_id")
 
-            d_avg = grouped_by_element.sum().d_weighted/grouped_by_element.sum().volume
-            rho_avg = grouped_by_element.sum().rho_weighted/grouped_by_element.sum().volume
+            d_avg = (
+                grouped_by_element.sum().d_weighted / grouped_by_element.sum().volume
+            )
+            rho_avg = (
+                grouped_by_element.sum().rho_weighted / grouped_by_element.sum().volume
+            )
 
             self._d_mean_active = np.zeros(self._grid.size("link"))
             self._d_mean_active[d_avg.element_id.values.astype(int)] = d_avg.values
 
             self._rhos_mean_active = np.zeros(self._grid.size("link"))
-            self._rhos_mean_active[rho_avg.element_id.values.astype(int)] = rho_avg.values
+            self._rhos_mean_active[
+                rho_avg.element_id.values.astype(int)
+            ] = rho_avg.values
 
         tau = (
             self._fluid_density
@@ -468,7 +483,9 @@ class NetworkSedimentTransporter(Component):
[…]
