### 2020-01-22 7d3a0d7
a bit more cleanup
