+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -375,7 +375,6 @@ class NetworkSedimentTransporter(Component):
                     :, self._time_idx
                 ] = self._parcels.dataset[at].values[:, self._time_idx - 1]
 
-        self._find_now = self._parcels.dataset.time == self._time
         self._this_timesteps_parcels = np.zeros_like(
             self._parcels.dataset.element_id, dtype=bool
         )
@@ -409,15 +408,11 @@ class NetworkSedimentTransporter(Component):
         elevations.
         """
         try:
-            vol_tot = self._parcels.calc_aggregate_value(
-                np.sum,
-                "volume",
-                at="link",
-                filter_array=self._this_timesteps_parcels,
-                fill_value=0.0,
+            self._vol_tot = self._parcels.calc_aggregate_value(
+                np.sum, "volume", at="link", filter_array=self._this_timesteps_parcels, fill_value=0.
             )
         except ValueError:
-            vol_tot = np.zeros(self._grid.size("link"))
+            self._vol_tot = np.zeros(self._grid.size("link"))
         # Wong et al. (2007) approximation for active layer thickness.
         # NOTE: calculated using grain size and grain density calculated for
         # the active layer grains in each link at the **previous** timestep.
@@ -426,9 +421,7 @@ class NetworkSedimentTransporter(Component):
         # active layer.
 
         current_parcels = self._parcels.dataset.isel(time=self._time_idx)
-        if (
-            self._time_idx == 0
-        ):  # KRB-Q: Why only the first timestep. Parcels will move around... so shouldn't this happen every timestep.
+        if self._time_idx == 0: #KRB-Q: Why only the first timestep. Parcels will move around... so shouldn't this happen every timestep.
 
             # In the first full timestep, we need to calc grain size & rho_sed.
             # Assume all parcels are in the active layer for the purposes of
@@ -438,33 +431,26 @@ class NetworkSedimentTransporter(Component):
             # has already been calculated (e.g. during 'zeroing' runs)
 
             # Calculate mean values for density and grain size (weighted by volume).
-            sel_parcels = current_parcels.where(
-                current_parcels.element_id != _OUT_OF_NETWORK
-            )
+            sel_parcels = current_parcels.where(current_parcels.element_id != _OUT_OF_NETWORK)
 
-            d_weighted = sel_parcels.D * sel_parcels.volume
-            rho_weighted = sel_parcels.density * sel_parcels.volume
+            d_weighted = sel_parcels.D*sel_parcels.volume
+            rho_weighted = sel_parcels.density*sel_parcels.volume
             d_weighted.name = "d_weighted"
             rho_weighted.name = "rho_weighted"
 
-            grouped_by_element = xr.merge(
-                (sel_parcels.element_id, sel_parcels.volume, d_weighted, rho_weighted)
-            ).groupby("element_id")
+            grouped_by_element = xr.merge((sel_parcels.element_id,
+                             sel_parcels.volume,
+                             d_weighted,
+                             rho_weighted)).groupby("element_id")
 
-            d_avg = (
-                grouped_by_element.sum().d_weighted / grouped_by_element.sum().volume
-            )
-            rho_avg = (
-                grouped_by_element.sum().rho_weighted / grouped_by_element.sum().volume
-            )
+            d_avg = grouped_by_element.sum().d_weighted/grouped_by_element.sum().volume
+            rho_avg = grouped_by_element.sum().rho_weighted/grouped_by_element.sum().volume
 
             self._d_mean_active = np.zeros(self._grid.size("link"))
             self._d_mean_active[d_avg.element_id.values.astype(int)] = d_avg.values
 
[…]
