### 2020-01-22 17849fc
fix pip update docs (#1119)
