+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -55,6 +55,7 @@ _REQUIRED_PARCEL_ATTRIBUTES = [
 _ACTIVE = 1
 _INACTIVE = 0
 
+_SAND_SIZE = 0.002
 
 class NetworkSedimentTransporter(Component):
     """
@@ -64,6 +65,9 @@ class NetworkSedimentTransporter(Component):
     described in Czuba (2018). Additions include: particle abrasion, variable
     active layer thickness (Wong et al., 2007).
 
+    This component cares about units.
+
+    TODO: statement about what units are required, including for dt (seconds)
 
     **Usage:**
     Option 1 - Basic::
@@ -347,7 +351,10 @@ class NetworkSedimentTransporter(Component):
         # Note that at present FlowDirector is just used for network connectivity.
 
         # get alluvium depth and calculate topography from br+alluvium, then update slopes.
+
         self._create_new_parcel_time()
+        self._calculate_mean_D_and_rho()
+
         self._partition_active_and_storage_layers()
         self._adjust_node_elevation()
         self._update_channel_slopes()
@@ -380,10 +387,10 @@ class NetworkSedimentTransporter(Component):
         )
         self._this_timesteps_parcels[:, -1] = True
 
-        self._parcels_off_grid = (
+        parcels_off_grid = (
             self._parcels.dataset.element_id[:, -1] == _OUT_OF_NETWORK
         )
-        self._this_timesteps_parcels[self._parcels_off_grid, -1] = False
+        self._this_timesteps_parcels[parcels_off_grid, -1] = False
 
         self._num_parcels = self._parcels.number_of_items
         # ^ needs to run just in case we've added more parcels
@@ -402,11 +409,48 @@ class NetworkSedimentTransporter(Component):
                 self._grid.at_link["reach_length"][i],
             )
 
+    def _calculate_mean_D_and_rho(self):
+        """Calculate mean grain size and density on each link"""
+
+        current_parcels = self._parcels.dataset.isel(time=self._time_idx)
+
+        # In the first full timestep, we need to calc grain size & rho_sed.
+        # Assume all parcels are in the active layer for the purposes of
+        # grain size and mean sediment density calculations
+
+        # FUTURE: make it possible to circumvent this if mean grain size
+        # has already been calculated (e.g. during 'zeroing' runs)
+
+        # Calculate mean values for density and grain size (weighted by volume).
+        sel_parcels = current_parcels.where(current_parcels.element_id != _OUT_OF_NETWORK)
+
+        d_weighted = sel_parcels.D*sel_parcels.volume
+        rho_weighted = sel_parcels.density*sel_parcels.volume
+        d_weighted.name = "d_weighted"
+        rho_weighted.name = "rho_weighted"
+
+        grouped_by_element = xr.merge((sel_parcels.element_id,
+                         sel_parcels.volume,
+                         d_weighted,
+                         rho_weighted)).groupby("element_id")
+
+        d_avg = grouped_by_element.sum().d_weighted/grouped_by_element.sum().volume
+        rho_avg = grouped_by_element.sum().rho_weighted/grouped_by_element.sum().volume
+
+        self._d_mean_active = np.zeros(self._grid.size("link"))
+        self._d_mean_active[d_avg.element_id.values.astype(int)] = d_avg.values
+
+        self._rhos_mean_active = np.zeros(self._grid.size("link"))
+        self._rhos_mean_active[rho_avg.element_id.values.astype(int)] = rho_avg.values
+
     def _partition_active_and_storage_layers(self, **kwds):
         """For each parcel in the network, determines whether it is in the
         active or storage layer during this timestep, then updates node
         elevations.
+
         """
+        current_parcels = self._parcels.dataset.isel(time=self._time_idx)
+
         try:
[…]
