+++ landlab/data_record/data_record.py
@@ -1103,8 +1103,8 @@ class DataRecord(object):
         >>> v_f
         array([  8.,   3.,   4.,   5.,  0.,  0.,  0.,  0.,  0.])
 
-        An array of np.nan is returned when ``filter_array`` is all `False` and
-        ``fill_value`` is not specified.
+        An array of ``fill_value`` is returned when ``filter_array`` is all
+        ``False`` (np.nan is the default value).
 
         >>> f = dr.dataset['ages'] > 4000.
         >>> v_f = dr.calc_aggregate_value(func=np.sum,
@@ -1113,8 +1113,7 @@ class DataRecord(object):
         >>> v_f
         array([  nan,   nan,   nan,   nan,  nan,  nan,  nan,  nan,  nan])
 
-        An array of ``fill_value`` is returned when this parameter is
-        specified and ``filter_array`` is all `False`.
+        Other values can be specified for ``fill_value``.
 
         >>> f = dr.dataset['ages'] > 4000.
         >>> v_f = dr.calc_aggregate_value(func=np.sum,
@@ -1126,32 +1125,34 @@ class DataRecord(object):
         """
         filter_at = self._dataset["grid_element"] == at
 
-        valid = np.arange(self._grid[at].size)
-
-        filter_valid_element = np.isin(self._dataset["element_id"], valid)
+        filter_valid_element = np.isin(
+            self._dataset["element_id"], np.arange(self._grid[at].size)
+        )
 
         if filter_array is None:
             my_filter = filter_at & filter_valid_element
-        elif np.all(filter_array is False):
-            return np.repeat(fill_value, self._grid[at].size)
         else:
             my_filter = filter_at & filter_array & filter_valid_element
 
-        # Filter DataRecord with my_filter and groupby element_id:
-        filtered = self._dataset.where(my_filter).groupby("element_id")
+        if np.any(my_filter):
 
-        vals = filtered.apply(func, *args, **kwargs)  # .reduce
+            # Filter DataRecord with my_filter and groupby element_id:
+            filtered = self._dataset.where(my_filter).groupby("element_id")
 
-        # create a nan array that we will fill with the results of the sum
-        # this should be the size of the number of elements, even if there are
-        # no items living at some grid elements.
-        out = fill_value * np.ones(self._grid[at].size)
+            vals = filtered.apply(func, *args, **kwargs)  # .reduce
 
-        # put the values of the specified variable into the correct location
-        # of the out array.
-        out[vals.element_id.values.astype(int)] = vals[data_variable]
+            # create a nan array that we will fill with the results of the sum
+            # this should be the size of the number of elements, even if there are
+            # no items living at some grid elements.
+            out = fill_value * np.ones(self._grid[at].size)
 
-        return out
+            # put the values of the specified variable into the correct location
+            # of the out array.
+            out[vals.element_id.values.astype(int)] = vals[data_variable]
+
+            return out
+        else:
+            return np.repeat(fill_value, self._grid[at].size)
 
     def ffill_grid_element_and_id(self):
         """Fill NaN values of the fields 'grid_element' and 'element_id'.
