### 2020-01-22 738344a
remove unnecessary bits now that data record is fixed.
