+++ landlab/data_record/data_record.py
@@ -1125,12 +1125,12 @@ class DataRecord(object):
         """
         filter_at = self._dataset["grid_element"] == at
 
-        filter_valid_element = (self._dataset["element_id"] > 0) and (self._dataset["element_id"] < self._grid[at].size - 1)
+        filter_valid_element = (self._dataset["element_id"] >= 0) * (self._dataset["element_id"] < self._grid[at].size)
 
-        if filter_array is None:
-            my_filter = filter_at & filter_valid_element
-        else:
+        if np.any(filter_array):
             my_filter = filter_at & filter_array & filter_valid_element
+        else:
+            my_filter = filter_at & filter_valid_element
 
         if np.any(my_filter):
             # Filter DataRecord with my_filter and groupby element_id:
