### 2020-01-22 cdb967d
use variable for init thickness
