### 2020-01-22 7fb963f
bit more syntax cleaning
