### 2020-01-22 75933a5
remove one additional calc_aggregate
