+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -526,7 +526,7 @@ class NetworkSedimentTransporter(Component):
                 idxinactive = np.where(cumvol > capacity[i])
                 make_inactive = parcel_id_time_sorted[idxinactive]
 
-                active_inactive[parcel_id_thislink] = _ACTIVE
+                active_inactive[parcel_id_thislink] = _ACTIVE # KRB-Q I think that this is an error. This is setting all parcels on this link to active. 
 
         self._parcels.dataset.active_layer[:, -1] = active_inactive
 
