### 2020-01-22 482edce
removing bug comment... I had made a mistake but git diff showed me right
