### 2020-01-22 5f40563
clean up active/inactive in partitioning
