### 2020-01-22 7839f2f
bit more clean up and formatting
currently at a 50x speedup (on a grid with 30 links and 50 parcels per link)
