+++ landlab/components/network_sediment_transporter/network_sediment_transporter.py
@@ -389,9 +389,7 @@ class NetworkSedimentTransporter(Component):
         )
         self._this_timesteps_parcels[:, -1] = True
 
-        parcels_off_grid = (
-            self._parcels.dataset.element_id[:, -1] == _OUT_OF_NETWORK
-        )
+        parcels_off_grid = self._parcels.dataset.element_id[:, -1] == _OUT_OF_NETWORK
         self._this_timesteps_parcels[parcels_off_grid, -1] = False
 
         self._num_parcels = self._parcels.number_of_items
@@ -424,20 +422,23 @@ class NetworkSedimentTransporter(Component):
         # has already been calculated (e.g. during 'zeroing' runs)
 
         # Calculate mean values for density and grain size (weighted by volume).
-        sel_parcels = current_parcels.where(current_parcels.element_id != _OUT_OF_NETWORK)
+        sel_parcels = current_parcels.where(
+            current_parcels.element_id != _OUT_OF_NETWORK
+        )
 
-        d_weighted = sel_parcels.D*sel_parcels.volume
-        rho_weighted = sel_parcels.density*sel_parcels.volume
+        d_weighted = sel_parcels.D * sel_parcels.volume
+        rho_weighted = sel_parcels.density * sel_parcels.volume
         d_weighted.name = "d_weighted"
         rho_weighted.name = "rho_weighted"
 
-        grouped_by_element = xr.merge((sel_parcels.element_id,
-                         sel_parcels.volume,
-                         d_weighted,
-                         rho_weighted)).groupby("element_id")
+        grouped_by_element = xr.merge(
+            (sel_parcels.element_id, sel_parcels.volume, d_weighted, rho_weighted)
+        ).groupby("element_id")
 
-        d_avg = grouped_by_element.sum().d_weighted/grouped_by_element.sum().volume
-        rho_avg = grouped_by_element.sum().rho_weighted/grouped_by_element.sum().volume
+        d_avg = grouped_by_element.sum().d_weighted / grouped_by_element.sum().volume
+        rho_avg = (
+            grouped_by_element.sum().rho_weighted / grouped_by_element.sum().volume
+        )
 
         self._d_mean_active = np.zeros(self._grid.size("link"))
         self._d_mean_active[d_avg.element_id.values.astype(int)] = d_avg.values
@@ -454,8 +455,12 @@ class NetworkSedimentTransporter(Component):
         current_parcels = self._parcels.dataset.isel(time=self._time_idx)
 
         self._vol_tot = self._parcels.calc_aggregate_value(
-                np.sum, "volume", at="link", filter_array=self._this_timesteps_parcels, fill_value=0.
-            )
+            np.sum,
+            "volume",
+            at="link",
+            filter_array=self._this_timesteps_parcels,
+            fill_value=0.0,
+        )
 
         # Wong et al. (2007) approximation for active layer thickness.
         # NOTE: calculated using grain size and grain density calculated for
@@ -485,7 +490,9 @@ class NetworkSedimentTransporter(Component):
         )  # in units of m
 
         links_with_no_active_layer = np.isnan(self._active_layer_thickness)
-        self._active_layer_thickness[links_with_no_active_layer] = np.mean(self._active_layer_thickness[links_with_no_active_layer==0])  # assign links with no parcels an average value
+        self._active_layer_thickness[links_with_no_active_layer] = np.mean(
+            self._active_layer_thickness[links_with_no_active_layer == 0]
+        )  # assign links with no parcels an average value
 
         if np.sum(np.isfinite(self._active_layer_thickness)) == 0:
             self._active_layer_thickness.fill(_INIT_ACTIVE_LAYER_THICKNESS)
[…]
