### 2020-01-23 cc1b851
Add'l doc and readme updates (#1120)
* fix few small things about the component readme

* add more info about where unit tests are located. few other changes.

* Better structure for development/index.rst
