+++ docs/source/development/index.rst
@@ -8,6 +8,9 @@ If you're intending to make changes to the Landlab code base, or want to
 develop your own components, this set of pages will provide information you
 need.
 
+Key Development Recommendations
+-------------------------------
+
 The most important things to remember are to:
 
   * Start by creating a fork of Landlab's repository;
@@ -26,6 +29,9 @@ Python 3.6, 3.7, and 3.8
 If you need to introduce a new dependency, that dependency must be compatible
 with Python  3.6+.
 
+Installation, Contribution, and Development Practices
+-----------------------------------------------------
+
 .. toctree::
    :maxdepth: 2
 
@@ -33,6 +39,9 @@ with Python  3.6+.
    contribution/index
    practices/index
    package_organization
+
+Quick Links For Package Maintenance
+-----------------------------------
 
 There are a few pages that are particularly important for the maintenance of
 the package. These are:
