### 2020-01-23 landlab/landlab#1108 @kbarnhart
@rreinecke I've now updated the main and component readme, as well as the documentation in response to your comments.

In many cases I've provided links into our more comprehensive documentation (hosted at https://landlab.readthedocs.io/). This is based on experience that duplicated information is difficult to maintain.

If these changes are insufficient, please indicate what needs improving and one of the manuscript authors will address the remaining improvements.
