### 2020-01-23 landlab/landlab#1108 @rreinecke
Thank you. This is much more comprehensible now. Of course there is always a trade of in duplicating information but I think this makes the project much more discoverable now.

I have to agree though with @DavidLitwin that the test structure could be somewhat different. But that is something out of the scope of this review. Before closing this: Am I correct that this is the only test file for the submitted component? landlab/landlab/tests/components/groundwater/test_dupuit_percolator.py
