### 2020-01-23 1a06376
Update resource_redistribution.py
Added documentation.
TODO: Add doctests.
