### 2020-01-23 landlab/landlab#1108 @kbarnhart
@rreinecke thanks for your comments on the readme/docs. I'll let you close this issue.

Regarding the location of tests for this component:
- The docstrings for the component (in the source code)
- The unit tests within `landlab/landlab/tests/components/groundwater/`

In addition, our continuous integration includes running all the notebooks and asserting that they run without error. This isn't such as strict test because we don't enforce any requirements on the output values (other than they are not errors).

Regarding test structure. I agree that it is beyond the scope of the review. Our reason for choosing this structure was so that our binary distributions don't include the unit test files.
