### 2020-01-23 72fef33
edits to JOSS paper
