### 2020-01-23 landlab/landlab#1105 @rreinecke
@DavidLitwin very nice paragraph. I think this provides a much more complete view on the issue. (even though I like to again to self-promote my publication here: https://www.geosci-model-dev.net/12/2401/2019" - maybe be future research into this topic ;))
