### 2020-01-23 landlab/landlab#1105 @rreinecke
@DavidLitwin can be closed as soon as it is merged.
