### 2020-01-23 landlab/landlab#1108 @rreinecke
as a side-note @DavidLitwin the tests are very nicely described and documented! A pleasure to read. Could always be extended in the future but to this point great addition to the rest of the tests. I've already recommended this project to some colleagues.
