### 2020-01-23 e4db79a
Update gdp tutorial for #1106
Correct conservation of mass plots, add info on fixed gradient boundaries
