### 2020-01-23 landlab/landlab#1106 @DavidLitwin
Referencing pull request #1123 Thank you for suggesting this, I think it adds a lot to the tutorial!
