### 2020-01-23 landlab/landlab#1108 @DavidLitwin
@rreinecke Thank you for the feedback on the tests! A good of thought went into them over several months.
