### 2020-01-23 f9499a2
Clarifications in joss paper
