### 2020-01-24 9464e1d
address #1118
Index area_of_cell to fix flux in and storage calculations
