### 2020-01-24 60c3628
Update groundwater tutorial for #1106 (#1123)
* various doc improvements

* a bit more about docs

* bug: missing `

* remove duplicate information.

* Address @mcflugen comments

* Start updates to gdp tutorial

* add timestep test

* Update gdp tutorial for #1106

Correct conservation of mass plots, add info on fixed gradient boundaries

* add text requests in three places

* add explanatory text to gdp tutorial

Co-authored-by: Katy Barnhart <katy.barnhart@gmail.com>
