### 2020-01-24 683e8c3
Address #1105 (#1122)
* Joss paper edits to address #1105

* edits to JOSS paper

* Clarifications in joss paper

Co-authored-by: Katy Barnhart <katy.barnhart@gmail.com>
