### 2020-01-24 176d614
indicate python requires 3.6
