### 2020-01-24 c51f13d
clean up init to use initialize_output_fields
