### 2020-01-26 835368d
doctests added to resource_redistribution
- added doctests to the ResourceRedistribution component. Tests are passing.
