### 2020-01-26 #1935 @kbarnhart
@rreinecke could you make an issue in Landlab and paste int the output of the failing tests? I'm surprised that they are failing, and want to make sure that this is addressed. Thanks for your review!
