### 2020-01-26 a7f6cb5
Update resource_redistribution.py
Documentation added to ResourceRedistribution component. Deviation from using all lower case for most variables is intentional. TODO: Minimize these deviations.
