+++ landlab/components/resource_redistribution/resource_redistribution.py
@@ -363,7 +363,7 @@ class ResourceRedistribution(Component):
     def erode(self):
         """
         This method removes resource R from a cell depending on the
-        vegetation (PFT) occupying that cell in discrete units of e.
+        vegetation (PFT) occupying each cell in discrete units of e.
         """
         V = self.grid.at_cell["vegetation__plant_functional_type"]
         R = self.grid.at_cell["soil__resources"]
@@ -403,6 +403,30 @@ class ResourceRedistribution(Component):
         eroded_soil_shrub: float, (-)
             R eroded from cells with burnt shrubs. This R
             is deposited in the cells neighboring burnt shrubs.
+
+        Returns
+        -------
+        A tuple: (burnt_shrubs_neigh,
+                  exclusive,
+                  shrub_exclusive,
+                  grass_exclusive,
+                  bare_exclusive,
+                  eroded_soil_part);
+        burnt_shrubs_neigh: numpy array of ints, (-)
+            cell_ids of looped shrub neighbors (neighbors of cells
+            occupied by BURNTSHRUB).
+        exclusive: numpy array of ints, (-)
+            cell_ids of cells except those occupied by BURNTSHRUB and
+            burnt_shrubs_neigh. (purpose - debugging)
+        shrub_exclusive: numpy array of ints, (-)
+            cell_ids of cells within exclusive that are occupied
+            by SHRUB. (purpose - debugging)
+        grass_exclusive: numpy array of ints, (-)
+            cell_ids of cells within exclusive that are occupied
+            by GRASS. (purpose - debugging)
+        bare_exclusive: numpy array of ints, (-)
+            cell_ids of cells within exclusive that are occupied
+            by BARE. (purpose - debugging)
         """
         V = self.grid.at_cell["vegetation__plant_functional_type"]
         R = self.grid.at_cell["soil__resources"]
@@ -443,6 +467,24 @@ class ResourceRedistribution(Component):
         """
         This method re-adjusts resource R based on the
         user-set maximum and minimum thresholds.
+        Returns
+        -------
+        A tuple: (resource_adjusted,
+                  eligible_locs_to_adj_neigh,
+                  Elig_locs,
+                  sed_to_borrow);
+        resource_adjusted: float, (-)
+            cumulative amount of R in excess of R_threshold.
+            (purpose - debugging)
+        eligible_locs_to_adj_neigh: numpy array of ints, (-)
+            cell_ids of neighbors of cells with
+            R > R_threshold. (purpose - debugging)
+        Elig_locs: numpy array of ints, (-)
+            cell_ids of cells where R < R_threshold.
+            R is deposited into these cells. (purpose - debugging)
+        sed_to_borrow: float, (-)
+            cumulative amount of R < R_threshold that
+            is borrowed. (purpose - debugging)
         """
         R = self.grid.at_cell["soil__resources"]
     ## Resource exceeding R_threshold will be distributed to its neighbors                                
@@ -481,17 +523,38 @@ class ResourceRedistribution(Component):
 
         Parameters
         ----------
-        V_age: int, array, shape=[number_of_cells] (yrs)
-            Age of the PFT (V) occupying each cell. This
+        V_age: numpy array of ints, shape=[number_of_cells] (yrs)
+            age of the PFT (V) occupying each cell.
+
+        Returns
+        -------
+        A tuple: (V_age, est_1, est_2, est_3, est_4, est_5);
+        V_age: numpy array of ints, (yrs)
+            age of the PFT (V) occupying each cell. This
             array is updated by this method.
+        est_1: numpy array of ints, (-)
[…]
