### 2020-01-27 d003e8f
Added some documentation to SpatialDisturbance
Working on SpatialDisturbance
- style edits
- added documentation (still in progress)
- TODO: add doctests.
