+++ landlab/components/spatial_disturbance/spatial_disturbance.py
@@ -9,15 +9,14 @@
 ## introduce fire suscessibility thresholds for shrubs and trees.
 ##############################################################################
 
-#%% Import Packages
+# Import Packages
 import numpy as np
 from landlab import FieldError, Component
 from ...utils.decorators import use_file_name_or_kwds
 from .funcs import (convert_phy_pft_to_distr_pft,
-                   convert_distr_pft_to_phy_pft)
+                    convert_distr_pft_to_phy_pft)
 
-
-_VALID_SCHEMES = set(["zhou_et_al_2013", "ravi_et_al_2009"])
+_VALID_SCHEMES = set(["ravi_et_al_2009", "zhou_et_al_2013"])
 
 
 def _assert_pft_scheme_is_valid(scheme):
@@ -25,7 +24,7 @@ def _assert_pft_scheme_is_valid(scheme):
         raise ValueError("%s: Invalid PFT scheme" % scheme)
 
 
-# %% Declare Global Variables (If any)
+# Declare Global Variables (If any)
 # 'ravi_et_al_2009' pft_scheme - used internally
 BARE = 0
 GRASS = 1
@@ -37,58 +36,118 @@ BURNTTREE = 6
 SHRUBSEED = 7
 TREESEED = 8
 
-# %%
-
 
 class SpatialDisturbance(Component):
     """
-    What is this component (brief description)?
+    Landlab component that implements that implements
+    spatial disturbances, such as wildfires and
+    grazing, conceptualized by Ravi and D'Odorico (2009).
+    These disturbances modify the vegetation occupying
+    the cells (e.g. convert SHRUB to BURNTSHRUB) depending on
+    their spatial reach.
+    
+    This component is designed to work with two different
+    plant functional type (PFT) schemes: 'ravi_et_al_2009',
+    and 'zhou_et_al_2013'.
+    'ravi_et_al_2009': Each cell in the RasterModelGrid object
+    can take an integer from 0 through 8 to
+    represent the cell states for bare soil [0],
+    grass [1], shrub [2], burnt grass [3], burnt shrub [4],
+    tree [5], burnt tree [6], shrub seed [7], and tree seed [8].
+    This scheme is an extended version of the PFTs used in 
+    Ravi and D'Odorico (2009).
+    'zhou_et_al_2013': Each cell in the RasterModelGrid object
+    can take an integer from 0 through 5 to
+    represent the cell states for grass [0],
+    shrub [1], tree [2], bare soil [3], shrub seedling [4],
+    and tree seedling [5]. This is the default scheme
+    used in the component.
 
-    A little more information about this component!
+    There are two key process-representing methods in this
+    component: graze(), and initiate_fires().
+    
+    References:
+    Ravi, S., & D’Odorico, P. (2009). Post-fire 
+    resource redistribution and fertility island dynamics
+    in shrub encroached desert grasslands: a modeling approach.
+    Landscape Ecology, 24(3), 325-335.
+    Zhou, X., Istanbulluoglu, E., & Vivoni, E. R. (2013). Modeling the
+    ecohydrological role of aspect controlled radiation on tree grass shrub
+    coexistence in a semiarid climate. Water Resources Research,
+    49(5), 2872-2895.
 
-    What does this component output? Are there multiple
-    processes? What are the key methods?
-
-    Default Vegetation Type classification: 0: Bare; 1: Grass;
-    2: Shrub; 3: Burnt Grass; 4: Burnt Shrub; 5: Trees;
-    6: Burnt Trees; 7: Shrub Seed; 8: Tree Seed
-
-    Parameters:
-    ----------
-    grid: RasterModelGrid
-        grid, Landlab's RasterModelGrid object
-    V: array_like
[…]
