### 2020-01-27 70021fa
Address #1118 (#1124)
* address #1118

Index area_of_cell to fix flux in and storage calculations

* use area_of_cell_at_node

* area_of_cell_at_node -> cell_area_at_node

* Add gdp test

test inactive interior nodes

Co-authored-by: Katy Barnhart <katy.barnhart@gmail.com>
