### 2020-01-27 035f1a1
NST plotting: plot pathway values
This "completes" the plotting function "plot_pathway_values" that plots an attribute along a given network pathway.
