+++ landlab/plot/network_sediment_transporter/__init__.py
@@ -7,6 +7,7 @@ Created on Tue Jun 18 14:22:27 2019
 from .locate_parcel_xy import locate_parcel_xy
 from .plot_network_links import plot_network_links
 from .plot_network_parcels import plot_network_parcels
+from .plot_pathway_values import plot_pathway_values
 
 # from .other_example_plot import other_example_plot
 
@@ -15,4 +16,5 @@ __all__ = [
     "plot_network_links",
     "locate_parcel_xy",
     "plot_network_parcels",
+    "plot_pathway_values",
 ]
