### 2020-01-27 ef0216f
state that slope is negative downwards.
