### 2020-01-27 2a87846
a new method for gdp stability
