### 2020-01-28 94dfb56
feat: add setters for erodibility for some stream power like components
