+++ landlab/components/stream_power/stream_power.py
@@ -262,7 +262,7 @@ class StreamPowerEroder(Component):
 
         self._A = return_array_at_node(grid, discharge_field)
         self._elevs = return_array_at_node(grid, "topographic__elevation")
-        self._K_unit_time = return_array_at_node(grid, K_sp)
+        self.K = return_array_at_node(grid, K_sp)
         self._sp_crit = return_array_at_node(grid, threshold_sp)
 
         assert np.all(self._sp_crit >= 0.0)
@@ -338,6 +338,15 @@ class StreamPowerEroder(Component):
         self._stream_power_erosion = self._grid.zeros(centering="node")
         self._alpha = self._grid.zeros("node")
 
+    @property
+    def K(self):
+        """Erodibility (units depend on m_sp)."""
+        return self._K
+
+    @K.setter
+    def K(self, new_val):
+        self._K = return_array_at_node(self._grid, new_val)
+
     def run_one_step(self, dt):
         """A simple, explicit implementation of a stream power algorithm.
 
@@ -377,7 +386,7 @@ class StreamPowerEroder(Component):
         # Operate the main function:
         if self._use_W:
             self._alpha[defined_flow_receivers] = (
-                self._K_unit_time[defined_flow_receivers]
+                self._K[defined_flow_receivers]
                 * dt
                 * self._A[defined_flow_receivers] ** self._m
                 / self._W[defined_flow_receivers]
@@ -386,7 +395,7 @@ class StreamPowerEroder(Component):
 
         else:
             self._alpha[defined_flow_receivers] = (
-                self._K_unit_time[defined_flow_receivers]
+                self._K[defined_flow_receivers]
                 * dt
                 * self._A[defined_flow_receivers] ** self._m
                 / (flow_link_lengths ** self._n)
