### 2020-01-28 120ea03
add notes that setters are being used.
