### 2020-01-29 fd6de59
feat: add setters for erodibility for some stream power like components (#1130)
* feat: add setters for erodibility for some stream power like components

* add notes that setters are being used.

* stream power smooth threshold eroder doesn't have a self._grid yet
