### 2020-01-30 6f7aeee
slight change to stability def
