### 2020-01-30 843969d
gdp doc updates, setter and getter
