### 2020-01-30 d0ab54f
use setters in init
