### 2020-01-31 0ecb323
Update spatial_disturbance.py
Added some documentation.
TODO: add doctests and complete adding documentation.
