### 2020-02-03 1d38a83
Updated SpatialDisturbance and its tests
- added documentation and doctests to SpatialDisturbance component.
- tests pass.
