+++ landlab/components/spatial_disturbance/spatial_disturbance.py
@@ -1,14 +1,4 @@
 
-##############################################################################
-## 21 May 2017
-## Authors: Sai Nudurupati & Erkan Istanbulluoglu
-## 
-## This is to replicate create_fires function from
-## Sujith Ravi's model as implemented in resource_redistribution_funcs.py.
-## In this fire creation method, we will also consider trees. Also, we
-## introduce fire suscessibility thresholds for shrubs and trees.
-##############################################################################
-
 # Import Packages
 import numpy as np
 from landlab import FieldError, Component
@@ -68,6 +58,12 @@ class SpatialDisturbance(Component):
     the other scheme, pass the vegetation
     field as input to the methods (i.e.,
     SpatialDisturbance.initiate_fires(V=veg_field)).
+    NOTE: The methods of this component internally
+    convert the input PFTs in 'zhou_et_al_2013' scheme
+    to 'ravi_et_al_2009', implement the algorithm,
+    and convert the PFTs back to 'zhou_et_al_2013'
+    when the component is instantiated with
+    'zhou_et_al_2013' scheme.
 
     There are two key process-representing methods in this
     component: graze(), and initiate_fires().
@@ -89,71 +85,117 @@ class SpatialDisturbance(Component):
     >>> import numpy as np
     >>> np.random.seed(0)
     >>> from landlab import RasterModelGrid as rmg
-    >>> from landlab.components import ResourceRedistribution
-    >>> grid = rmg((5, 4), xy_spacing=(0.2, 0.2))
-    >>> ResourceRedistribution.name
-    'Resource Redistribution'
-    >>> sorted(ResourceRedistribution.output_var_names)
-    ['soil__resources',
-     'vegetation__plant_functional_type']
-    >>> sorted(ResourceRedistribution.units)
-    [('soil__resources', 'None'),
-     ('vegetation__plant_functional_type', 'None')]
+    >>> from landlab.components import SpatialDisturbance
+    >>> grid = rmg((10, 10), xy_spacing=(0.2, 0.2))
+    >>> SpatialDisturbance.name
+    'Spatial Disturbance'
+    >>> sorted(SpatialDisturbance.output_var_names)
+    ['vegetation__plant_functional_type']
+    >>> sorted(SpatialDisturbance.units)
+    [('vegetation__plant_functional_type', 'None')]
+
+    Let us look at an example using the default
+    'zhou_et_al_2013' PFT scheme.
+
     >>> grid.at_cell["vegetation__plant_functional_type"] = (
-    ...     np.random.randint(0, 5, size=grid.number_of_cells))
-    >>> np.allclose(
-    ...    grid.at_cell["vegetation__plant_functional_type"],
-    ...    np.array([4, 0, 3, 3, 3, 1]))
+    ...     np.random.randint(0, 4, size=grid.number_of_cells))
+    >>> np.where(
+    ...     grid.at_cell["vegetation__plant_functional_type"]==0)[0].shape
+    (15,)
+    >>> sd = SpatialDisturbance(grid)
+    >>> sd._grid.number_of_cell_rows
+    8
+    >>> sd._grid.number_of_cell_columns
+    8
+    >>> sd._grid is grid
     True
-    >>> grid.at_cell["soil__resources"] = (
-    ...     np.ones(grid.number_of_cells, dtype=float))
-    >>> rr = ResourceRedistribution(grid)
-    >>> rr.grid.number_of_cell_rows
-    3
[…]
