### 2020-02-04 c9c329b
remove Python 2.7 from CI builds
