+++ .travis.yml
@@ -6,7 +6,6 @@ env:
   matrix:
   - CONDA_ENV=3.7
   - CONDA_ENV=3.6
-  - CONDA_ENV=2.7
   global:
   - MPLBACKEND=Agg
   - secure: exvIFiQjIunvyGXJB27wUskWoXOL/qgu1tLjpgSopZz34WRK7kszqD0fPBLpX7wWgvBuUPYt9tXjdj80gXsouhW/R44AxauRpeAvrcodOwKpFIKM5DOyLmBs+YHQvS1UDJ/n8Lr+n6Q0H36l3pwZC269UKQkpwDwUuGkobeMqXQ=
