### 2020-02-05 #1935 @rreinecke
@kbarnhart Sorry I'm super swamped right now. It might also only be a problem connected to my manual setup without anaconda. If I find the time I will open an issue in landlab. But I don't think this is blocking the completion of this review from my side.
