### 2020-02-06 #1935 @kthyng
@DavidLitwin Just read through your paper and have one small typo in the paper [#1144](https://github.com/landlab/landlab/pull/1144) and a few in the bib file [#1145](https://github.com/landlab/landlab/pull/1145). See what you think.

After that, let me know: what is the appropriate version number to use?

Also, can you set up an archive of your code at, for example, Zenodo, and report back here with the doi?
