### 2020-02-06 #1935 @DavidLitwin
@kthyng thank you for the typo corrections.

@kbarnhart I want to make sure the version and doi are handled correctly.. If we do it similarly to the JOSS paper for `lithology`, the landlab release where the component was introduced is referenced, along with the current landlab release and associated DOI. The component first appears in v.1.10 I believe, and current release is v2.0.0b5. @kbarnhart can you confirm this?
