### 2020-02-06 #1935 @DavidLitwin
Just fixed one other typo in landlab/landlab#1146.
