### 2020-02-06 #1935 @kbarnhart
@DavidLitwin you are correct. the typo you just fixed isn't in v2.0.0b5, but I'm not sure if that is important. if it is, we can make a new release.
