### 2020-02-07 #1935 @kbarnhart
@DavidLitwin I would manually make a Zenodo archive rather than use the auto-created Zenodo archive that sometimes get made when a new tag is created.

I also say this because I know in the [JOSS docs](https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance) there is a statement that the title and authors of the zenodo archive should match. The docs say:

> Check the archive deposit has the correct metadata (title and author list), and request the author edit it if it doesn’t match the paper.

So I think you'll need to download a .zip of the current version of Landlab and archive it with the name and authors of this contribution. I'll let @kthyng chime in if she thinks this is not correct.
