### 2020-02-08 #1935 @kthyng
@kbarnhart Yes this sounds correct to me. I haven't gone through the process myself but we do want the relevant version of the code archived at something like Zenodo, and we want the authors/title to match the paper exactly. Having the whole Landlab in the archive does make sense, given this is a submodule.
