### 2020-02-08 0f54b72
added SpeciesEvolver component abstract
