### 2020-02-08 d2e6395
clarify many-to-many zone relationship
