+++ joss/papers.bib
@@ -1,249 +1,0 @@
-
-@article{dunne_partial_1970,
-	title = {Partial {Area} {Contributions} to {Storm} {Runoff} in a {Small} {New} {England} {Watershed}},
-	volume = {6},
-	issn = {19447973},
-	doi = {10.1029/WR006i005p01296},
-	number = {5},
-	journal = {Water Resources Research},
-	author = {Dunne, Thomas and Black, Richard D.},
-	year = {1970},
-	note = {ISBN: 0043-1397},
-	pages = {1296--1311},
-}
-
-@article{sklash_role_1979,
-	title = {The role of groundwater in storm runoff},
-	issn = {0022-1694},
-	doi = {10.1016/S0167-5648(09)70009-7},
-	journal = {Developments in Water Science},
-	author = {Sklash, Michael G. and Farvolden, Robert N. and Farvolden, Robert N.},
-	year = {1979},
-}
-
-@article{troch_hillslope-storage_2003,
-	title = {Hillslope-storage {Boussinesq} model for subsurface flow and variable source areas along complex hillslopes: 1. {Formulation} and characteristic response},
-	volume = {39},
-	issn = {0043-1397},
-	doi = {10.1029/2002WR001728},
-	number = {11},
-	journal = {Water Resources Research},
-	author = {Troch, Peter A. and Paniconi, Claudio and Van Loon, E. Emiel},
-	year = {2003},
-	keywords = {Boussinesq equation, Subsurface flow, Groundwater modeling, Hillslope drainage, Linearization, Model simplification},
-}
-
-@article{marcais_dynamic_2017,
-	title = {Dynamic coupling of subsurface and seepage flows solved within a regularized partition formulation},
-	volume = {109},
-	doi = {10.1016/j.advwatres.2017.09.008},
-	journal = {Advances in Water Resources},
-	author = {Marçais, J. and de Dreuzy, J. R. and Erhel, J.},
-	year = {2017},
-	keywords = {Subsurface flow, Saturation excess, Boussinesq equations, ODE with discontinuous right-hand side, Regularization, Seepage},
-	pages = {94--105},
-}
-
-@article{huang_modelling_2006,
-	title = {Modelling the potential impacts of groundwater hydrology on long-term drainage basin evolution},
-	issn = {0197-9337},
-	doi = {10.1002/esp.1369},
-	journal = {Earth Surface Processes and Landforms},
-	author = {Huang, Xiangjiang and Niemann, Jeffrey D.},
-	year = {2006},
-	keywords = {Numerical modeling, Basin evolution, Groundwater hydrology, WE-38},
-}
-
-@article{hilberts_hillslope-storage_2004,
-	title = {The hillslope-storage {Boussinesq} model for non-constant bedrock slope},
-	volume = {291},
-	issn = {0022-1694},
-	doi = {10.1016/j.jhydrol.2003.12.043},
-	number = {3-4},
-	journal = {Journal of Hydrology},
-	author = {Hilberts, A. G J and Van Loon, E. E. and Troch, P. A. and Paniconi, C.},
-	year = {2004},
-	keywords = {Boussinesq equation, Hillslope hydrology, Subsurface flow, Kinematic wave equation, Richards equation},
-	pages = {160--173},
-}
-
-@article{jasechko_pronounced_2014,
-	title = {The pronounced seasonality of global groundwater recharge},
-	volume = {50},
-	issn = {1944-7973},
-	url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1002/2014WR015809},
-	doi = {10.1002/2014WR015809},
-	language = {en},
-	number = {11},
-	urldate = {2019-10-24},
[…]
